% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr_ames_housing.R
\name{ames_housing}
\alias{ames_housing}
\alias{mlr_tasks_ames_housing}
\title{House Sales in Ames, Iowa}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.
}
\description{
A regression task to predict house sale prices for Ames, Iowa.
This is the processed version from the \code{\link[AmesHousing:make_ames]{AmesHousing::make_ames()}} package.

Contains 80 features and 2930 observations.
Target column is \code{"Sale_Price"}.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_tasks$get("ames_housing")
tsk("ames_housing")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Dimensions: 2930x81
\item Properties: -
\item Has Missings: \code{FALSE}
\item Target: \dQuote{Sale_Price}
\item Features: \dQuote{Alley}, \dQuote{Bedroom_AbvGr}, \dQuote{Bldg_Type}, \dQuote{BsmtFin_SF_1}, \dQuote{BsmtFin_SF_2}, \dQuote{BsmtFin_Type_1}, \dQuote{BsmtFin_Type_2}, \dQuote{Bsmt_Cond}, \dQuote{Bsmt_Exposure}, \dQuote{Bsmt_Full_Bath}, \dQuote{Bsmt_Half_Bath}, \dQuote{Bsmt_Qual}, \dQuote{Bsmt_Unf_SF}, \dQuote{Central_Air}, \dQuote{Condition_1}, \dQuote{Condition_2}, \dQuote{Electrical}, \dQuote{Enclosed_Porch}, \dQuote{Exter_Cond}, \dQuote{Exter_Qual}, \dQuote{Exterior_1st}, \dQuote{Exterior_2nd}, \dQuote{Fence}, \dQuote{Fireplace_Qu}, \dQuote{Fireplaces}, \dQuote{First_Flr_SF}, \dQuote{Foundation}, \dQuote{Full_Bath}, \dQuote{Functional}, \dQuote{Garage_Area}, \dQuote{Garage_Cars}, \dQuote{Garage_Cond}, \dQuote{Garage_Finish}, \dQuote{Garage_Qual}, \dQuote{Garage_Type}, \dQuote{Gr_Liv_Area}, \dQuote{Half_Bath}, \dQuote{Heating}, \dQuote{Heating_QC}, \dQuote{House_Style}, \dQuote{Kitchen_AbvGr}, \dQuote{Kitchen_Qual}, \dQuote{Land_Contour}, \dQuote{Land_Slope}, \dQuote{Latitude}, \dQuote{Longitude}, \dQuote{Lot_Area}, \dQuote{Lot_Config}, \dQuote{Lot_Frontage}, \dQuote{Lot_Shape}, \dQuote{Low_Qual_Fin_SF}, \dQuote{MS_SubClass}, \dQuote{MS_Zoning}, \dQuote{Mas_Vnr_Area}, \dQuote{Mas_Vnr_Type}, \dQuote{Misc_Feature}, \dQuote{Misc_Val}, \dQuote{Mo_Sold}, \dQuote{Neighborhood}, \dQuote{Open_Porch_SF}, \dQuote{Overall_Cond}, \dQuote{Overall_Qual}, \dQuote{Paved_Drive}, \dQuote{Pool_Area}, \dQuote{Pool_QC}, \dQuote{Roof_Matl}, \dQuote{Roof_Style}, \dQuote{Sale_Condition}, \dQuote{Sale_Type}, \dQuote{Screen_Porch}, \dQuote{Second_Flr_SF}, \dQuote{Street}, \dQuote{Three_season_porch}, \dQuote{TotRms_AbvGrd}, \dQuote{Total_Bsmt_SF}, \dQuote{Utilities}, \dQuote{Wood_Deck_SF}, \dQuote{Year_Built}, \dQuote{Year_Remod_Add}, \dQuote{Year_Sold}
}
}

\seealso{
\itemize{
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}:
\url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html}
\item Package \CRANpkg{mlr3data} for more toy tasks.
\item Package \CRANpkg{mlr3oml} for downloading tasks from \url{https://www.openml.org}.
\item Package \CRANpkg{mlr3viz} for some generic visualizations.
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}
\item \code{as.data.table(mlr_tasks)} for a table of available \link[=Task]{Tasks} in the running session (depending on the loaded packages).
\item \CRANpkg{mlr3fselect} and \CRANpkg{mlr3filters} for feature selection and feature filtering.
\item Extension packages for additional task types:
\itemize{
\item Unsupervised clustering: \CRANpkg{mlr3cluster}
\item Probabilistic supervised regression and survival analysis: \url{https://mlr3proba.mlr-org.com/}.
}
}

Other Task: 
\code{\link{Task}},
\code{\link{TaskClassif}},
\code{\link{TaskRegr}},
\code{\link{TaskSupervised}},
\code{\link{TaskUnsupervised}},
\code{\link{mlr_tasks}},
\code{\link{mlr_tasks_breast_cancer}},
\code{\link{mlr_tasks_german_credit}},
\code{\link{mlr_tasks_iris}},
\code{\link{mlr_tasks_mtcars}},
\code{\link{mlr_tasks_penguins}},
\code{\link{mlr_tasks_pima}},
\code{\link{mlr_tasks_sonar}},
\code{\link{mlr_tasks_spam}},
\code{\link{mlr_tasks_wine}},
\code{\link{mlr_tasks_zoo}}
}
\concept{Task}
