% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{as.data.table}
\alias{as.data.table}
\title{Re-export of \code{as.data.table}

See \link[data.table:as.data.table]{data.table::as.data.table}.}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{as.data.table}}}
}}

