% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtype_image.R
\name{as.imagepath}
\alias{as.imagepath}
\title{Convert a file path to an image path}
\usage{
as.imagepath(x)
}
\arguments{
\item{x}{(\code{character}) \cr
Vector of (relative) image paths}
}
\description{
Convert a file path to an image path
}
