% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_resample_result.R
\name{as_resample_result}
\alias{as_resample_result}
\alias{as_resample_result.ResampleResult}
\alias{as_resample_result.ResultData}
\title{Convert to ResampleResult}
\usage{
as_resample_result(x, ...)

\method{as_resample_result}{ResampleResult}(x, ...)

\method{as_resample_result}{ResultData}(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Currently not used.}
}
\value{
(\link{ResampleResult}).
}
\description{
Convert object to a \link{ResampleResult}.
}
