% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\name{as_task_regr}
\alias{as_task_regr}
\alias{as_task_regr.TaskRegr}
\alias{as_task_regr.data.frame}
\alias{as_task_regr.DataBackend}
\title{Convert to a Regression Task}
\usage{
as_task_regr(x, ...)

\method{as_task_regr}{TaskRegr}(x, clone = FALSE, ...)

\method{as_task_regr}{data.frame}(x, target, id = deparse(substitute(x)), ...)

\method{as_task_regr}{DataBackend}(x, target, id = deparse(substitute(x)), ...)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert, e.g. a \code{data.frame()}.}

\item{...}{(\code{any})\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}, e.g.
by cloning it or constructing it from a \link[mlr3misc:Dictionary]{dictionary} such as \link{mlr_learners}.}

\item{target}{(\code{character(1)})\cr
Name of the target column.}

\item{id}{(\code{character(1)})\cr
Id for the new task.
Defaults to the (deparsed and substituted) name of \code{x}.}
}
\description{
Convert object to a regression task (\link{TaskRegr}).
}
