% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_unsupervised.R
\name{as_task_unsupervised}
\alias{as_task_unsupervised}
\alias{as_task_unsupervised.Task}
\alias{as_task_unsupervised.data.frame}
\alias{as_tasks_unsupervised}
\alias{as_task_unsupervised.list}
\title{Convert to an Unsupervised Task}
\usage{
as_task_unsupervised(x, ...)

\method{as_task_unsupervised}{Task}(x, clone = FALSE, ...)

\method{as_task_unsupervised}{data.frame}(
  x,
  id = deparse(substitute(x)),
  label = NA_character_,
  ...
)

as_tasks_unsupervised(x, ...)

\method{as_task_unsupervised}{list}(x, clone = FALSE, ...)

\method{as_task_unsupervised}{Task}(x, clone = FALSE, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}

\item{id}{(\code{character(1)})\cr
Id for the new task.
Defaults to the (deparsed and substituted) name of the data argument.}

\item{label}{(\code{character(1)})\cr
Label for the new instance.}
}
\description{
Convert object to a \link{TaskUnsupervised} or a list of \link{TaskUnsupervised}.
}
