% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_empty_ellipsis}
\alias{assert_empty_ellipsis}
\title{Assert Empty Ellipsis}
\usage{
assert_empty_ellipsis(...)
}
\arguments{
\item{...}{(any)\cr
Ellipsis arguments to check.}
}
\value{
\code{NULL}
}
\description{
Assert that \code{...} arguments are empty.
Use this function in S3-methods to ensure that misspelling of arguments does not go unnoticed.
}
\keyword{internal}
