% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackEvaluation.R
\name{assert_evaluation_callback}
\alias{assert_evaluation_callback}
\alias{assert_evaluation_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_evaluation_callback(callback, null_ok = FALSE)

assert_evaluation_callbacks(callbacks)
}
\arguments{
\item{callback}{(\link{CallbackEvaluation}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link{CallbackEvaluation}).}
}
\value{
[CallbackEvaluation | List of \link{CallbackEvaluation}s.
}
\description{
Assertions for \link{CallbackEvaluation} class.
}
