% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_internal_tuning.R
\name{assert_internal_tuning}
\alias{assert_internal_tuning}
\title{Assert Internal Tuning}
\usage{
assert_internal_tuning(learner, ids, ...)
}
\arguments{
\item{learner}{(\code{\link{Learner}})\cr
The learner.}

\item{ids}{(\code{character()})\cr
The parameter ids to check.}
}
\description{
Asserts that internal tuning is correctly configured for a \code{\link{Learner}}.
}
\keyword{internal}
