% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackResample.R
\name{assert_resample_callback}
\alias{assert_resample_callback}
\alias{assert_resample_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_resample_callback(callback, null_ok = FALSE)

assert_resample_callbacks(callbacks, null_ok = FALSE)
}
\arguments{
\item{callback}{(\link{CallbackResample}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link{CallbackResample}).}
}
\value{
\link{CallbackResample} | List of \link{CallbackResample}s.
}
\description{
Assertions for \link{CallbackResample} class.
}
