% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{assert_validate}
\alias{assert_validate}
\title{Assert Validate}
\usage{
assert_validate(x)
}
\arguments{
\item{x}{(any)\cr
The input to check.}
}
\description{
Asserts whether the input is a valid value for the \verb{$validate} field of a \code{\link{Learner}}.
}
