% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackWorkhorse.R
\name{assert_workhorse_callback}
\alias{assert_workhorse_callback}
\alias{assert_workhorse_callbacks}
\title{Assertions for Callbacks}
\usage{
assert_workhorse_callback(callback, null_ok = FALSE)

assert_workhorse_callbacks(callbacks)
}
\arguments{
\item{callback}{(\link{CallbackWorkhorse}).}

\item{null_ok}{(\code{logical(1)})\cr
If \code{TRUE}, \code{NULL} is allowed.}

\item{callbacks}{(list of \link{CallbackWorkhorse}).}
}
\value{
[CallbackWorkhorse | List of \link{CallbackWorkhorse}s.
}
\description{
Assertions for \link{CallbackWorkhorse} class.
}
