% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backend.R
\name{Backend}
\alias{Backend}
\title{Backend Interface}
\description{
All objects of type Backend provide the following interface:\preformatted{b$data(rows, cols)
b$head(n = 6)
b$rownames
b$colnames
b$nrow
b$ncol
print(b)
}

See \link{BackendDataTable} for an implementation of this interface.
}
\section{Arguments}{

\itemize{
\item \code{rows} (\code{integer()} or \code{character()}):
Vector of row indices to subset rows using the primary key in the data backend.
\item \code{cols} (\code{character()}):
Vector of column names to select specific columns.
\item \code{n} (\code{integer(1)}): Number of rows to return.
}
}

\section{Details}{

\code{$data()} returns a slice of the data as \link[data.table:data.table]{data.table}:
rows are matched by the \code{primary_key} column, columns are selected by name.

\code{$head()} (\code{data.table}) returns a \link[data.table:data.table]{data.table} of the first \code{n} data rows.

\code{$rownames} (\code{character(1)}) returns all rownames of \code{data} as integer or character vector.

\code{$colnames} (\code{character(1)}) returns all colnames of \code{data} as character vector.

\code{$nrow} (\code{integer(1)}) returns the number of total rows.

\code{$ncol} (\code{integer(1)}) returns the number of total columns, including primary key column.
}

\seealso{
Other Backend: \code{\link{BackendDataTable}}
}
\concept{Backend}
