% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackWorkhorse.R
\name{callback_workhorse}
\alias{callback_workhorse}
\title{Create Workhorse Callback}
\usage{
callback_workhorse(
  id,
  label = NA_character_,
  man = NA_character_,
  on_workhorse_before_train = NULL,
  on_workhorse_before_predict = NULL,
  on_workhorse_before_result = NULL
)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Identifier for the new instance.}

\item{label}{(\code{character(1)})\cr
Label for the new instance.}

\item{man}{(\code{character(1)})\cr
String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.
The referenced help package can be opened via method \verb{$help()}.}
}
\description{
Function to create a \link{CallbackWorkhorse}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Start Workhorse
 - on_workhorse_before_train
 - on_workhorse_before_predict
 - on_workhorse_before_result
End Tuning
}\if{html}{\out{</div>}}
}
\details{
When implementing a callback, each function must have two arguments named \code{callback} and \code{context}.
A callback can write data to the state (\verb{$state}), e.g. settings that affect the callback itself.
Workhorse callbacks access \link{ContextWorkhorse}.
}
