% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast_from_dict}
\alias{cast_from_dict}
\title{Cast objects using a Dictionary}
\usage{
cast_from_dict(x, type, dict, clone, multiple = TRUE)
}
\arguments{
\item{`x`}{(\code{character()} | \code{list()})\cr
Object to cast.}

\item{`type`}{(\code{character(1)})\cr
Expected type of objects.}

\item{`dict`}{(\link{Dictionary})\cr
Expected type of objects.}

\item{`clone`}{(\code{logical(1)})\cr
Clone objects, if necessary. Default is \code{FALSE}.}

\item{`multiple`}{(\code{logical(1)})\cr
Cast multiple objects of type \code{type} or just a single one?}
}
\value{
Object of type \code{type} or list of objects of type \code{type}.
}
\description{
Uses a dictionary to cast objects of a specific type.
Intended for package developers, internal use only.
}
\keyword{internal}
