% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{cjoin}
\alias{cjoin}
\title{Cross Join}
\usage{
cjoin(tasks, learners, resamplings)
}
\arguments{
\item{tasks}{(list of \link{Task}).}

\item{learners}{(list of \link{Learner}).}

\item{resamplings}{(list of \link{Resampling}).}
}
\value{
(\code{\link[=data.table]{data.table()}}) with the cross product of the input vectors.
}
\description{
Builds an exhaustive grid of tasks, learners, and resampling strategies.
Resampling strategies will be instantiated.
}
