% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\name{col_info}
\alias{col_info}
\alias{col_info.data.table}
\alias{col_info.DataBackend}
\title{Column Information for Backend}
\usage{
col_info(x, ...)

\method{col_info}{data.table}(x, primary_key = character(), ...)

\method{col_info}{DataBackend}(x, ...)
}
\arguments{
\item{x}{(any)\cr
A backend-like object for which to retrieve column information.}

\item{...}{(any)\cr
Additional arguments.}

\item{primary_key}{(\code{character()})\cr
The primary key of the backend.}
}
\description{
Collects column information for backend.

Currently, this includes:
\itemize{
\item storage type
\item levels (factor / ordered), but not for the primary key column
}
}
\keyword{internal}
