% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextEvaluation.R
\name{ContextEvaluation}
\alias{ContextEvaluation}
\title{Evaluation Context}
\description{
A \link{CallbackEvaluation} accesses and modifies data during \code{\link[=resample]{resample()}} and \code{\link[=benchmark]{benchmark()}} via the \code{ContextEvaluation}.
See the section on fields for a list of modifiable objects.
See \code{\link[=callback_evaluation]{callback_evaluation()}} for a list of stages that access \code{ContextEvaluation}.
}
\section{Super class}{
\code{\link[mlr3misc:Context]{mlr3misc::Context}} -> \code{ContextEvaluation}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{(\link{Task})\cr
The task to be evaluated.
The task is unchanged during the evaluation.}

\item{\code{learner}}{(\link{Learner})\cr
The learner to be evaluated.
The learner contains the models after stage \code{on_evaluation_before_train}.}

\item{\code{resampling}}{\link{Resampling}\cr
The resampling strategy to be used.
The resampling is unchanged during the evaluation.}

\item{\code{param_values}}{\code{list()}\cr
The parameter values to be used.
Is usually only set while tuning.}

\item{\code{iteration}}{(\code{integer()})\cr
The current iteration.}

\item{\code{sets}}{(\code{list()})\cr
The train and test set.
The sets are available on stage `on_evaluation_before_train``.}

\item{\code{test_set}}{(\code{integer()})\cr
Validation test set.
The set is only available when using internal validation.}

\item{\code{predict_sets}}{(\code{list()})\cr
The prediction sets stored in \code{learner$predict_sets}.
The sets are available on stage \code{on_evaluation_before_predict}.}

\item{\code{pdatas}}{(List of \link{PredictionData})\cr
The prediction data.
The data is available on stage \code{on_evaluation_end}.}

\item{\code{data_extra}}{(list())\cr
Data saved in the \link{ResampleResult} or \link{BenchmarkResult}.
Use this field to save results.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContextEvaluation-new}{\code{ContextEvaluation$new()}}
\item \href{#method-ContextEvaluation-clone}{\code{ContextEvaluation$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="format"><a href='../../mlr3misc/html/Context.html#method-Context-format'><code>mlr3misc::Context$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="print"><a href='../../mlr3misc/html/Context.html#method-Context-print'><code>mlr3misc::Context$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextEvaluation-new"></a>}}
\if{latex}{\out{\hypertarget{method-ContextEvaluation-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextEvaluation$new(task, learner, resampling, param_values, iteration)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{(\link{Task})\cr
The task to be evaluated.}

\item{\code{learner}}{(\link{Learner})\cr
The learner to be evaluated.}

\item{\code{resampling}}{(\link{Resampling})\cr
The resampling strategy to be used.}

\item{\code{param_values}}{(\code{list()})\cr
The parameter values to be used.}

\item{\code{iteration}}{(\code{integer()})\cr
The current iteration.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextEvaluation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContextEvaluation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextEvaluation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
