% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextWorkhorse.R
\name{ContextWorkhorse}
\alias{ContextWorkhorse}
\title{Batch Tuning Context}
\description{
A \link{CallbackResample} accesses and modifies data during the optimization via the \code{ContextBatchTuning}.
See the section on active bindings for a list of modifiable objects.
See \code{\link[=callback_batch_tuning]{callback_batch_tuning()}} for a list of stages that access \code{ContextBatchTuning}.
}
\section{Super class}{
\code{\link[mlr3misc:Context]{mlr3misc::Context}} -> \code{ContextResample}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContextResample-clone}{\code{ContextWorkhorse$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="format"><a href='../../mlr3misc/html/Context.html#method-Context-format'><code>mlr3misc::Context$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="initialize"><a href='../../mlr3misc/html/Context.html#method-Context-initialize'><code>mlr3misc::Context$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3misc" data-topic="Context" data-id="print"><a href='../../mlr3misc/html/Context.html#method-Context-print'><code>mlr3misc::Context$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContextResample-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContextResample-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContextWorkhorse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
