% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendCbind.R
\name{DataBackendCbind}
\alias{DataBackendCbind}
\title{DataBackendCbind}
\description{
Returns a named list of vectors of distinct values for each column
specified. If \code{na_rm} is \code{TRUE}, missing values are removed from the
returned vectors of distinct values. Non-existing rows and columns are
silently ignored.

Returns the number of missing values per column in the specified slice
of data. Non-existing rows and columns are silently ignored.
}
\section{Super class}{
\code{\link[mlr3:DataBackend]{mlr3::DataBackend}} -> \code{DataBackendCbind}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{rownames}}{\code{integer()}\cr
Returns vector of all distinct row identifiers, i.e. the primary key
column.}

\item{\code{colnames}}{\code{character()}\cr
Returns vector of all column names, including the primary key column.}

\item{\code{nrow}}{\code{integer(1)}\cr
Number of rows (observations).}

\item{\code{ncol}}{\code{integer(1)}\cr
Number of columns (variables), including the primary key column.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataBackendCbind$new()}}
\item \href{#method-data}{\code{DataBackendCbind$data()}}
\item \href{#method-head}{\code{DataBackendCbind$head()}}
\item \href{#method-distinct}{\code{DataBackendCbind$distinct()}}
\item \href{#method-missings}{\code{DataBackendCbind$missings()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="format">}\href{../../mlr3/html/DataBackend.html#method-format}{\code{mlr3::DataBackend$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="DataBackend" data-id="print">}\href{../../mlr3/html/DataBackend.html#method-print}{\code{mlr3::DataBackend$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Column bind two \link{DataBackends}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCbind$new(b1, b2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{b1}}{Backend 1}

\item{\code{b2}}{Backend 2}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\subsection{Method \code{data()}}{
Returns a slice of the data in the specified format. Currently, the
only supported formats are \code{"data.table"} and \code{"Matrix"}. The rows must
be addressed as vector of primary key values, columns must be referred
to via column names. Queries for rows with no matching row id and
queries for columns with no matching column name are silently ignored.
Rows are guaranteed to be returned in the same order as \code{rows}, columns
may be returned in an arbitrary order. Duplicated row ids result in
duplicated rows, duplicated column names lead to an exception.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCbind$data(rows, cols, data_format = self$data_formats[1L])}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{data_format}}{\cr
Desired data format, e.g. \code{"data.table"} or \code{"Matrix"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-head"></a>}}
\subsection{Method \code{head()}}{
Returns the first up-to \code{n} rows of the data as
\code{\link[data.table:data.table]{data.table::data.table()}}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCbind$head(n = 6L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-distinct"></a>}}
\subsection{Method \code{distinct()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCbind$distinct(rows, cols, na_rm = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{na_rm}}{\code{logical(1)}\cr
Whether to remove NAs or not.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-missings"></a>}}
\subsection{Method \code{missings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataBackendCbind$missings(rows, cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{rows}}{\code{integer()}\cr
Row indices.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{cols}}{\code{character()}\cr
Column names.}

\item{\code{cols}}{\code{character()}\cr
Column names.}
}
\if{html}{\out{</div>}}
}
}
}
