% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendMatrix.R
\name{DataBackendMatrix}
\alias{DataBackendMatrix}
\title{DataBackend for Matrix}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{DataBackend}.}
\description{
\link{DataBackend} for \CRANpkg{Matrix}.
Data is split into a (numerical) sparse part and a dense part.
These parts are automatically merged to a sparse format during \verb{$data()}.
Note that merging the data parts may cause data loss during the conversion.
}
\section{Construction}{
\preformatted{DataBackendMatrix$new(data, primary_key = NULL)
as_data_backend(data, primary_key = NULL, ...)
}
\itemize{
\item \code{data} :: \code{\link[Matrix:Matrix]{Matrix::Matrix()}}\cr
Sparse (numeric) data stored as \code{\link[Matrix:Matrix]{Matrix::Matrix()}}.
\item \code{dense} :: \code{\link[=data.frame]{data.frame()}}.
Dense data, converted to \code{\link[data.table:data.table]{data.table::data.table()}}.
\item \code{primary_key} :: \code{character(1)}\cr
Name of the primary key column in \code{dense}
}
}

\section{Fields}{

See \link{DataBackend}.
}

\section{Methods}{

See \link{DataBackend}.
}

\examples{
requireNamespace("Matrix")
data = Matrix::Matrix(sample(0:1, 20, replace = TRUE), ncol = 2)
colnames(data) = c("x1", "x2")
dense = data.frame(..row_id = 1:10, num = runif(10), fact = sample(c("a", "b"), 10, replace = TRUE))

b = as_data_backend(data, dense, "..row_id")
b$head()
b$data(1:3, b$colnames, data_format = "Matrix")
b$data(1:3, b$colnames, data_format = "data.table")
}
\seealso{
Other DataBackend: 
\code{\link{DataBackendDataTable}},
\code{\link{DataBackend}},
\code{\link{as_data_backend}()}
}
\concept{DataBackend}
