% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_fallback.R
\name{default_fallback}
\alias{default_fallback}
\alias{default_fallback.LearnerClassif}
\alias{default_fallback.LearnerRegr}
\title{Create a Fallback Learner}
\usage{
default_fallback(learner, ...)

\method{default_fallback}{LearnerClassif}(learner)

\method{default_fallback}{LearnerRegr}(learner)
}
\arguments{
\item{learner}{\link{Learner}\cr
The learner for which a fallback learner should be created.}
}
\value{
\link{Learner}
}
\description{
Create a fallback learner for a given learner.
The function searches for a suitable fallback learner based on the task type.
Additional checks are performed to ensure that the fallback learner supports the predict type.
}
