% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_measures.R
\name{default_measures}
\alias{default_measures}
\title{Get a Default Measure}
\usage{
default_measures(x)
}
\arguments{
\item{x}{:: \code{character(1)} | \link{Task} | \link{Learner}\cr
Get the default measure for the task type \code{x}, e.g., \code{"classif"} or \code{"regr"}.
If you pass an object of type \link{Task} or \link{Learner}, the corresponding task type is extracted.
If \code{x} is \code{NULL}, an empty list is returned.}
}
\value{
list of \link{Measure}.
}
\description{
Gets the default measures using the information in \link[=mlr_reflections]{mlr_reflections$default_measures}:
\itemize{
\item \code{\link[=mlr_measures_classif.ce]{"classif.ce"}} for classification
\item \code{\link[=mlr_measures_regr.mse]{"regr.mse"}} for regression
\item Add-on package may register additional default measures for their own task types
}
}
\examples{
default_measures("classif")
default_measures("regr")
}
