% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_values.R
\name{default_values}
\alias{default_values}
\alias{default_values.Learner}
\alias{default_values.LearnerClassifRpart}
\title{Extract Hyperparameter Default Values}
\usage{
default_values(x, search_space, task, ...)

\method{default_values}{Learner}(x, search_space, task, ...)

\method{default_values}{LearnerClassifRpart}(x, search_space, task, ...)
}
\arguments{
\item{x}{(\link{Learner})\cr
Learner to extract hyperparameter default values from.}

\item{search_space}{(\link[paradox:ParamSet]{paradox::ParamSet})\cr
Learner hyperparameter values are subsetted to hyperparameters in search space.}

\item{task}{(\link{Task})\cr
Task to train the learner on.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
\code{list()}.
}
\description{
Extract hyperparameter default values.
}
