% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\name{Dictionary}
\alias{Dictionary}
\title{Key-Value Storage}
\description{
A simple key-value store for \link[R6:R6]{R6::R6} objects.
On retrieval of an object, the following applies:
\itemize{
\item R6 Factories (objects of class \code{R6ClassGenerator}) are initialized (with no arguments).
\item \link[R6:R6Class]{R6::R6Class} objects are cloned.
\item Functions are called (with no arguments).
\item All other objects are returned as-is.
}
}
\section{Usage}{
\preformatted{# Construction
d = Dictionary$new()
# Getters
d$ids(pattern)
d$add(value)
d$get(id)
d$mget(ids)
d$has(ids)
d$remove(id, value)
d$remove(ids)
# S3 methods
as.data.frame(d)
as.data.table(d)
}
}

\section{Arguments}{

\itemize{
\item \code{pattern} (\code{character(1)}):\cr
Restrict to ids which match the regular expression \code{pattern}.
\item \code{id} (\code{character(1)}):\cr
Single id as string.
\item \code{value}:\cr
Arbitrary value.
\item \code{ids} (\code{character()}):\cr
Vector of ids.
}
}

\section{Details}{

\itemize{
\item \code{$new()} initializes a new object of class \link{Dictionary}.
\item \code{$ids()} (\code{character()}) returns a vector with all ids (or all ids matching \code{pattern}).
\item \code{$get()} retrieves a single object with id \code{id} (or raises an exception).
\item \code{$mget()} (named \code{list}) creates a list of objects with ids \code{ids} (or raises an exception).
\item \code{$has()} (\code{logical()}) returns a named logical of the same length as \code{ids} with value \code{TRUE} if the respective id is found in the Dictionary.
\item \code{$add()} adds item \code{value} with id \code{id} to the Dictionary.
\item \code{$remove()} removes item with id \code{id} from the Dictionary.
\item \code{as.data.frame()} and \code{as.data.table()} give a summarizing overview as \code{data.frame} or \code{data.table}, respectively.
}
}

\seealso{
Other Dictionary: \code{\link{mlr_learners}},
  \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}
}
\concept{Dictionary}
\keyword{internal}
