% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifDebug.R
\name{disable_inner_tuning.LearnerClassifDebug}
\alias{disable_inner_tuning.LearnerClassifDebug}
\title{Configure Inner Tuning for the Debug Classifier}
\usage{
\method{disable_inner_tuning}{LearnerClassifDebug}(learner, ids)
}
\arguments{
\item{validate}{(\code{numeric(1)}, \code{"inner_valid"} or \code{NULL})\cr
How to set the \code{validate} field of the learner, optional.}

\item{iter}{(\code{integer(1)})\cr
Value for parameter \code{iter}, optional.}
}
\description{
Configure the inner tuning for \code{\link{LearnerClassifDebug}}.
There are two modi:
\enumerate{
\item Disabling inner tuning (\code{.disable = TRUE}), sets \code{early_stopping = FALSE} and all other arguments are ignored.
\item Enablig inner tuning (\code{disable = FALSE}) sets \code{validate} and \code{iter} parameters if given and afterwards ensures
that the llearner is configured to actually perform inner tuning.
}

This function specifies the \code{early_stopping} parameter and ensures that the learner is correctly configured.
}
