% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_inner_tuning.R
\name{disable_inner_tuning}
\alias{disable_inner_tuning}
\title{Disable Inner Tuning}
\usage{
disable_inner_tuning(learner, ids, ...)
}
\arguments{
\item{learner}{(\code{\link{Learner}})\cr
The learner.}

\item{ids}{(\code{character()})\cr
IDs of the parameters for which to disable inner tuning.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
Modified \code{\link{Learner}}
}
\description{
Disable the inner tuning for some parameters of a \code{\link{Learner}}.
}
\details{
When implementing inner tuning for a \code{\link{Learner}}, this method must be implemented.
This is necessary to be able to use inner tuning with the \code{\link[mlr3tuning:AutoTuner]{mlr3tuning::AutoTuner}}, which -- after using
inner tuning during to find the best hyperparameters -- then needs to deactive it for the final model fit.

Note that calling this function should not change the value of the \verb{$validate} parameter.
}
\examples{
learner = lrn("classif.debug", validate = 0.3, early_stopping = TRUE, iter = 100)
disable_inner_tuning(learner, "iter")
learner$param_set$values
learner$validate
}
\keyword{internal}
