% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\docType{data}
\name{Experiment}
\alias{Experiment}
\title{Experiment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Experiment
}
\description{
Container object for machine learning experiments.
Holds all important information as computed by the steps \code{\link[=train]{train()}}, \code{\link[=predict]{predict()}} and \code{\link[=score]{score()}}.
}
\keyword{datasets}
