% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faggregate.R
\name{fscore}
\alias{fscore}
\title{Fast Scoring of ResampleResults and BenchmarkResults}
\usage{
fscore(obj, measure)
}
\arguments{
\item{obj}{(\link{ResampleResult} | \link{BenchmarkResult}).}

\item{measure}{(\link{Measure}).}
}
\description{
Scores a \link{ResampleResult} or \link{BenchmarkResult} for a single simple measure.
Returns the score for each iteration and uhash.
}
\details{
This function is faster than \verb{$score()} because it does not reassemble the resampling results.
It also does not require the task, learner, model or train set to be available.
It only works for simple measures.
}
