% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tasks.R
\name{generate_tasks}
\alias{generate_tasks}
\title{Generate Tasks for a Learner}
\usage{
generate_tasks(learner, N = 20L)
}
\arguments{
\item{learner}{(\link{Learner}).}

\item{N}{(\code{integer(1)}). Number of rows of generated tasks.}
}
\value{
(\code{list()}) of \link{Task}.
}
\description{
Generates multiple tasks for a given \link{Learner}, based on its properties.
This function is primarily used for unit tests, but can also assist while
writing custom learners.
}
\examples{
tasks = generate_tasks(mlr_learners$get("classif.rpart"))
tasks$missings_binary$data()
}
\keyword{internal}
