% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratorSmiley.R
\docType{data}
\name{GeneratorSmiley}
\alias{GeneratorSmiley}
\alias{mlr_generators_smiley}
\title{Smiley Classification Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Generator}.}
\usage{
GeneratorSmiley
}
\description{
A \link{Generator} for the smiley task in \code{\link[mlbench:mlbench.smiley]{mlbench::mlbench.smiley()}}.
}
\keyword{datasets}
