% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratorXor.R
\docType{data}
\name{GeneratorXor}
\alias{GeneratorXor}
\alias{mlr_generators_xor}
\title{XOR Classification Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Generator}.}
\usage{
GeneratorXor
}
\description{
A \link{Generator} for the xor task in \code{\link[mlbench:mlbench.xor]{mlbench::mlbench.xor()}}.
}
\keyword{datasets}
