% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_learner.R
\name{get_base_learner}
\alias{get_base_learner}
\title{Extract Base Learner}
\usage{
get_base_learner(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object to extract the base learner from.}

\item{...}{(any)\cr
Currently not used.}
}
\value{
\link{Learner} or \code{NULL}.
}
\description{
Extracts the base learners from nested objects such as the
\code{GraphLearner} in \CRANpkg{mlr3pipelines}.
Returns the \link{Learner} itself for regular learners,
and \code{NULL} if there is no learner to extract.
}
\keyword{internal}
