% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_resample_result.R
\name{hash_resample_result}
\alias{hash_resample_result}
\title{Calculate the Hash for a ResampleResult}
\usage{
hash_resample_result(task, learner, resampling)
}
\arguments{
\item{task}{:: \link{Task}.}

\item{learner}{:: \link{Learner}.}

\item{resampling}{:: \link{Resampling}.}
}
\value{
(\code{character(1)}) hash value.
}
\description{
Exported for add-on packages.
Calculates the hash of the \link{ResampleResult} given a \link{Task}, a \link{Learner} and a \link{Resampling}.
}
\keyword{internal}
