% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{ids}
\alias{ids}
\title{Extract the id slot}
\usage{
ids(x)
}
\arguments{
\item{x}{[\code{list()}]\cr
List of objects which have a slot named \code{"id"}.}
}
\value{
[\code{character}] of the same length as input \code{x}.
}
\description{
A simple helper which extracts the id slot of a list of objects.
}
\examples{
x = mlr_tasks$mget(c("iris", "sonar"))
ids(x)
}
