% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifFeatureless.R
\docType{data}
\name{LearnerClassifFeatureless}
\alias{LearnerClassifFeatureless}
\alias{mlr_learners_classif.featureless}
\title{Featureless Classification Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifFeatureless
}
\description{
A simple \link{LearnerClassif} which only analyses the labels during train, ignoring all features.
Hyperparameter \code{method} determines the mode of operation during prediction:
\describe{
\item{mode:}{Predicts the most frequent label. If there are two or more labels tied, randomly selects one per prediction.}
\item{sample:}{Randomly predict a label uniformly.}
\item{weighed.sample:}{Randomly predict a label, with probability estimated from the training distribution.}
}
}
\section{Construction}{
\preformatted{LearnerClassifFeatureless$new()
mlr_learners$get("classif.featureless")
lrn("classif.featureless")
}
}

\keyword{datasets}
