% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_rpart.R
\docType{data}
\name{LearnerRegrRpart}
\alias{LearnerRegrRpart}
\alias{mlr_learners_regr.rpart}
\title{Regression Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrRpart
}
\description{
A learner for a regression tree implemented in \link[rpart:rpart]{rpart::rpart}.
}
\keyword{datasets}
