% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mlr_tasks.R
\name{list_mlr_tasks}
\alias{list_mlr_tasks}
\title{List Tasks in mlr3}
\usage{
list_mlr_tasks(task = NULL)
}
\arguments{
\item{task}{(\code{character(1)}) \cr Task type to filter by.}
}
\value{
(\code{data.table()})
}
\description{
Lists all the predefined tasks in mlr3. Can be filtered by task types.
}
\examples{
# list all tasks
list_mlr_tasks()

# list regression only tasks
list_mlr_tasks("regr")

# list classif tasks with twoclass property
list_mlr_tasks("classif")[twoclass == TRUE]
}
