% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifCE.R
\docType{data}
\name{MeasureClassifCE}
\alias{MeasureClassifCE}
\alias{mlr_measures_classif.ce}
\title{Classification Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\description{
Calls \code{\link[Metrics:ce]{Metrics::ce()}}.
}
\section{Construction}{
\preformatted{MeasureClassifCE$new()
mlr_measures$get("classif.ce")
msr("classif.ce")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
}
\keyword{datasets}
