% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifConfusion.R
\docType{data}
\name{MeasureClassifConfusion}
\alias{MeasureClassifConfusion}
\alias{mlr_measures_classif.confusion}
\alias{mlr_measures_classif.tp}
\alias{mlr_measures_classif.fn}
\alias{mlr_measures_classif.fp}
\alias{mlr_measures_classif.tn}
\alias{mlr_measures_classif.tpr}
\alias{mlr_measures_classif.fnr}
\alias{mlr_measures_classif.fpr}
\alias{mlr_measures_classif.tnr}
\alias{mlr_measures_classif.ppv}
\alias{mlr_measures_classif.fdr}
\alias{mlr_measures_classif.for}
\alias{mlr_measures_classif.npv}
\alias{mlr_measures_classif.precision}
\alias{mlr_measures_classif.recall}
\alias{mlr_measures_classif.sensitivity}
\alias{mlr_measures_classif.specificity}
\title{Binary Classification Measures Derived from a Confusion Matrix}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifConfusion
}
\description{
All implemented \link{Measure}s call \code{\link[=confusion_measures]{confusion_measures()}} with the respective \code{type} internally.
}
\examples{
task = tsk("german_credit")
learner = lrn("classif.rpart")
p = learner$train(task)$predict(task)
round(p$score(c("classif.sensitivity", "classif.specificity")), 2)
}
\keyword{datasets}
