% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifF1.R
\docType{data}
\name{MeasureClassifF1}
\alias{MeasureClassifF1}
\alias{mlr_measures_classif.f1}
\title{F1 Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\description{
Calls \code{\link[Metrics:f1]{Metrics::f1()}}.
}
\section{Construction}{
\preformatted{MeasureClassifF1$new()
mlr_measures$get("classif.f1")
msr("classif.f1")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
}
\keyword{datasets}
