% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifFScore.R
\docType{data}
\name{MeasureClassifFScore}
\alias{MeasureClassifFScore}
\alias{mlr_measures_classif.f_score}
\title{F-score Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\description{
Calls \code{\link[Metrics:fbeta_score]{Metrics::fbeta_score()}}.
Argument \code{beta} defaults to \code{1}.
}
\section{Construction}{
\preformatted{MeasureClassifFScore$new(beta = 1)
mlr_measures$get("classif.f_score")
msr("classif.f_score")
}
\itemize{
\item \code{beta} :: \code{numeric(1)}\cr
Non-negative parameter balancing precision and recall.
Passed down to \code{\link[Metrics:fbeta_score]{Metrics::fbeta_score()}}
}
}

\references{
\cite{mlr3}{sasaki_2007}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
}
\keyword{datasets}
