% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_classif_mmce.R
\docType{data}
\name{MeasureClassifMMCE}
\alias{MeasureClassifMMCE}
\alias{mlr_measures_classif.mmce}
\title{Mean Misclassification Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifMMCE
}
\description{
Calls \link[Metrics:ce]{Metrics::ce}.
}
\keyword{datasets}
