% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureDebug.R
\docType{data}
\name{MeasureDebug}
\alias{MeasureDebug}
\alias{mlr_measures_debug}
\title{Debug Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\description{
This measure returns the number of observations in the \link{Prediction} object.
Its main purpose is debugging.
}
\section{Construction}{
\preformatted{MeasureDebug$new()
mlr_measures$get("debug")
msr("debug")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
}
\keyword{datasets}
