% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureElapsedTime.R
\docType{data}
\name{MeasureElapsedTime}
\alias{MeasureElapsedTime}
\alias{mlr_measures_elapsed_time}
\title{Elapsed Time Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\usage{
MeasureElapsedTime
}
\description{
Measures the elapsed time during train ("time_train"), predict ("time_predict"), or both ("time_both").
}
\keyword{datasets}
