% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureNumberObservations.R
\docType{data}
\name{MeasureNPredict}
\alias{MeasureNPredict}
\alias{mlr_measures_n_predict}
\title{Number of Observations in Prediction Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\description{
Returns the number of observations in the \link{Prediction} object.
Its main purpose is debugging.
}
\section{Construction}{
\preformatted{MeasureNPredict$new()
mlr_measures$get("n_predict")
msr("n_predict")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
}
\keyword{datasets}
