% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureOOBError.R
\docType{data}
\name{MeasureOOBError}
\alias{MeasureOOBError}
\alias{mlr_measures_oob_error}
\title{Out-of-bag Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\usage{
MeasureOOBError
}
\description{
Returns the out-of-bag error of the learner for learners that support it
(these learners have property "oob_error").
}
\keyword{datasets}
