% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegrMAE.R
\docType{data}
\name{MeasureRegrMAE}
\alias{MeasureRegrMAE}
\alias{mlr_measures_regr.mae}
\title{Absolute Errors Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{MeasureRegr}.}
\description{
Calls \code{\link[Metrics:mae]{Metrics::mae()}}.
}
\section{Construction}{
\preformatted{MeasureRegrMAE$new()
mlr_measures$get("regr.mae")
msr("regr.mae")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
}
\keyword{datasets}
