% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegrRMSE.R
\docType{data}
\name{MeasureRegrRMSE}
\alias{MeasureRegrRMSE}
\alias{mlr_measures_regr.rmse}
\title{Root Mean Squared Error Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{MeasureClassif}.}
\usage{
MeasureRegrRMSE
}
\description{
Calls \code{\link[Metrics:rmse]{Metrics::rmse()}}.
}
\keyword{datasets}
