% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSelectedFeatures.R
\name{MeasureSelectedFeatures}
\alias{MeasureSelectedFeatures}
\alias{mlr_measures_selected_features}
\title{Selected Features Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\description{
Measures the number of selected features.
}
\section{Construction}{
\preformatted{MeasureSelectedFeatures$new(normalize = FALSE)
mlr_measures$get("selected_features")
msr("selected_features")
}
\itemize{
\item \code{normalize} :: \code{logical(1)}\cr
If \code{normalize} is set to \code{TRUE}, divides the number of features by the total number of features.
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
}
