% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureSimple.R
\name{MeasureSimple}
\alias{MeasureSimple}
\alias{mlr_measures_classif.tpr}
\alias{mlr_measures_classif.auc}
\alias{mlr_measures_regr.mse}
\alias{mlr_measures_classif.ce}
\alias{mlr_measures_regr.rse}
\alias{mlr_measures_classif.tn}
\alias{mlr_measures_classif.precision}
\alias{mlr_measures_classif.tp}
\alias{mlr_measures_classif.specificity}
\alias{mlr_measures_classif.sensitivity}
\alias{mlr_measures_classif.fbeta}
\alias{mlr_measures_classif.tnr}
\alias{mlr_measures_classif.dor}
\alias{mlr_measures_classif.recall}
\alias{mlr_measures_regr.msle}
\alias{mlr_measures_regr.medae}
\alias{mlr_measures_regr.sae}
\alias{mlr_measures_regr.rsq}
\alias{mlr_measures_classif.fpr}
\alias{mlr_measures_regr.sse}
\alias{mlr_measures_regr.medse}
\alias{mlr_measures_classif.ppv}
\alias{mlr_measures_classif.fn}
\alias{mlr_measures_classif.fp}
\alias{mlr_measures_classif.fdr}
\alias{mlr_measures_classif.fomr}
\alias{mlr_measures_classif.fnr}
\alias{mlr_measures_classif.mcc}
\alias{mlr_measures_classif.logloss}
\alias{mlr_measures_regr.pbias}
\alias{mlr_measures_regr.maxae}
\alias{mlr_measures_regr.rmse}
\alias{mlr_measures_regr.mape}
\alias{mlr_measures_regr.smape}
\alias{mlr_measures_regr.rrse}
\alias{mlr_measures_classif.npv}
\alias{mlr_measures_classif.acc}
\alias{mlr_measures_regr.bias}
\alias{mlr_measures_regr.maxse}
\alias{mlr_measures_regr.rae}
\alias{mlr_measures_regr.ktau}
\alias{mlr_measures_regr.mae}
\alias{mlr_measures_regr.rmsle}
\alias{mlr_measures_regr.srho}
\title{Measures implemented in \CRANpkg{mlr3measures}}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\description{
The following measures from \CRANpkg{mlr3measures} are interfaced using a simple utility class:

Binary classification (inheriting from \link{MeasureClassif}):
\itemize{
\item \link[mlr3measures:auc]{auc}: Area Under the ROC Curve
\item \link[mlr3measures:dor]{dor}: Diagnostic Odds Ratio
\item \link[mlr3measures:fbeta]{fbeta}: F-beta score
\item \link[mlr3measures:fdr]{fdr}: False Discovery Rate
\item \link[mlr3measures:fn]{fn}: False Negatives
\item \link[mlr3measures:fnr]{fnr}: False Negative Rate
\item \link[mlr3measures:fomr]{fomr}: False Omission Rate
\item \link[mlr3measures:fp]{fp}: False Positives
\item \link[mlr3measures:fpr]{fpr}: False Positive Rate
\item \link[mlr3measures:mcc]{mcc}: Matthews Correlation Coefficient
\item \link[mlr3measures:npv]{npv}: Negative Predictive Value
\item \link[mlr3measures:ppv]{ppv}: Positive Predictive Value
\item \link[mlr3measures:precision]{precision}: Precision
\item \link[mlr3measures:recall]{recall}: Recall
\item \link[mlr3measures:sensitivity]{sensitivity}: Sensitivity
\item \link[mlr3measures:specificity]{specificity}: Specificity
\item \link[mlr3measures:tn]{tn}: True Negatives
\item \link[mlr3measures:tnr]{tnr}: True Negative Rate
\item \link[mlr3measures:tp]{tp}: True Positives
\item \link[mlr3measures:tpr]{tpr}: True Positive Rate
}

Multiclass classification (inheriting from \link{MeasureClassif}):
\itemize{
\item \link[mlr3measures:acc]{acc}: Classification Accuracy
\item \link[mlr3measures:ce]{ce}: Classification Error
\item \link[mlr3measures:logloss]{logloss}: Log Loss
}

Regression (inheriting from \link{MeasureRegr}):
\itemize{
\item \link[mlr3measures:bias]{bias}: Bias
\item \link[mlr3measures:ktau]{ktau}: Kendall's tau
\item \link[mlr3measures:mae]{mae}: Mean Absolute Errors
\item \link[mlr3measures:mape]{mape}: Mean Absolute Percent Error
\item \link[mlr3measures:maxae]{maxae}: Max Absolute Error
\item \link[mlr3measures:maxse]{maxse}: Max Squared Error
\item \link[mlr3measures:medae]{medae}: Median Absolute Errors
\item \link[mlr3measures:medse]{medse}: Median Squared Error
\item \link[mlr3measures:mse]{mse}: Mean Squared Error
\item \link[mlr3measures:msle]{msle}: Mean Squared Log Error
\item \link[mlr3measures:pbias]{pbias}: Percent Bias
\item \link[mlr3measures:rae]{rae}: Relative Absolute Error
\item \link[mlr3measures:rmse]{rmse}: Root Mean Squared Error
\item \link[mlr3measures:rmsle]{rmsle}: Root Mean Squared Log Error
\item \link[mlr3measures:rrse]{rrse}: Root Relative Squared Error
\item \link[mlr3measures:rse]{rse}: Relative Squared Error
\item \link[mlr3measures:rsq]{rsq}: R Squared
\item \link[mlr3measures:sae]{sae}: Sum of Absolute Errors
\item \link[mlr3measures:smape]{smape}: Symmetric Mean Absolute Percent Error
\item \link[mlr3measures:srho]{srho}: Spearman's rho
\item \link[mlr3measures:sse]{sse}: Sum of Squared Errors
}

Details about the implementation of the respective measures can be found on the corresponding
help page in \CRANpkg{mlr3measures}.
It is possible to customize the value of each measure which is returned if the measure
is not defined for the input by setting the public field \code{na_value} to an arbitrary \code{numeric(1)} (default is \code{NaN}).
}
\section{Construction}{

The measures can be retrieved from the dictionary \link{mlr_measures}.
With \code{"<id>"} being the name of the measure to construct:\preformatted{mlr_measures$get("<id>")
msr("<id>")
}
}

\examples{
mse = msr("regr.mse")
print(mse)

# score prediction on mtcars
task = tsk("mtcars")
pred = lrn("regr.featureless")$train(task)$predict(task)
mse$score(pred)

# rae: not defined for constant truth
pred = PredictionRegr$new(row_ids = 1:10, truth = rep(0, 10), response = runif(10))
rae = msr("regr.rae")
rae$score(pred)

# return worst possible value instead of NaN
rae$na_value = rae$range[2]
rae$score(pred)
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
}
