% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner.R, R/mlr.learners.R
\docType{class}
\name{mlr.learners}
\alias{mlr.learners}
\title{Registered Learners}
\format{\code{\link{R6Class}} object}
\usage{
mlr.learners

mlr.learners
}
\description{
\code{mlr.learners} is a \code{\link{Dictionary}} used to manage learners.

\code{mlr.learners} is a \code{\link{Dictionary}} used to manage learners.
}
\examples{
mlr.learners$ids
mlr.learners$contains("classif.dummy")
mlr.learners$get("classif.dummy")
mlr.learners$ids
mlr.learners$get("classif.dummy")
}
\keyword{datasets}
