% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\docType{class}
\name{mlr.resamplings}
\alias{mlr.resamplings}
\title{Registered Resampling Methods}
\format{\code{\link{R6Class}} object}
\usage{
mlr.resamplings
}
\description{
\code{Resamplings} is a \code{\link{Dictionary}} used to manage resampling methods.
}
\keyword{datasets}
