% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr.tasks.R
\docType{class}
\name{mlr.tasks}
\alias{mlr.tasks}
\title{Dictionary of defined Tasks}
\format{\code{\link{R6Class}} object}
\usage{
mlr.tasks
}
\description{
\code{Tasks} is a \code{\link{Dictionary}} used to manage tasks.
}
\examples{
# List task ids:
mlr.tasks$ids

# Get a briew summary:
as.data.frame(mlr.tasks)

# Retrieve a specific task:
mlr.tasks$get("iris")

# Add a new task, based on a subset of iris:
data = iris
data$Species = ifelse(data$Species == "setosa", "1", "0")
task = TaskClassif$new("iris.binary", data = data, target = "Species")
task$nclasses
mlr.tasks$add(task)
mlr.tasks$remove("iris.binary")
}
\keyword{datasets}
