% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_callbacks.R
\name{mlr3.score_measures}
\alias{mlr3.score_measures}
\title{Score Measures Callback}
\description{
This \code{CallbackEvaluation} scores measures directly on the worker.
This way measures that require a model can be scores without saving the model.
}
\examples{
clbk("mlr3.score_measures", measures = msr("classif.ce"))

task = tsk("pima")
learner = lrn("classif.rpart")
resampling = rsmp("cv", folds = 3)

callback = clbk("mlr3.score_measures", measures = msr("selected_features"))

rr = resample(task, learner, resampling = resampling, callbacks = callback)

rr$learners[[1]]$state$selected_features
}
