% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_asserts.R
\name{mlr3_assertions}
\alias{mlr3_assertions}
\alias{assert_backend}
\alias{assert_experiment}
\alias{assert_task}
\alias{assert_learner}
\alias{assert_measure}
\alias{assert_measures}
\alias{assert_resampling}
\alias{assert_resample_result}
\alias{assert_par_set}
\alias{assert_par_vals}
\title{Assertion for mlr3 objects}
\usage{
assert_backend(b)

assert_experiment(e)

assert_task(task)

assert_learner(learner, task = NULL)

assert_measure(measure, task = NULL, learner = NULL)

assert_measures(measures, task = NULL, learner = NULL)

assert_resampling(resampling)

assert_resample_result(resample_result)

assert_par_set(par_set)

assert_par_vals(par_vals, par_set)
}
\arguments{
\item{b}{(\code{[DataBackend]}).}

\item{e}{(\code{[Experiment]}).}

\item{task}{(\code{[Task]}).}

\item{learner}{(\code{[Learner]}).}

\item{measure}{(\code{[Measure]}).}

\item{measures}{(list of \code{[Measure]}).}

\item{resampling}{(\code{[Resampling]}).}

\item{resample_result}{(\code{[ResampleResult]}).}

\item{par_set}{(\code{[paradox::ParamSet()]}).}

\item{par_vals}{(\code{named list}).}
}
\description{
Functions intended to be used in packages extending \pkg{mlr3}.
}
\keyword{internal}
