% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mlr_assertions}
\alias{mlr_assertions}
\alias{assert_backend}
\alias{assert_experiment}
\alias{assert_task}
\alias{assert_learner}
\alias{assert_measure}
\alias{assert_measures}
\alias{assert_resampling}
\alias{assert_resample_result}
\alias{assert_benchmark_result}
\alias{assert_param_set}
\alias{assert_id}
\title{Assertion for mlr3 Objects}
\usage{
assert_backend(b, .var.name = vname(b))

assert_experiment(e, .var.name = vname(e))

assert_task(task, feature_types = NULL, task_properties = NULL,
  .var.name = vname(task))

assert_learner(learner, task = NULL, properties = character(0L),
  .var.name = vname(learner))

assert_measure(measure, task = NULL, predict_types = NULL,
  .var.name = vname(measure))

assert_measures(measures, task = NULL, predict_types = NULL,
  .var.name = vname(measures))

assert_resampling(resampling, instantiated = NULL,
  .var.name = vname(resampling))

assert_resample_result(resample_result,
  .var.name = vname(resample_result))

assert_benchmark_result(bmr, .var.name = vname(bmr))

assert_param_set(param_set, .var.name = vname(param_set))

assert_id(id, .var.name = vname(id))
}
\arguments{
\item{b}{(\link{DataBackend}).}

\item{e}{(\link{Experiment}).}

\item{task}{(\link{Task}).}

\item{feature_types}{(\code{character()})\cr
Set of allowed feature types.}

\item{task_properties}{(\code{character()})\cr
Set of required task properties.}

\item{learner}{(\link{Learner}).}

\item{measure}{(\link{Measure}).}

\item{predict_types}{(\code{character()}). Vector of predict types provided by the experiment/learner.}

\item{measures}{(list of \link{Measure}).}

\item{resampling}{(\link{Resampling}).}

\item{resample_result}{(\link{ResampleResult}).}

\item{bmr}{(\link{BenchmarkResult}).}

\item{param_set}{(\link[paradox:ParamSet]{paradox::ParamSet}).}

\item{id}{(\code{character(1)}).}

\item{task}{(\link{Task}).}
}
\description{
Functions intended to be used in packages extending \pkg{mlr3}.
}
\keyword{internal}
