% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_coercions.R
\name{mlr_coercions}
\alias{mlr_coercions}
\alias{as_task}
\alias{as_task.Task}
\alias{as_tasks}
\alias{as_tasks.list}
\alias{as_tasks.Task}
\alias{as_learner}
\alias{as_learner.Learner}
\alias{as_learners}
\alias{as_learners.list}
\alias{as_learners.Learner}
\alias{as_resampling}
\alias{as_resampling.Resampling}
\alias{as_resamplings}
\alias{as_resamplings.list}
\alias{as_resamplings.Resampling}
\alias{as_measure}
\alias{as_measure.NULL}
\alias{as_measure.Measure}
\alias{as_measures}
\alias{as_measures.NULL}
\alias{as_measures.list}
\alias{as_measures.Measure}
\title{Object Coercion}
\usage{
as_task(x, ...)

\method{as_task}{Task}(x, clone = FALSE, ...)

as_tasks(x, ...)

\method{as_tasks}{list}(x, clone = FALSE, ...)

\method{as_tasks}{Task}(x, clone = FALSE, ...)

as_learner(x, ...)

\method{as_learner}{Learner}(x, clone = FALSE, ...)

as_learners(x, clone = FALSE, ...)

\method{as_learners}{list}(x, clone = FALSE, ...)

\method{as_learners}{Learner}(x, clone = FALSE, ...)

as_resampling(x, ...)

\method{as_resampling}{Resampling}(x, clone = FALSE, ...)

as_resamplings(x, ...)

\method{as_resamplings}{list}(x, clone = FALSE, ...)

\method{as_resamplings}{Resampling}(x, clone = FALSE, ...)

as_measure(x, ...)

\method{as_measure}{`NULL`}(x, task_type = NULL, clone = FALSE, ...)

\method{as_measure}{Measure}(x, clone = FALSE, ...)

as_measures(x, ...)

\method{as_measures}{`NULL`}(x, task_type = NULL, clone = FALSE, ...)

\method{as_measures}{list}(x, clone = FALSE, ...)

\method{as_measures}{Measure}(x, clone = FALSE, ...)
}
\arguments{
\item{x}{(\code{any})\cr
Object to coerce.}

\item{...}{(\code{any})\cr
Additional arguments, currently ignored.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}, e.g.
by cloning it or constructing it from a \link[mlr3misc:Dictionary]{dictionary} such as \link{mlr_learners}.}

\item{task_type}{(\code{character(1)})\cr
Used if \code{x} is \code{NULL} to construct a default measure for the respective task type.
The default measures are stored in \code{\link[=mlr_reflections]{mlr_reflections$default_measures}}.}
}
\value{
Coerced object. The default method will return the object as-is.
Failed coercions have to be handled by on of the assertions in \link{mlr_assertions}.
}
\description{
S3 generics and methods to coerce to (lists of) \link{Task}, \link{Learner}, \link{Resampling}, \link{Measure} and \link{Prediction}.
}
\examples{
# convert single measure to list of measures
measure = msr("classif.ce")
as_measures(measure)
}
