% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generator2dNormals.R
\docType{data}
\name{mlr_generators_2dnormals}
\alias{mlr_generators_2dnormals}
\alias{Generator2DNormals}
\title{2d Normals Classification Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Generator}.}
\usage{
Generator2DNormals
}
\description{
A \link{Generator} for the 2d normals task in \code{\link[mlbench:mlbench.2dnormals]{mlbench::mlbench.2dnormals()}}.
}
\keyword{datasets}
