% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_debug.R
\docType{data}
\name{mlr_learners_classif_debug}
\alias{mlr_learners_classif_debug}
\alias{LearnerClassifDebug}
\title{Toy Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifDebug
}
\description{
A simple learner used primarily in the unit tests and for debugging purposes.
If no hyperparameter is set, it simply constantly predicts a randomly selected label.
The following hyperparameters trigger the following actions:
\describe{
\item{message_train:}{Outputs a message during train.}
\item{message_predict:}{Outputs a message during predict.}
\item{warning_train:}{Signals a warning during train.}
\item{warning_predict:}{Signals a warning during predict.}
\item{error_train:}{Raises an exception during train.}
\item{error_predict:}{Raises an exception during predict.}
\item{segfault_train:}{Provokes a segfault during train.}
\item{segfault_predict:}{Provokes a segfault during predict.}
}
Note that segfaults may not work on your operating system.
Also note that if they work, they will tear down your R session immediately!
}
\keyword{internal}
