% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_featureless.R
\docType{data}
\name{mlr_learners_classif_featureless}
\alias{mlr_learners_classif_featureless}
\alias{LearnerClassifFeatureless}
\title{Featureless Classification Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifFeatureless
}
\description{
A simple learner which only analyses the labels during train, ignoring all features.
Hyperparameter \code{method} determines the mode of operation during prediction:
\describe{
\item{mode:}{Predicts the most frequent label. If there are two or more labels tied, randomly selects one per prediction.}
\item{sample:}{Randomly predict a label uniformly.}
\item{weighed.sample:}{Randomly predict a label, with probability estimated from the training distribution.}
}
}
\keyword{internal}
