% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_rpart.R
\docType{data}
\name{mlr_learners_classif_rpart}
\alias{mlr_learners_classif_rpart}
\alias{LearnerClassifRpart}
\title{Classification Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifRpart
}
\description{
A learner for a classification tree implemented in \link[rpart:rpart]{rpart::rpart}.
}
\references{
\href{https://mlr3.mlr-org.com/reference/mlr_learners_classif_rpart.html}{HTML help page}
}
\keyword{datasets}
