% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_rpart.R
\docType{data}
\name{mlr_learners_regr_rpart}
\alias{mlr_learners_regr_rpart}
\alias{LearnerRegrRpart}
\title{Regression Tree Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrRpart
}
\description{
A learner for a regression tree implemented in package \pkg{rpart}.
}
\keyword{datasets}
