% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_acc.R
\docType{data}
\name{mlr_measures_acc}
\alias{mlr_measures_acc}
\alias{MeasureClassifACC}
\title{Accuracy Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifACC
}
\description{
Calls \link[Metrics:accuracy]{Metrics::accuracy}.
}
\references{
\href{https://mlr3.mlr-org.com/reference/mlr_measures_acc.html}{HTML help page}
}
\keyword{datasets}
