% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifF1.R
\docType{data}
\name{mlr_measures_classif.f1}
\alias{mlr_measures_classif.f1}
\alias{MeasureClassifF1}
\title{F1 Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifF1
}
\description{
Calls \code{\link[Metrics:f1]{Metrics::f1()}}.
}
\keyword{datasets}
