% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_mse.R
\docType{data}
\name{mlr_measures_mse}
\alias{mlr_measures_mse}
\alias{MeasureRegrMSE}
\title{Mean Squared Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureRegrMSE
}
\description{
Calls \link[Metrics:mse]{Metrics::mse}.
}
\keyword{datasets}
