% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureStabPhi.R
\name{mlr_measures_stab.phi}
\alias{mlr_measures_stab.phi}
\alias{MeasureStabPhi}
\title{Phi Stability Measure}
\description{
Calculates the stability of the sets of selected features
as the mean correlation between all pairs of 0/1 dummy vectors \eqn{z}.
These vectors have a 1 in the \eqn{i}-th position if the \eqn{i}-th feature has been selected.
With \eqn{m} features and \eqn{p} sets, this can be written as
\deqn{\frac{2}{m (m-1)} \sum_{i =1}^{p-1} \sum_{j = i + 1}^{p} \mathrm{Cor}(z_i, z_j).}
}
\details{
This measure operates on the \link{ResampleResult} and requires the model to be
stored (see \code{store_model} in \code{\link[=resample]{resample()}}/\code{\link[=benchmark]{benchmark()}}.
It is not possible to score a \link{Prediction} individually or to score resamplings with less
than two iterations.
}
\note{
More stability measures can be found in the package \CRANpkg{stabm}.
For a comparison of stability measures, see Bommert (2017) where this measure is
called \dQuote{SC}.
}
\section{Dictionary}{

This \link{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures} or with the associated sugar function \code{\link[=msr]{msr()}}:\preformatted{mlr_measures$get("phi")
msr("phi")
}
}

\section{Meta Information}{

\itemize{
\item Task Type: \dQuote{NA}
\item Range: \eqn{[-1, 1]}{[-1, 1]}
\item Minimize: \dQuote{FALSE}
\item Required prediction: \dQuote{response}
}
}

\references{
Bommert A, Rahnenführer J, Lang M (2017).
\dQuote{A Multicriteria Approach to Find Predictive and Sparse Models with Stable Feature Selection for High-Dimensional Data.}
\emph{Computational and Mathematical Methods in Medicine}, \bold{2017}, 1--18.
\doi{10.1155/2017/7907163}.

Nogueira S, Brown G (2016).
\dQuote{Measuring the Stability of Feature Selection.}
In \emph{Machine Learning and Knowledge Discovery in Databases}, 442--457.
Springer International Publishing.
\doi{10.1007/978-3-319-46227-1_28}.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.

Other stability measures: 
\code{\link{mlr_measures_stab.jaccard}}

Other Measure: 
\code{\link{MeasureClassif}},
\code{\link{MeasureRegr}},
\code{\link{Measure}},
\code{\link{mlr_measures_classif.costs}},
\code{\link{mlr_measures_debug}},
\code{\link{mlr_measures_elapsed_time}},
\code{\link{mlr_measures_oob_error}},
\code{\link{mlr_measures_selected_features}},
\code{\link{mlr_measures_stab.jaccard}},
\code{\link{mlr_measures}}
}
\concept{Measure}
\concept{stability measures}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureStabPhi}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{MeasureStabPhi$new()}}
\item \href{#method-aggregate}{\code{MeasureStabPhi$aggregate()}}
\item \href{#method-clone}{\code{MeasureStabPhi$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format">}\href{../../mlr3/html/Measure.html#method-format}{\code{mlr3::Measure$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help">}\href{../../mlr3/html/Measure.html#method-help}{\code{mlr3::Measure$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print">}\href{../../mlr3/html/Measure.html#method-print}{\code{mlr3::Measure$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score">}\href{../../mlr3/html/Measure.html#method-score}{\code{mlr3::Measure$score()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureStabPhi$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aggregate"></a>}}
\if{latex}{\out{\hypertarget{method-aggregate}{}}}
\subsection{Method \code{aggregate()}}{
A custom aggregation function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureStabPhi$aggregate(rr)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rr}}{\link{ResampleResult}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureStabPhi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
