% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_options.R
\name{mlr_options}
\alias{mlr_options}
\title{Options for mlr3}
\usage{
mlr_options()
}
\value{
(named \code{list}). Currently set options (without the prefix \code{"mlr3."}).
}
\description{
The following options are currently supported by \pkg{mlr3}:
\itemize{
\item \code{mlr3.verbose}: Set to \code{FALSE} to suppress output of informational messages.
\item \code{mlr3.debug}: Set to \code{TRUE} to enable extra output and additional (potentially expensive) assertions.
\item \code{mlr3.use_future}: Set to \code{FALSE} to disable parallelization via \pkg{future}.
Explicitly disabling this simplifies debugging.
}
}
\examples{
# get a list of currently set options (without the prefix)
mlr_options()
}
