% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_sugar.R
\name{mlr_sugar}
\alias{mlr_sugar}
\alias{tsk}
\alias{gen}
\alias{lrn}
\alias{rsp}
\alias{mea}
\title{Syntactic Sugar for Object Construction}
\usage{
tsk(.key, ...)

gen(.key, ...)

lrn(.key, ...)

rsp(.key, ...)

mea(.key, ...)
}
\arguments{
\item{.key}{:: \code{character(1)}\cr
Key passed to the respective \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} to retrieve the object.}

\item{...}{:: named \code{list()}\cr
Parameter values are added to the \link[paradox:ParamSet]{paradox::ParamSet} of the object if applicable.
Additional parameters are assigned to the respective field of the object.}
}
\value{
\link[R6:R6Class]{R6::R6Class} of the respective type.
}
\description{
Currently, there are three mechanisms in place to ease the construction of objects stored in a \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
\enumerate{
\item Whenever an specific object has to passed or created, it is possible to just pass a scalar string.
This string is used as key to retrieve the object form the respective dictionary.
\item Functions to retrieve objects, set hyperparameters and assign to fields in one go:
\itemize{
\item \code{tsk()} for a \link{Task} from \link{mlr_tasks}.
\item \code{gen()} for a \link{Generator} from \link{mlr_generators}.
\item \code{lrn()} for a \link{Learner} from \link{mlr_learners}.
\item \code{rsp()} for a \link{Resampling} from \link{mlr_resamplings}.
\item \code{mea()} for a \link{Measure} from \link{mlr_measures}.
}
\item Each task type has an associated default \link{Measure} (stored in \link{mlr_reflections}) which is used if no measure is provided.
Classification tasks default to the classification error in \link[=mlr_measures_classif.ce]{"classif.ce"},
regression tasks to the mean squared error in \link[=mlr_measures_regr.mse]{"regr.mse"}.
}
}
\examples{
task = tsk("iris", id = "iris2")
print(task)

learner = lrn("classif.rpart", cp = 0.1, predict_type = "prob")
print(learner)
}
