% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorCircle.R
\name{mlr_task_generators_circle}
\alias{mlr_task_generators_circle}
\alias{TaskGeneratorCircle}
\title{Circle Classification Task Generator}
\description{
A \link{TaskGenerator} for the circle binary classification task in \code{\link[mlbench:mlbench.circle]{mlbench::mlbench.circle()}}.
Creates a large circle containing a smaller circle.
}
\section{Dictionary}{

This \link{TaskGenerator} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_task_generators} or with the associated sugar function \code{\link[=tgen]{tgen()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_task_generators$get("circle")
tgen("circle")
}\if{html}{\out{</div>}}
}

\section{Parameters}{
\tabular{llll}{
   Id \tab Type \tab Default \tab Range \cr
   d \tab integer \tab 2 \tab \eqn{[2, \infty)}{[2, Inf)} \cr
}
}

\examples{
generator = tgen("circle")
plot(generator, n = 200)

task = generator$generate(200)
str(task$data())
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=TaskGenerator]{TaskGenerators}: \link{mlr_task_generators}
\item \code{as.data.table(mlr_task_generators)}  for a table of available \link[=TaskGenerator]{TaskGenerators} in the running session (depending on the loaded packages).
\item Extension packages for additional task types:
\itemize{
\item \CRANpkg{mlr3proba} for probabilistic supervised regression and survival analysis.
\item \CRANpkg{mlr3cluster} for unsupervised clustering.
}
}

Other TaskGenerator:
\code{\link{TaskGenerator}},
\code{\link{mlr_task_generators}},
\code{\link{mlr_task_generators_2dnormals}},
\code{\link{mlr_task_generators_cassini}},
\code{\link{mlr_task_generators_friedman1}},
\code{\link{mlr_task_generators_moons}},
\code{\link{mlr_task_generators_peak}},
\code{\link{mlr_task_generators_simplex}},
\code{\link{mlr_task_generators_smiley}},
\code{\link{mlr_task_generators_spirals}},
\code{\link{mlr_task_generators_xor}}
}
\concept{TaskGenerator}
\section{Super class}{
\code{\link[mlr3:TaskGenerator]{TaskGenerator}} -> \code{TaskGeneratorCircle}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TaskGeneratorCircle-initialize}{\code{TaskGeneratorCircle$new()}}
    \item \href{#method-TaskGeneratorCircle-plot}{\code{TaskGeneratorCircle$plot()}}
    \item \href{#method-TaskGeneratorCircle-clone}{\code{TaskGeneratorCircle$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="format"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-format'><code>TaskGenerator$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="generate"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-generate'><code>TaskGenerator$generate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="print"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-print'><code>TaskGenerator$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorCircle-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorCircle-initialize}{}}}
\subsection{\code{TaskGeneratorCircle$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskGeneratorCircle$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorCircle-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorCircle-plot}{}}}
\subsection{\code{TaskGeneratorCircle$plot()}}{
  Creates a simple plot of generated data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskGeneratorCircle$plot(n = 200L, pch = 19L, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{(\code{integer(1)})\cr
Number of samples to draw for the plot. Default is 200.}
      \item{\code{pch}}{(\code{integer(1)})\cr
Point char. Passed to \code{\link[=plot]{plot()}}.}
      \item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[=plot]{plot()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorCircle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorCircle-clone}{}}}
\subsection{\code{TaskGeneratorCircle$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskGeneratorCircle$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
