% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorSmiley.R
\name{mlr_task_generators_smiley}
\alias{mlr_task_generators_smiley}
\alias{TaskGeneratorSmiley}
\title{Smiley Classification Task Generator}
\description{
A \link{TaskGenerator} for the smiley task in \code{\link[mlbench:mlbench.smiley]{mlbench::mlbench.smiley()}}.
}
\section{Dictionary}{

This \link{TaskGenerator} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_task_generators} or with the associated sugar function \code{\link[=tgen]{tgen()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_task_generators$get("smiley")
tgen("smiley")
}\if{html}{\out{</div>}}
}

\section{Parameters}{
\tabular{llll}{
   Id \tab Type \tab Default \tab Range \cr
   sd1 \tab numeric \tab - \tab \eqn{[0, \infty)}{[0, Inf)} \cr
   sd2 \tab numeric \tab - \tab \eqn{[0, \infty)}{[0, Inf)} \cr
}
}

\examples{
generator = tgen("smiley")
plot(generator, n = 200)

task = generator$generate(200)
str(task$data())
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=TaskGenerator]{TaskGenerators}: \link{mlr_task_generators}
\item \code{as.data.table(mlr_task_generators)}  for a table of available \link[=TaskGenerator]{TaskGenerators} in the running session (depending on the loaded packages).
\item Extension packages for additional task types:
\itemize{
\item \CRANpkg{mlr3proba} for probabilistic supervised regression and survival analysis.
\item \CRANpkg{mlr3cluster} for unsupervised clustering.
}
}

Other TaskGenerator:
\code{\link{TaskGenerator}},
\code{\link{mlr_task_generators}},
\code{\link{mlr_task_generators_2dnormals}},
\code{\link{mlr_task_generators_cassini}},
\code{\link{mlr_task_generators_circle}},
\code{\link{mlr_task_generators_friedman1}},
\code{\link{mlr_task_generators_moons}},
\code{\link{mlr_task_generators_peak}},
\code{\link{mlr_task_generators_simplex}},
\code{\link{mlr_task_generators_spirals}},
\code{\link{mlr_task_generators_xor}}
}
\concept{TaskGenerator}
\section{Super class}{
\code{\link[mlr3:TaskGenerator]{TaskGenerator}} -> \code{TaskGeneratorSmiley}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-TaskGeneratorSmiley-initialize}{\code{TaskGeneratorSmiley$new()}}
    \item \href{#method-TaskGeneratorSmiley-plot}{\code{TaskGeneratorSmiley$plot()}}
    \item \href{#method-TaskGeneratorSmiley-clone}{\code{TaskGeneratorSmiley$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="format"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-format'><code>TaskGenerator$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="generate"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-generate'><code>TaskGenerator$generate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="print"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-print'><code>TaskGenerator$print()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorSmiley-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorSmiley-initialize}{}}}
\subsection{\code{TaskGeneratorSmiley$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskGeneratorSmiley$new()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorSmiley-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorSmiley-plot}{}}}
\subsection{\code{TaskGeneratorSmiley$plot()}}{
  Creates a simple plot of generated data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskGeneratorSmiley$plot(n = 200L, pch = 19L, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n}}{(\code{integer(1)})\cr
Number of samples to draw for the plot. Default is 200.}
      \item{\code{pch}}{(\code{integer(1)})\cr
Point char. Passed to \code{\link[=plot]{plot()}}.}
      \item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[=plot]{plot()}}.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGeneratorSmiley-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGeneratorSmiley-clone}{}}}
\subsection{\code{TaskGeneratorSmiley$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{TaskGeneratorSmiley$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
