% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegrAirpassengers.R
\name{mlr_tasks_airpassengers}
\alias{mlr_tasks_airpassengers}
\title{Air Passengers Regression Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.
}
\description{
A toy regression task for the \link[datasets:AirPassengers]{datasets::AirPassengers} data set.
The task represents a monthly time series and is ordered by
its only feature \code{date}.
}
\section{Construction}{
\preformatted{mlr_tasks$get("airpassengers")
tsk("airpassengers")
}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Dimensions: 144x2
\item Properties: \dQuote{ordered}
\item Has Missings: \code{FALSE}
\item Target: \dQuote{passengers}
\item Features: \dQuote{date}
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.

Other Task: 
\code{\link{TaskClassif}},
\code{\link{TaskRegr}},
\code{\link{TaskSupervised}},
\code{\link{TaskUnsupervised}},
\code{\link{Task}},
\code{\link{mlr_tasks_boston_housing}},
\code{\link{mlr_tasks_breast_cancer}},
\code{\link{mlr_tasks_german_credit}},
\code{\link{mlr_tasks_iris}},
\code{\link{mlr_tasks_mtcars}},
\code{\link{mlr_tasks_pima}},
\code{\link{mlr_tasks_sonar}},
\code{\link{mlr_tasks_spam}},
\code{\link{mlr_tasks_wine}},
\code{\link{mlr_tasks_zoo}},
\code{\link{mlr_tasks}}
}
\concept{Task}
