% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks_bh.R
\name{mlr_tasks_bh}
\alias{mlr_tasks_bh}
\title{Boston Housing Regression Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.}
\description{
A regression task for the \link[mlbench:BostonHousing2]{mlbench::BostonHousing2} data set.
}
