% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr_boston_housing.R
\name{mlr_tasks_boston_housing}
\alias{mlr_tasks_boston_housing}
\title{Boston Housing Regression Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.}
\description{
A regression task for the \link[mlbench:BostonHousing2]{mlbench::BostonHousing2} data set.
}
\section{Construction}{
\preformatted{mlr_tasks$get("boston_housing")
tsk("boston_housing")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}
}
