% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskOpenML.R
\name{mlr_tasks_openml}
\alias{mlr_tasks_openml}
\title{OpenML Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif} or \link{TaskRegr}.
}
\description{
A Task loaded from OpenML.

The data can either be identified with \code{data_id} (first default parameter), or with \code{task_id},
which gets the data of the referenced (OpenML) task.
}
\section{Construction}{
\preformatted{mlr_tasks$get("openml", 13)
tsk("openml", 13)
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.

Other Task: 
\code{\link{TaskClassif}},
\code{\link{TaskRegr}},
\code{\link{TaskSupervised}},
\code{\link{Task}},
\code{\link{mlr_tasks_boston_housing}},
\code{\link{mlr_tasks_german_credit}},
\code{\link{mlr_tasks_iris}},
\code{\link{mlr_tasks_mtcars}},
\code{\link{mlr_tasks_pima}},
\code{\link{mlr_tasks_sonar}},
\code{\link{mlr_tasks_spam}},
\code{\link{mlr_tasks_wine}},
\code{\link{mlr_tasks_zoo}},
\code{\link{mlr_tasks}}
}
\concept{Task}
