% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks.R
\name{mlr_tasks_titanic}
\alias{mlr_tasks_titanic}
\title{Titanic classification task}
\description{
A classification task for the \link[titanic:titanic]{titanic::titanic} data set.

The following preprocessing steps are already performed:
\itemize{
\item Training and test data has been merged. In the test data, the target variable "Survived" is NA.
\item Columns \code{Survived} and \code{Sex} have been converted to a factor.
\item Columns \code{Embarked} and \code{Cabin} have been converted to a factor and empty string have been replaced with NA.
\item The positive class is set to "1" (Survived).
}
}
