% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif.R
\name{new_task_classif}
\alias{new_task_classif}
\title{Create a New Classification Task}
\usage{
new_task_classif(backend, target, id = deparse(substitute(backend)))
}
\arguments{
\item{backend}{(\link{DataBackend})\cr
Either a \link{DataBackend}, or any object which is convertible to a \link{DataBackend} with \code{as_data_backend()}.
E.g., a \code{data.frame()} will be converted to a \link{DataBackendDataTable}.}

\item{target}{(\code{character(1)})\cr
Name of the target column.}

\item{id}{(\code{character(1)})\cr
Identifier for the new task.
Defaults to the (deparsed and substituted) name of \code{backend}.}
}
\description{
A wrapper around the constructor of \link{TaskClassif}.
}
\examples{
task = new_task_classif(palmerpenguins::penguins, target = "species")
}
