% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner.R
\name{predict_newdata_s3}
\alias{predict_newdata_s3}
\alias{predict_newdata_s3.data.frame}
\title{Predict model on new datasets}
\usage{
predict_newdata_s3(x, ...)

\method{predict_newdata_s3}{data.frame}(x, task = NULL, self = NULL)
}
\arguments{
\item{x}{(\code{data.frame()})\cr
New data to predict on.
Row ids are automatically set to \code{1:nrow(newdata)}.}

\item{...}{Passed down to the respective predict method.}

\item{task}{(\link{Task}).}
}
\value{
\link{Prediction}.

\link{Prediction}.
}
\description{
Uses the model fitted during \verb{$train()} to create a new \link{Prediction} based on the new data in \code{newdata}.
Object \code{task} is the task used during \verb{$train()} and required for conversion of \code{newdata}.
If the learner's \verb{$train()} method has been called, there is a (size reduced) version
of the training task stored in the learner.
If the learner has been fitted via \code{\link[=resample]{resample()}} or \code{\link[=benchmark]{benchmark()}}, you need to pass the corresponding task stored
in the \link{ResampleResult} or \link{BenchmarkResult}, respectively.

Uses the model fitted during \verb{$train()} to create a new \link{Prediction} based on the new data in \code{newdata}.
Object \code{task} is the task used during \verb{$train()} and required for conversion of \code{newdata}.
If the learner's \verb{$train()} method has been called, there is a (size reduced) version
of the training task stored in the learner.
If the learner has been fitted via \code{\link[=resample]{resample()}} or \code{\link[=benchmark]{benchmark()}}, you need to pass the corresponding task stored
in the \link{ResampleResult} or \link{BenchmarkResult}, respectively.
}
