% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_roc_measures.R
\name{print.roc_measures}
\alias{print.roc_measures}
\title{Print ROC Measures}
\usage{
\method{print}{roc_measures}(x, abbreviations = TRUE, digits = 2, ...)
}
\arguments{
\item{x}{(\code{roc_measures})\cr
The object returned by \code{score_roc_measures}.}

\item{abbreviations}{(\code{logical(1)})\cr
If \code{TRUE}, print a list of abbreviations for the measures.}

\item{digits}{(\code{integer(1)})\cr
Number of digits to round the measures to.}

\item{...}{(\code{any})\cr
Additional parameters, currently unused.}
}
\description{
Print the confusion matrix and a set of roc performance measures.
}
