% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.data.table}
\alias{data.table}
\alias{mlr_callbacks}
\alias{clbk}
\alias{clbks}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:as.data.table]{as.data.table()}}, \code{\link[data.table:data.table]{data.table()}}}

  \item{mlr3misc}{\code{\link[mlr3misc:clbk]{clbk()}}, \code{\link[mlr3misc:clbks]{clbks()}}, \code{\link[mlr3misc:mlr_callbacks]{mlr_callbacks}}}
}}

