% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingCustom.R
\docType{data}
\name{ResamplingCustom}
\alias{ResamplingCustom}
\alias{mlr_resamplings_custom}
\title{Custom Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\usage{
ResamplingCustom
}
\description{
A custom resampling class where the training and test indices can be set manually.
}
\section{Fields}{

See \link{Resampling}.
}

\section{Methods}{

See \link{Resampling}.
}

\examples{
# Create a task with 10 observations
task = mlr_tasks$get("iris")
task$filter(1:10)

# Instantiate Resampling
rc = mlr_resamplings$get("custom")
train_sets = list(1:5, 5:10)
test_sets = list(5:10, 1:5)
rc$instantiate(task, train_sets, test_sets)

rc$train_set(1)
rc$test_set(1)

# Resampling with custom observation weights for the training set
rc = mlr_resamplings$get("custom")
train_sets = list(1:5, 1:6)
test_sets = list(6:10, 7:10)
obs_weights_train_sets =
   list(train_sets[[1]] / sum(train_sets[[1]]), train_sets[[2]] / sum(train_sets[[2]]))
rc$instantiate(task, train_sets, test_sets, obs_weights_train_sets)
rc$train_set(1)
rc$test_set(1)
rc$obs_weights_train_set(1)
}
\keyword{datasets}
