% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner.R
\name{retrain_max_default}
\alias{retrain_max_default}
\title{Default Forward Which Retrain Function}
\usage{
retrain_max_default(retrain_value, xss)
}
\arguments{
\item{retrain_value}{(\code{numeric(1)}).}

\item{xss}{(\code{list()}).}
}
\description{
The list \code{xss} contains multiple "\code{retrain}" hyperparameter values.
This function determines which one allows to retrain to a learner
with the "\code{retrain}" hyperparameter value in \code{retrain_value} most efficiently.
For a \link{Learner} that can only retrain forwards.
}
