% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner.R
\name{retrain_min_default}
\alias{retrain_min_default}
\title{Default Backward Which Retrain Function}
\usage{
retrain_min_default(retrain_value, xss)
}
\description{
The list \code{xss} contains multiple "\code{retrain}" hyperparameter values.
This function determines which one allows to retrain to a learner
with the "\code{retrain}" hyperparameter value in \code{retrain_value} most efficiently.
For a \link{Learner} that can retrain backwards.
}
