% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{Score a Model on a Task}
\usage{
score(experiment, measures = NULL)
}
\arguments{
\item{experiment}{\link{Experiment}.\cr
Experiment with fitted model and predictions.}

\item{measures}{\link{Measure}\cr
Measure or list of measures.}
}
\value{
\code{\link{Experiment}}.
}
\description{
Scores the fitted model on a (new) subset of the task.
}
