% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_fallback.R
\name{set_fallback}
\alias{set_fallback}
\title{Set a Fallback Learner}
\usage{
set_fallback(learner)
}
\arguments{
\item{learner}{\link{Learner}\cr
The learner for which a fallback learner should be set.}
}
\value{
Returns the learner itself, but modified \strong{by reference}.
}
\description{
Set a fallback learner for a given learner.
The function searches for a suitable fallback learner based on the task type.
Additional checks are performed to ensure that the fallback learner supports the predict type.
}
