% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifDebug.R
\name{set_inner_tuning.LearnerClassifDebug}
\alias{set_inner_tuning.LearnerClassifDebug}
\title{Configure Inner Tuning for Debug Classifier}
\usage{
\method{set_inner_tuning}{LearnerClassifDebug}(.learner, .disable = FALSE, validate = NA, iter = NULL, ...)
}
\arguments{
\item{.learner}{(\code{\link{Learner}})\cr
The learner for which to set the inner tuning.}

\item{.disable}{(\code{logical(1)})\cr
Whether to disable the inner tuning, default is \code{FALSE}.}

\item{validate}{(\code{NA}, \code{numeric(1)}, \code{"inner_valid"} or \code{NULL})\cr
How to set the \code{validate} field of the learner.
If \code{NA} (default), it is se to \code{NULL} when disabling inner tuning and left unchanged otherwise.}

\item{iter}{(\code{integer(1)} or \code{NULL})\cr
Value for parameter \code{iter}.}

\item{...}{(any)\cr
Additional arguments.}
}
\description{
Configure the inner tuning for \code{\link{LearnerClassifDebug}}.
This function specifies the \code{early_stopping} parameter and ensures that the learner is correctly configured.
}
\examples{
learner = lrn("classif.debug")
set_inner_tuning(learner, validate = 0.3, iter = 100)
learner$validate
learner$early_stopping

set_inner_tuning(learner, .disable = TRUE)
learner$validate
learner$early_stopping
}
