% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_inner_tuning.R
\name{set_inner_tuning}
\alias{set_inner_tuning}
\title{Activate Inner Tuning}
\usage{
set_inner_tuning(.learner, .disable = FALSE, ...)
}
\arguments{
\item{.learner}{(\code{\link{Learner}})\cr
The learner for which to set the inner tuning.}

\item{.disable}{(\code{logical(1)})\cr
Whether to disable the inner tuning, default is \code{FALSE}.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
The input \code{\link{Learner}}
}
\description{
(De)activate the inner tuning for a learner.
}
\details{
Calling this function does not autoconfigure the inner tuning of the learner.
However, if the necessary parameter to configure the inner tuning are neither set nor specified via the \code{...}
parameter, the function should raise an informative error message.

Disabling the inner tuning needs to happen automatically, as it is required to make the \code{\link[mlr3tuning:AutoTuner]{mlr3tuning::AutoTuner}}
work.

Many learners need to ensure that validation is enabled to allow for inner tuning, however there are some learners
that don't have the \code{"validation"} property but can still do inner tuning. Examples include \code{cv.glmnet}.
}
