% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_validate.R
\name{set_validate}
\alias{set_validate}
\title{Enable Validation for Learner}
\usage{
set_validate(learner, validate, ...)
}
\arguments{
\item{learner}{(any)\cr
The learner.}

\item{validate}{(\code{numeric(1)}, \code{"inner_valid"}, \code{"test"}, or \code{NULL})\cr
Which validation set to use.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
Modified \code{\link{Learner}}
}
\description{
Helper function to configure the \verb{$validate} field(s) of a \code{\link{Learner}}.

This is especially useful for learners such as \code{\link[mlr3tuning:AutoTuner]{mlr3tuning::AutoTuner}} or \code{\link[mlr3pipelines:mlr_learners_graph]{mlr3pipelines::GraphLearner}},
where the \verb{$validate} fields need to be configured on multiple levels.
}
\examples{
learner = lrn("classif.debug")
set_validate(learner, 0.2)
learner$validate
}
