% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\name{task_check_col_roles}
\alias{task_check_col_roles}
\alias{task_check_col_roles.Task}
\alias{task_check_col_roles.TaskClassif}
\alias{task_check_col_roles.TaskRegr}
\alias{task_check_col_roles.TaskSupervised}
\alias{task_check_col_roles.TaskUnsupervised}
\title{Check Column Roles}
\usage{
task_check_col_roles(task, new_roles, ...)

\method{task_check_col_roles}{Task}(task, new_roles, ...)

\method{task_check_col_roles}{TaskClassif}(task, new_roles, ...)

\method{task_check_col_roles}{TaskRegr}(task, new_roles, ...)

\method{task_check_col_roles}{TaskSupervised}(task, new_roles, ...)

\method{task_check_col_roles}{TaskUnsupervised}(task, new_roles, ...)
}
\arguments{
\item{task}{(\link{Task})\cr
Task.}

\item{new_roles}{(\code{list()})\cr
Column roles.}
}
\description{
Internal function to check column roles.
}
\keyword{internal}
