% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorFriedman1.R
\name{TaskGeneratorFriedman1}
\alias{TaskGeneratorFriedman1}
\alias{mlr_task_generators_friedman1}
\title{Friedman1 Regression Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskGenerator}.}
\description{
A \link{TaskGenerator} for the friedman1 task in \code{\link[mlbench:mlbench.friedman1]{mlbench::mlbench.friedman1()}}.
}
\section{Construction}{
\preformatted{TaskGeneratorFriedman1$new()
mlr_task_generators$get("friedman1")
tgen("friedman1")
}
}

\examples{
tgen("friedman1")$generate(10)$data()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=TaskGenerator]{TaskGenerators}: \link{mlr_task_generators}

\code{as.data.table(mlr_resamplings)} for a complete table of all (also dynamically created) \link{TaskGenerator} implementations.
}
