% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorXor.R
\docType{data}
\name{TaskGeneratorXor}
\alias{TaskGeneratorXor}
\alias{mlr_task_generators_xor}
\title{XOR Classification Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskGenerator}.}
\usage{
TaskGeneratorXor
}
\description{
A \link{TaskGenerator} for the xor task in \code{\link[mlbench:mlbench.xor]{mlbench::mlbench.xor()}}.
}
\examples{
mlr_task_generators$get("xor")$generate(10)$data()
}
\keyword{datasets}
