% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_learner_classif.R
\name{test_learner_classif}
\alias{test_learner_classif}
\title{Runs Various Unit Tests on a Learner}
\usage{
test_learner_classif(learner)
}
\arguments{
\item{learner}{(\link{Learner}):\cr
Object of type \link{Learner}.}
}
\value{
None.
}
\description{
A learner is tested on many different tasks which cover the learner's specified feature types and properties.
}
\examples{
  lrn = LearnerClassifRpart$new()
  test_learner_classif(lrn)
}
