% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendCbind.R
\name{virtual_cbind}
\alias{virtual_cbind}
\title{Cbind two data formats
d2 can be NULL
must return the rows in the correct order, columns can be in any order}
\usage{
virtual_cbind(d1, d2, rows, cols, primary_key, ...)
}
\description{
Cbind two data formats
d2 can be NULL
must return the rows in the correct order, columns can be in any order
}
