% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendRbind.R
\name{virtual_rbind}
\alias{virtual_rbind}
\title{Rbind two data formats
data2 can be NULL
must return the rows in the correct order, columns can be in any order.}
\usage{
virtual_rbind(d1, d2, rows, cols, primary_key, ...)
}
\description{
Rbind two data formats
data2 can be NULL
must return the rows in the correct order, columns can be in any order.
}
