% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\name{as.data.table}
\alias{as.data.table}
\title{Re-export of as.data.table

See \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}
\description{
Re-export of as.data.table

See \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.
}
