% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R, R/ResampleResult.R
\name{as_benchmark_result}
\alias{as_benchmark_result}
\alias{as_benchmark_result.ResampleResult}
\title{Convert to BenchmarkResult}
\usage{
as_benchmark_result(x, ...)

\method{as_benchmark_result}{ResampleResult}(x, ...)
}
\arguments{
\item{x}{(\code{any})\cr
Object to dispatch on, e.g. a \link{ResampleResult}.}

\item{...}{(\code{any})\cr
Currently not used.}
}
\value{
(\link{BenchmarkResult}).
}
\description{
Simple S3 method to convert objects to a \link{BenchmarkResult}.
}
