% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_backend.R
\name{as_data_backend}
\alias{as_data_backend}
\title{Create a Data Backend.}
\usage{
as_data_backend(data, ...)
}
\arguments{
\item{data}{Data to wrap the backend around.
For a \code{data.frame()} (this includes \code{tibble()} from \pkg{tibble} and \code{data.table()}),
creates a \link{DataBackendDataTable}.
See \code{methods("as_data_backend")} for possible input formats.}

\item{...}{Additional arguments passed to the respective \link{DataBackend} method.}
}
\value{
\link{DataBackend}.
}
\description{
Wraps a \link{DataBackend} around data.
}
\seealso{
Other DataBackend: \code{\link{DataBackendDataTable}},
  \code{\link{DataBackendMatrix}},
  \code{\link{DataBackend}}
}
\concept{DataBackend}
