% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_backend.R
\name{as_data_backend}
\alias{as_data_backend}
\title{Create a Data Backend}
\usage{
as_data_backend(data, ...)
}
\arguments{
\item{data}{:: \code{any}\cr
Data to create a \link{DataBackend} for.
For a \code{data.frame()} (this includes \code{tibble()} from \CRANpkg{tibble} and \code{\link[data.table:data.table]{data.table::data.table()}})
this function creates a \link{DataBackendDataTable}.
See \code{methods("as_data_backend")} for possible input formats.
Note that third-party packages may extend this functionality.}

\item{...}{:: \code{any}\cr
Additional arguments passed to the respective \link{DataBackend} method.}
}
\value{
\link{DataBackend}.
}
\description{
Wraps a \link{DataBackend} around data.
}
\examples{
# create a new backend using the iris data:
as_data_backend(iris)
}
\seealso{
Other DataBackend: \code{\link{DataBackendDataTable}},
  \code{\link{DataBackendMatrix}},
  \code{\link{DataBackend}}
}
\concept{DataBackend}
