% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_learner.R
\name{as_learner}
\alias{as_learner}
\alias{as_learner.Learner}
\alias{as_learners}
\alias{as_learners.list}
\alias{as_learners.Learner}
\title{Convert to a Learner}
\usage{
as_learner(x, ...)

\method{as_learner}{Learner}(x, clone = FALSE, ...)

as_learners(x, clone = FALSE, ...)

\method{as_learners}{list}(x, clone = FALSE, ...)

\method{as_learners}{Learner}(x, clone = FALSE, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}
}
\value{
\link{Learner}.
}
\description{
Convert object to a \link{Learner} or a list of \link{Learner}.
}
