% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{as_prediction_data}
\alias{as_prediction_data}
\alias{new_prediction}
\title{Prediction Object Helpers

\code{as_prediction_data()} is called on the slave to check and convert the return
value of the \code{predict()} function of a \link{Learner}.

\code{new_prediction()} is used to construct a \link{Prediction} object from the data
returned by \code{as_prediction_data()}.}
\usage{
as_prediction_data(task, ...)

new_prediction(task, data)
}
\arguments{
\item{task}{:: \link{Task}\cr
Task as passed to \code{$predict()} of a \link{Learner}.}

\item{...}{:: any\cr
Arguments depending on the task type and predict type.
E.g., for classification, you may pass \code{response} and \code{prob}.}

\item{data}{:: named \code{list()}\cr
Data as returned by \code{as_prediction}, used to construct a \link{Prediction} object in
\code{new_prediction()}.}
}
\description{
Prediction Object Helpers

\code{as_prediction_data()} is called on the slave to check and convert the return
value of the \code{predict()} function of a \link{Learner}.

\code{new_prediction()} is used to construct a \link{Prediction} object from the data
returned by \code{as_prediction_data()}.
}
\keyword{internal}
