% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\name{as_resample_result}
\alias{as_resample_result}
\title{Convert to ResampleResult}
\usage{
as_resample_result(x, ...)
}
\arguments{
\item{x}{(\code{any})\cr
Object to dispatch on, e.g. a \link{ResampleResult}.}

\item{...}{(\code{any})\cr
Currently not used.}
}
\value{
(\link{ResampleResult}).
}
\description{
Simple S3 method to convert objects to a \link{ResampleResult}.
}
