% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_resampling.R
\name{as_resampling}
\alias{as_resampling}
\alias{as_resampling.Resampling}
\alias{as_resamplings}
\alias{as_resamplings.list}
\alias{as_resamplings.Resampling}
\title{Convert to a Resampling}
\usage{
as_resampling(x, ...)

\method{as_resampling}{Resampling}(x, clone = FALSE, ...)

as_resamplings(x, ...)

\method{as_resamplings}{list}(x, clone = FALSE, ...)

\method{as_resamplings}{Resampling}(x, clone = FALSE, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}
}
\description{
Convert object to a \link{Resampling} or a list of \link{Resampling}.
}
