% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task.R
\name{as_task}
\alias{as_task}
\alias{as_task.Task}
\alias{as_tasks}
\alias{as_tasks.list}
\alias{as_tasks.Task}
\title{Convert to a Task}
\usage{
as_task(x, ...)

\method{as_task}{Task}(x, clone = FALSE, ...)

as_tasks(x, ...)

\method{as_tasks}{list}(x, clone = FALSE, ...)

\method{as_tasks}{Task}(x, clone = FALSE, ...)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert.}

\item{...}{(\code{any})\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}
}
\description{
Convert object to a \link{Task} or a list of \link{Task}.
}
