% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_classif.R
\name{as_task_classif}
\alias{as_task_classif}
\alias{as_task_classif.TaskClassif}
\alias{as_task_classif.data.frame}
\alias{as_task_classif.DataBackend}
\title{Convert to a Classification Task}
\usage{
as_task_classif(x, ...)

\method{as_task_classif}{TaskClassif}(x, clone = FALSE, ...)

\method{as_task_classif}{data.frame}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  positive = NULL,
  ...
)

\method{as_task_classif}{DataBackend}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  positive = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert.}

\item{...}{(\code{any})\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}

\item{target}{(\code{character(1)})\cr
Name of the target column.}

\item{id}{(\code{character(1)})\cr
Id for the new task.
Defaults to the (deparsed and substituted) name of \code{x}.}

\item{positive}{(\code{character(1)})\cr
Level of the positive class. See \link{TaskClassif}.}
}
\description{
Convert object to a classification task (\link{TaskClassif}).
}
\examples{
as_task_classif(palmerpenguins::penguins, target = "species")
}
