% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark Multiple Learners on Multiple Tasks}
\usage{
benchmark(tasks, learners, resamplings, measures)
}
\arguments{
\item{tasks}{(\code{list} of \link{Task})\cr
List of objects of type \link{Task}.}

\item{learners}{(\code{list} of \link{Learner})\cr
List of objects of type \link{Learner}.}

\item{resamplings}{(\code{list} of \link{Resampling})\cr
List of objects of type \link{Resampling}.}

\item{measures}{(\code{list} of \link{Measure})\cr
List of objects of type \link{Measure}.}
}
\value{
\link{BenchmarkResult}.
}
\description{
Runs a benchmark (possibly in parallel).
}
\examples{
tasks = lapply(c("iris", "sonar"), mlr_tasks$get)
learners = lapply(c("classif.dummy", "classif.rpart"), mlr_learners$get)
resamplings = lapply("cv", mlr_resamplings$get)
measures = lapply("mmce", mlr_measures$get)
bmr = benchmark(tasks, learners, resamplings, measures)
bmr$performance
}
