% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{Benchmark Multiple Learners on Multiple Tasks}
\usage{
benchmark(tasks, learners, resamplings, measures = NULL,
  ctrl = exec_control())
}
\arguments{
\item{tasks}{[\code{list} of \link{Task}]:\cr
List of objects of type \link{Task}.}

\item{learners}{[\code{list} of \link{Learner}]:\cr
List of objects of type \link{Learner}.}

\item{resamplings}{[\code{list} of \link{Resampling}]:\cr
List of objects of type \link{Resampling}.}

\item{measures}{[\code{list} of \link{Measure}]:\cr
List of performance measures used to assess the predictive performance.
Defaults to the respective measures stored in \code{task}.}

\item{ctrl}{[\code{named list} as returned by \code{\link[=exec_control]{exec_control()}}]:\cr
Object to control experiment execution. See \code{\link[=exec_control]{exec_control()}}.}
}
\value{
\link{BenchmarkResult}.
}
\description{
Runs a benchmark of the cross-product of learners, tasks, and resampling strategies (possibly in parallel).
}
