% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\docType{data}
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{Container for Results of benchmark}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BenchmarkResult
}
\description{
Container for Results of benchmark
}
\keyword{datasets}
