% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\docType{data}
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{Container for Results of \code{benchmark()}}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
This is the result container object returned by \code{\link[=benchmark]{benchmark()}}.

Note that all stored objects are accessed by reference.
Do not modify any object without cloning it first.
}
\section{Construction}{
\preformatted{bmr = BenchmarkResult$new(data)
}
\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table with the data of one resampling iteration per row.
}
}

\section{Fields}{

\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Internal data storage.
\item \code{tasks} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of used tasks with three columns:
"task_hash" (\code{character(1)}), "task_id" (\code{character(1)}) and "task" (\link{Task}).
\item \code{learners} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of used learners with three columns:
"learner_hash" (\code{character(1)}), "learner_id" (\code{character(1)}) and "learner" (\link{Learner}).
\item \code{resamplings} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of used resamplings with three columns:
"resampling_hash" (\code{character(1)}), "resampling_id" (\code{character(1)}) and "resampling" (\link{Resampling}).
}
}

\section{Methods}{

\itemize{
\item \code{aggregated(objects = TRUE, ids = TRUE, params = FALSE)}\cr
(\code{logical(1)}, \code{logical(1)}, \code{logical(1)}) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a result table where experiments are aggregated per \link{ResampleResult}.
Arguments control the number of additional columns:
\itemize{
\item \code{objects} :: \code{logical(1)}\cr
Return objects as columns in the result \code{data.table()}.
\item \code{ids} :: \code{logical(1)}\cr
Return object ids as columns in the result \code{data.table()}.
\item \code{params} :: \code{logical(1)}\cr
Return learner hyperparameter values as list column \code{params} in the result \code{data.table()}.
}
\item \code{combine(bmr)}\cr
\link{BenchmarkResult} -> \code{self}\cr
Fuses a second \link{BenchmarkResult} into itself.
\item \code{get_best(id)}\cr
(\code{character(1)}) -> \link{ResampleResult}\cr
Returns the \link{ResampleResult} with best performance according to \link{Measure} with the provided id.
}
}

\section{S3 Methods}{

\itemize{
\item \code{as.data.table(bmr)}\cr
\link{BenchmarkResult} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Converts the data to a \code{data.table()}.
}
}

\examples{
set.seed(123)
design = expand_grid(
  tasks = "iris",
  learners = c("classif.featureless", "classif.rpart"),
  resamplings = "cv3"
)
print(design)

bmr = benchmark(design)
print(bmr)

bmr$tasks
bmr$learners

# aggregated results
bmr$aggregated()

# aggregated results with hyperparameters as separate columns
mlr3misc::unnest(bmr$aggregated(params = TRUE), "params")

# extract resample result for classif.rpart
rr = bmr$aggregated()[learner_id == "classif.rpart", resample_result][[1]]
print(rr)

# access the confusion matrix of the first resampling iteration
rr$experiment(1)$prediction$confusion
}
\keyword{datasets}
