% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{Container for Results of benchmark}
\description{
This is the object returned by \link{benchmark}.
}
\section{Usage}{
\preformatted{rr$experiment(iter)
rr$experiments(iters)
bmr$resample_result(i)
bmr$performance
bmr$aggregated
}
}

\section{Arguments}{

\itemize{
\item \code{i} (\code{integer}):
Counter of the \link{ResampleResult} to retrieve.
}
}

\section{Details}{

\code{experiment()} returns an \link{Experiment} for the \code{iter}-th resampling iteration.

\code{experiments()} returns a list with the slice of \link{Experiment}s for the provided \code{iters}.

\code{performance} provides a \link[data.table:data.table]{data.table::data.table} with column \code{iteration} (integer) and a numeric column for each
performance measure (columns named using the measure ids).

\code{aggregated} returns the aggregated performance measures. The aggregation method is part of the \link{Measure}.
}

