% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\docType{data}
\name{BenchmarkResult}
\alias{BenchmarkResult}
\title{Container for Results of \code{benchmark()}}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
This is the result container object returned by \code{\link[=benchmark]{benchmark()}}.
}
\section{Construction}{
\preformatted{bmr = BenchmarkResult$new(data)
}
\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table with the data of one \link{Experiment} per row.
See description of in \link{Experiment} for the exact structure.
Usually, this object is constructed via \code{\link[=benchmark]{benchmark()}}.
}
}

\section{Public}{

\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Experiment data with one \link{Experiment} per line.
\item \code{tasks} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of used tasks with three columns:
"task_hash" (\code{character(1)}), "task_id" (\code{character(1)}) and "task" (\link{Task}).
\item \code{learners} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of used learners with three columns:
"learner_hash" (\code{character(1)}), "learner_id" (\code{character(1)}) and "learner" (\link{Learner}).
\item \code{resamplings} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of used resamplings with three columns:
"resampling_hash" (\code{character(1)}), "resampling_id" (\code{character(1)}) and "resampling" (\link{Resampling}).
\item \code{measures} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of used measures with three columns:
"measure_hash" (\code{character(1)}), "measure_id" (\code{character(1)}) and "measure" (\link{Measure}).
\item \code{resample_results} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table of \link{ResampleResult} groups with 5 columns:
"hash" (\code{character(1)}), "task_id" (\code{character(1)}), "learner_id" (\code{character(1)}), "resamping_id" (\code{character(1)}) and "N" (\code{integer(1)}).
The last column ("N") is the number of \link{Experiment}s in the \link{ResampleResult}.
}
}

\section{Methods}{

\itemize{
\item \code{aggregated(objects = TRUE, ids = TRUE, params = FALSE)}\cr
(\code{logical(1)}, \code{logical(1)}, \code{logical(1)}) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a result table where experiments are aggregated per \link{ResampleResult}.
Arguments control the number of additional columns:
\itemize{
\item \code{objects} :: \code{logical(1)}\cr
Return objects as columns in the result \code{data.table()}.
\item \code{ids} :: \code{logical(1)}\cr
Return object ids as columns in the result \code{data.table()}.
\item \code{params} :: \code{logical(1)}\cr
Return learner hyperparameter values as list column \code{params} in the result \code{data.table()}.
}
\item \code{combine(bmr)}\cr
\link{BenchmarkResult} -> \code{self}\cr
Fuses a second \link{BenchmarkResult} into itself.
\item \code{get_best(measure)}\cr
\link{Measure} -> \link{ResampleResult}\cr
Returns the \link{ResampleResult} with best performance according to the provided \link{Measure}.
\item \code{resample_result(hash)}\cr
\code{character(1)} -> \link{ResampleResult}\cr
Retrieves the \link{ResampleResult} with provided \code{hash}.
}
}

\examples{
\dontshow{
   .threshold = logger::log_threshold(namespace = "mlr3")
   logger::log_threshold(logger::WARN, namespace = "mlr3")
}
set.seed(123)
bmr = benchmark(expand_grid(
  tasks = mlr_tasks$mget("iris"),
  learners = mlr_learners$mget(c("classif.featureless", "classif.rpart")),
  resamplings = mlr_resamplings$mget("cv")
))
print(bmr)
bmr$tasks
bmr$learners
bmr$resamplings
bmr$measures
bmr$aggregated(objects = FALSE)
rrs = bmr$resample_results
print(rrs)
rr = bmr$resample_result(rrs$hash[1])
print(rr)
rr$experiment(1)$model
\dontshow{
   logger::log_threshold(.threshold, namespace = "mlr3")
}
}
\keyword{datasets}
