% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallbackEvaluation.R
\name{callback_evaluation}
\alias{callback_evaluation}
\title{Create Workhorse Callback}
\usage{
callback_evaluation(
  id,
  label = NA_character_,
  man = NA_character_,
  on_evaluation_begin = NULL,
  on_evaluation_before_train = NULL,
  on_evaluation_before_predict = NULL,
  on_evaluation_end = NULL
)
}
\arguments{
\item{id}{(\code{character(1)})\cr
Identifier for the new instance.}

\item{label}{(\code{character(1)})\cr
Label for the new instance.}

\item{man}{(\code{character(1)})\cr
String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.
The referenced help package can be opened via method \verb{$help()}.}
}
\description{
Function to create a \link{CallbackEvaluation}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{Start Evaluation on Worker
 - on_evaluation_begin
 - on_evaluation_before_train
 - on_evaluation_before_predict
 - on_evaluation_end
End Evaluation on Worker
}\if{html}{\out{</div>}}
}
\details{
When implementing a callback, each function must have two arguments named \code{callback} and \code{context}.
A callback can write data to the state (\verb{$state}), e.g. settings that affect the callback itself.
Evaluation callbacks access \link{ContextEvaluation}.
}
