% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{convert_prediction}
\alias{convert_prediction}
\alias{as_prediction}
\title{Prediction Object Helpers

\code{convert_prediction()} is called on the slave to check and convert the return
value of the \code{predict()} function of a \link{Learner}.

\code{as_prediction()} is used to construct a \link{Prediction} object from the data
returned by \code{convert_prediction()}.}
\usage{
convert_prediction(task, predicted)

as_prediction(task, predicted)
}
\arguments{
\item{task}{:: \link{Task}\cr
Task as passed to \code{$predict()} of a \link{Learner}.}

\item{predicted}{:: named \code{list()}\cr
Return value of the \link{Learner}.}
}
\description{
Prediction Object Helpers

\code{convert_prediction()} is called on the slave to check and convert the return
value of the \code{predict()} function of a \link{Learner}.

\code{as_prediction()} is used to construct a \link{Prediction} object from the data
returned by \code{convert_prediction()}.
}
\keyword{internal}
