% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackend.R
\name{DataBackend}
\alias{DataBackend}
\title{DataBackend}
\description{
This is the abstract base class for data backends.
All objects of type \code{DataBackend} must provide the following interface:
}
\section{Usage}{
\preformatted{b$data(rows, cols)
b$head(n = 6)
b$distinct(col)
b$rownames
b$colnames
b$nrow
b$ncol
print(b)
}

See \link{DataBackendDataTable} for an exemplary implementation of this interface.
}

\section{Arguments}{

\itemize{
\item \code{rows} (\code{integer()} or \code{character()}):
Vector of row indices to subset rows using the primary key in the data backend.
\item \code{cols} (\code{character()}):
Vector of column names to select.
\item \code{n} (\code{integer(1)}): Number of rows to return.
}
}

\section{Details}{

\code{$data()} returns a slice of the data as \link[data.table:data.table]{data.table}:
rows are matched by the \code{primary_key} column, columns are selected by name.

\code{$head()} (\code{data.table}) returns a \link[data.table:data.table]{data.table} of the first \code{n} data rows.

\code{$distinct()} (\code{vector}) returns a named list of distinct values for selected \code{cols}.

\code{$rownames} (\code{character(1)}) returns all rownames of \code{data} as integer or character vector.

\code{$colnames} (\code{character(1)}) returns all colnames of \code{data} as character vector.

\code{$nrow} (\code{integer(1)}) returns the number of total rows.

\code{$ncol} (\code{integer(1)}) returns the number of total columns, including primary key column.
}

\seealso{
Other DataBackend: \code{\link{DataBackendDataTable}}
}
\concept{DataBackend}
