% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataBackend for data.table}
\description{
Abstraction for \code{\link[data.table:data.table]{data.table::data.table()}} as an in-memory data base.
Returns an object of class \link{DataBackend}.
}
\section{Usage}{
\preformatted{b = DataBackendDataTable$new(data, primary_key = NULL)
}
}

\section{Arguments}{

\itemize{
\item \code{data} (\code{data.frame} or \code{data.table}).
\item \code{primary_key} (\code{character(1)}): Name of the column in \code{data} which represents a unique
row identifier (as integer or character). If \code{NULL}, a new column with integer indices is
automatically created.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{DataBackend}.
}

\examples{
b = DataBackendDataTable$new(data = iris)
print(b)
b$head()
b$data(rows = 100:101, cols = "Species")

b$nrow
head(b$rownames)

b$ncol
b$colnames
}
\seealso{
Other DataBackend: \code{\link{DataBackend}}
}
\concept{DataBackend}
