% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\docType{data}
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataBackend for \code{data.table}}
\format{\link{R6Class} object inheriting from \link{DataBackend}.}
\usage{
DataBackendDataTable
}
\description{
\link{DataBackend} for \code{\link[=data.table]{data.table()}} as an in-memory data base.
}
\section{Construction}{

\itemize{
\item \code{new(data, primary_key = NULL, format = )}\cr
(\code{data.table()}, \code{character(1)}) -> \code{self}\cr
The input \code{data.table()} is not copied, and argument \code{primary_key} must be column of the \code{data}.
\item \code{as_data_backend(data, primary_key = NULL)}\cr
\code{data.frame()} -> \code{self}\cr
Converts \code{data} to \code{data.table()} (or copies it), and automatically creates a primary key column
if \code{primary_key} is `NULL.
}
}

\section{Public}{

\itemize{
\item \code{nrow} :: \code{integer(1)}\cr
Number of rows (observations).
\item \code{ncol} :: \code{integer(1)}\cr
Number of columns (variables), including the primary key column.
\item \code{colnames} :: \code{character()}\cr
Returns vector of all column names, including the primary key column.
\item \code{rownames} :: \code{integer()} | \code{character()}\cr
Returns vector of all distinct row identifiers, i.e. the primary key column.
}
}

\section{Methods}{

\itemize{
\item \code{data(rows = NULL, cols = NULL, format = "data.table")}\cr
(\code{integer()} | \code{character()}), \code{character()}) -> \code{any}\cr
Returns a slice of the data in a specific format.
Currently, the only supported format is "data.table".
The rows must be addressed as vector of primary key values, columns must be referred to via column names.
Non-existing rows and columns are silently ignored.
\item \code{distinct(cols)}\cr
(\code{character()}) -> named \code{list()}\cr
Returns a named list of vectors of distinct values for each column specified.
Non-existing columns are silently ignored.
\item \code{head(n = 6)}\cr
(\code{integer(1)}) -> \code{data.table()}\cr
Returns the first up-to \code{n} rows of the data as \code{data.table()}.
\item \code{missing(rows, cols)}\cr
(\code{integer()} | \code{character()}), \code{character()}) -> named \code{integer()}\cr
Returns the number of missing values per column in the specified slice of data.
Non-existing rows and columns are silently ignored.
}
}

\examples{
data = as.data.table(iris)
data$id = seq_len(nrow(iris))
b = DataBackendDataTable$new(data = data, primary_key = "id")
print(b)
b$head()
b$data(rows = 100:101, cols = "Species")

b$nrow
head(b$rownames)

b$ncol
b$colnames
}
\seealso{
Other DataBackend: \code{\link{DataBackendMatrix}},
  \code{\link{DataBackend}}, \code{\link{as_data_backend}}
}
\concept{DataBackend}
\keyword{datasets}
