% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataBackend for data.table}
\description{
Abstraction for \code{\link[data.table:data.table]{data.table()}} as an in-memory data base.
Returns an object of class \link{DataBackend}.
}
\section{Usage}{
\preformatted{# Construction
b = DataBackendDataTable$new(data, primary_key)
b = as_data_backend(data, primary_key = NULL)
}

The interface is described in \link{DataBackend}.
}

\section{Arguments}{

\itemize{
\item \code{data} [\code{\link[base:data.frame]{data.frame()}}].
\item \code{primary_key} [\code{character(1)}]:\cr
Name of the column in \code{data} which represents a unique
row identifier (as integer or character).
If `NULL, the constructor \code{\link[=as_data_backend]{as_data_backend()}} automatically creates an integer column of primary keys.
}
}

\examples{
data = as.data.table(iris)
data$id = seq_len(nrow(iris))
b = DataBackendDataTable$new(data = data, primary_key = "id")
print(b)
b$head()
b$data(rows = 100:101, cols = "Species")

b$nrow
head(b$rownames)

b$ncol
b$colnames
}
\seealso{
Other DataBackend: \code{\link{DataBackendMatrix}},
  \code{\link{DataBackend}}
}
\concept{DataBackend}
