% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataBackend for data.table}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{DataBackend}.}
\description{
\link{DataBackend} for \CRANpkg{data.table} as an in-memory data base.
}
\section{Construction}{
\preformatted{DataBackendDataTable$new(data, primary_key = NULL)
as_data_backend(data, primary_key = NULL, keep_rownames = FALSE, ...)
}
\itemize{
\item \code{data} :: \code{\link[=data.frame]{data.frame()}}\cr
The input \code{\link[=data.frame]{data.frame()}}.
Converted to a \code{\link[data.table:data.table]{data.table::data.table()}} automatically.
\item \code{primary_key} :: \code{character(1)}\cr
Name of the primary key column.
\item \code{keep_rownames} :: \code{logical(1)} | \code{character(1)}\cr
If \code{TRUE} or a single string, keeps the row names of \code{data} as a new column.
The column is named like the provided string, defaulting to \code{"..rownames"} for \code{keep_rownames == TRUE}.
Note that the created column will be used as a regular feature by the task unless you manually change the column role.
See also \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.
}

\code{DataBackendDataTable} does not copy the input data, while \code{as_data_backend} calls \code{\link[data.table:copy]{data.table::copy()}}.
\code{as_data_backend} creates a primary key column as integer column if \code{primary_key} is \code{NULL.}
}

\section{Fields}{

See \link{DataBackend}.
}

\section{Methods}{

See \link{DataBackend}.
}

\examples{
data = as.data.table(iris)
data$id = seq_len(nrow(iris))
b = DataBackendDataTable$new(data = data, primary_key = "id")
print(b)
b$head()
b$data(rows = 100:101, cols = "Species")

b$nrow
head(b$rownames)

b$ncol
b$colnames

# alternative construction
as_data_backend(iris)
}
\seealso{
Other DataBackend: 
\code{\link{DataBackendMatrix}},
\code{\link{DataBackend}},
\code{\link{as_data_backend}()}
}
\concept{DataBackend}
