% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\docType{data}
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataBackend for \code{data.table}}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{DataBackend}.}
\description{
\link{DataBackend} for \CRANpkg{data.table} as an in-memory data base.
}
\section{Construction}{
\preformatted{DataBackendDataTable$new(data, primary_key = NULL)
}
\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
The input \code{\link[data.table:data.table]{data.table::data.table()}} (as reference).
\item \code{primary_key} :: \code{character(1)}\cr
Name of the primary key column.
}

Alternatively, use \link{as_data_backend} on a \code{\link[data.table:data.table]{data.table::data.table()}} which will
construct a \link{DataBackend} with a copy of the data, and automatically
creates a primary key column if required.
}

\section{Fields}{

See \link{DataBackend}.
}

\section{Methods}{

See \link{DataBackend}.
}

\examples{
data = as.data.table(iris)
data$id = seq_len(nrow(iris))
b = DataBackendDataTable$new(data = data, primary_key = "id")
print(b)
b$head()
b$data(rows = 100:101, cols = "Species")

b$nrow
head(b$rownames)

b$ncol
b$colnames
}
\seealso{
Other DataBackend: \code{\link{DataBackendMatrix}},
  \code{\link{DataBackend}}, \code{\link{as_data_backend}}
}
\concept{DataBackend}
\keyword{datasets}
