% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataBackendDataTable.R
\name{DataBackendDataTable}
\alias{DataBackendDataTable}
\title{DataBackend for \code{data.table}}
\format{\link{R6Class} object inheriting from \link{DataBackend}.}
\description{
\link{DataBackend} for \code{\link[=data.table]{data.table()}} as an in-memory data base.
}
\section{Usage}{

Inherits from \link{DataBackend}.\preformatted{# Construction
b = DataBackendDataTable$new(data, primary_key)
b = as_data_backend(data, primary_key = NULL)
}

The interface is described in \link{DataBackend}.
}

\section{Arguments}{

\itemize{
\item \code{data} (\code{\link[=data.frame]{data.frame()}}). This includes special data.frames like \code{\link[=data.table]{data.table()}} or \code{tibble()} from package \pkg{tibble}.
\item \code{primary_key} (\code{character(1)}):
Name of the column in \code{data} which represents a unique row identifier (as integer or character).
If \code{NULL}, the constructor \code{\link[=as_data_backend]{as_data_backend()}} automatically creates an integer column of primary keys.
}
}

\examples{
data = as.data.table(iris)
data$id = seq_len(nrow(iris))
b = DataBackendDataTable$new(data = data, primary_key = "id")
print(b)
b$head()
b$data(rows = 100:101, cols = "Species")

b$nrow
head(b$rownames)

b$ncol
b$colnames
}
\seealso{
Other DataBackend: \code{\link{DataBackendMatrix}},
  \code{\link{DataBackend}}, \code{\link{as_data_backend}}
}
\concept{DataBackend}
