% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\docType{data}
\name{Dictionary}
\alias{Dictionary}
\title{Base Class for Dictionaries}
\format{\code{\link{R6Class}} object}
\usage{
Dictionary
}
\value{
[\code{Dictionary}].
}
\description{
A \code{\link[R6]{R6Class}} for a simple dictionary (hash map) storing R6 objects with slot \dQuote{id}.
}
\section{Fields}{

\describe{
\item{\code{ids}}{Returns the keys of registered items.}

\item{\code{length}}{Returns the number of registered items.}

\item{\code{items}}{Environment where all objects are stored.}
}}

\section{Methods}{

\describe{
 \item{\code{add(id, value)}}{...}
 \item{\code{get(id)}}{...}
 \item{\code{mget(ids)}}{...}
 \item{\code{remove(id)}}{...}
}
}

\keyword{datasets}
