% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\name{Dictionary}
\alias{Dictionary}
\title{Key-Value Storage}
\format{\link{R6Class} object.}
\description{
A simple key-value store for \link[R6:R6]{R6::R6} objects.
On retrieval of an object, the following applies:
\itemize{
\item Functions are called (with no arguments).
\item R6 Factories (objects of class \code{R6ClassGenerator}) are initialized (with no arguments).
\item \link[R6:R6Class]{R6::R6Class} objects are cloned.
\item All other objects are returned as-is.
}
}
\section{Usage}{
\preformatted{# Construction
d = Dictionary$new()

# Methods
d$add(value)
d$get(key)
d$has(keys)
d$keys(pattern)
d$mget(keys)
d$remove(key, value)
d$remove(keys)

## S3 methods
as.data.frame(d)
as.data.table(d)
}
}

\section{Arguments}{

\itemize{
\item \code{pattern} (\code{character(1)}): Restrict to ids which match the regular expression \code{pattern}.
\item \code{id} (\code{character(1)}): Single id as string.
\item \code{value}: Arbitrary value.
\item \code{ids} (\code{character()}): Vector of ids.
}
}

\section{Details}{

\itemize{
\item \code{$add()} adds item \code{value} with key \code{key} to the Dictionary.
\item \code{$get()} retrieves a single object with key \code{key} (or raises an exception).
\item \code{$has()} (\code{logical()}) returns a named logical of the same length as \code{keys} with value \code{TRUE} if the respective key is found in the Dictionary.
\item \code{$keys()} (\code{character()}) returns a vector with all keys (or all keys matching \code{pattern}).
\item \code{$mget()} (named \code{list}) creates a list of objects with keys \code{keys} (or raises an exception).
\item \code{$new()} initializes a new object of class \link{Dictionary}.
\item \code{$remove()} removes item with key \code{key} from the Dictionary.
\item \code{as.data.frame()} and \code{as.data.table()} give a summarizing overview as \code{\link[=data.frame]{data.frame()}} or \code{\link[=data.table]{data.table()}}, respectively.#'
}
}

\seealso{
Other Dictionary: \code{\link{mlr_learners}},
  \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}},
  \code{\link{mlr_task_generators}},
  \code{\link{mlr_tasks}}
}
\concept{Dictionary}
\keyword{internal}
