% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\name{Dictionary}
\alias{Dictionary}
\title{Key-value storage}
\description{
A simple key-value store for \pkg{R6} objects with support of lazy-loading objects.
}
\section{Usage}{
\preformatted{d = Dictionary$new(contains)

d$add(value)
d$get(id)
d$mget(ids)
d$remove(ids)
d$ids()
}
}

\section{Arguments}{

\itemize{
\item \code{contains} (\code{string}):
Class of objects to store. Used for assertions.
\item \code{id} (\code{string}):
Key of the object to work on.
\item \code{ids} (\code{string}):
Keys of the objects to work on.
}
}

\section{Details}{

\code{$new()} initializes a new object of class \link{Dictionary}.

\code{$get()} retrieves a single object with key \code{id} (or raises an exception).

\code{$mget()} retrieves a named list of objects with keys \code{ids} (or raises an exception).

\code{$remove()} removes item with id \code{id} from the Dictionary.

\code{$ids()} returns a vector of type \code{character} with all ids.
}

\seealso{
Other Dictionary: \code{\link{mlr_learners}},
  \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}
}
\concept{Dictionary}
\keyword{internal}
