% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\name{Dictionary}
\alias{Dictionary}
\title{Key-value storage}
\description{
A simple key-value store for \pkg{R6} objects with support of lazy-loading objects.
}
\section{Usage}{
\preformatted{d = Dictionary$new()

d$add(value)
d$get(key)
d$mget(keys)
d$remove(keys)
d$keys()
}
}

\section{Arguments}{

\itemize{
\item \code{key} (\code{string}):
Key of single object to work on.
\item \code{keys} (\code{string}):
Keys of multiple objects to work on.
}
}

\section{Details}{

\code{$new()} initializes a new object of class \link{Dictionary}.

\code{$get()} retrieves a single object with key \code{key} (or raises an exception).

\code{$mget()} retrieves a named list of objects with keys \code{keys} (or raises an exception).

\code{$remove()} removes item with key \code{key} from the Dictionary.

\code{$keys()} returns a vector of type \code{character} with all keys.
}

\seealso{
Other Dictionary: \code{\link{mlr_learners}},
  \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}
}
\concept{Dictionary}
\keyword{internal}
