% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dictionary.R
\docType{data}
\name{Dictionary}
\alias{Dictionary}
\title{Key-Value Storage}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
A simple key-value store for \link[R6:R6]{R6::R6} generator objects.
On retrieval of an object, the following applies:
\itemize{
\item R6 Factories (objects of class \code{R6ClassGenerator}) are initialized (with additional arguments).
\item Functions are called (with additional arguments) and must return an instance of a \link[R6:R6]{R6::R6} object.
}
}
\section{Construction}{
\preformatted{d = Dictionary$new()
}
}

\section{Methods}{

\itemize{
\item \code{get(key, ...)}\cr
(\code{character(1)}, ...) -> \code{any}
Retrieves object with key \code{key} from the dictionary.
\item \code{mget(keys, ...)}\cr
(\code{character()}, ...) -> named \code{list()}
Retrieves objects with keys \code{keys} from the dictionary, returns them in a list named with \code{keys}.
\item \code{has(keys)}\cr
\code{character()} -> \code{logical()}
Returns a logical vector with \code{TRUE} at its i-th position, if the i-th key exists.
\item \code{keys(pattern)}\cr
\code{character(1)} -> \code{character()}
Returns all keys which comply to the regular expression \code{pattern}.
\item \code{add(key, value)}\cr
(\code{character(1)}, \code{any}) -> \code{self}
Adds object \code{value} to the dictionary with key \code{key}, potentially overwriting a
previously stored value.
\item \code{remove(key)}\cr
\code{character()} -> \code{self}
Removes object with key \code{key} from the dictionary.
}
}

\seealso{
Other Dictionary: \code{\link{mlr_generators}},
  \code{\link{mlr_learners}}, \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}
}
\concept{Dictionary}
\keyword{datasets}
