% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec_control.R
\name{exec_control}
\alias{exec_control}
\title{Execution control object}
\usage{
exec_control(...)
}
\arguments{
\item{...}{Named arguments to overwrite the defaults / options.}
}
\description{
This function creates a named list of settings which control the execution of an \link{Experiment}.
It contains all options (without the \code{"mlr3"} prefix) and additionally:
\itemize{
\item \code{discard.model}: If \code{TRUE}, the model as returned by the learner is discarded after the experiment is finished.
Set this to save some memory. Note that you will be unable to further investigate the experiment or #'  predict on new data.
}
}
\examples{
# get a list of the currently active defaults
exec_control()

# enable debuging
exec_control(debug = TRUE)
}
