% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid}
\alias{expand_grid}
\title{Generate a Benchmark Design}
\usage{
expand_grid(tasks, learners, resamplings)
}
\arguments{
\item{tasks}{:: list of \link{Task}\cr
See also \link{mlr_sugar}.}

\item{learners}{:: list of \link{Learner}\cr
See also \link{mlr_sugar}.}

\item{resamplings}{:: list of \link{Resampling}\cr
See also \link{mlr_sugar}.}
}
\value{
(\code{\link[data.table:data.table]{data.table::data.table()}}) with the cross product of the input vectors.
}
\description{
Takes a lists of \link{Task}, a list of \link{Learner} and a list of \link{Resampling} to
generate a design in an \code{\link[=expand.grid]{expand.grid()}} fashion (a.k.a. cross join or Cartesian product).

Resampling strategies may not be instantiated, and will be instantiated per task internally.
}
