% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{expand_grid}
\alias{expand_grid}
\title{Generate a Benchmark Design}
\usage{
expand_grid(tasks, learners, resamplings)
}
\arguments{
\item{tasks}{(list of \link{Task} | \code{character()}).
Instead a \link{Task} object, it is also possible to provide a keys to retrieve tasks from the \link{mlr_tasks} dictionary.}

\item{learners}{(list of \link{Learner} | \code{character()}).
Instead if a \link{Learner} object, it is also possible to provide keys to retrieve learners from the \link{mlr_learners} dictionary.}

\item{resamplings}{(list of \link{Resampling} | \code{character()}).
Instead if a \link{Resampling} object, it is also possible to provide a key to retrieve a resampling from the \link{mlr_resamplings} dictionary.}
}
\value{
(\code{\link[=data.table]{data.table()}}) with the cross product of the input vectors.
}
\description{
Takes a lists of \link{Task}, a list of \link{Learner} and a list of \link{Resampling} to
generate a design in an \code{\link[=expand.grid]{expand.grid()}} fashion (a.k.a. cross join or Cartesian product).

Resampling strategies may not be instantiated, and will be instantiated per task internally.
}
