% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\docType{data}
\name{Experiment}
\alias{Experiment}
\title{Experiment}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Container object for a machine learning experiment.
After initialization with a \link{Task} and a \link{Learner}, the experiment is conducted
by calling the methods \code{$train()}, \code{$predict()} and \code{$score()}.
}
\section{Construction}{
\preformatted{Experiment$new(task = NULL, learner = NULL, ctrl = list())
}
\itemize{
\item \code{task} :: (\link{Task} | \code{character(1)})\cr
May be \code{NULL} during initialization, but is mandatory to train the Experiment.
Instead of a \link{Task} object, it is also possible to provide a key to retrieve a task from the \link{mlr_tasks} dictionary.
\item \code{learner} :: \link{Learner} | \code{character(1)})\cr
May be \code{NULL} during initialization, but is mandatory to train the Experiment.
Instead of a \link{Learner} object, it is also possible to provide a key to retrieve a learner from the \link{mlr_learners} dictionary.
\item \code{ctrl} :: named \code{list()}\cr
Control object, see \code{\link[=mlr_control]{mlr_control()}}.
}
}

\section{Fields}{

\itemize{
\item \code{ctrl} :: \code{list()}\cr
Control settings passed during initialization.
\item \code{data} :: named \code{list()}\cr
See section "Internal Data Storage".
\item \code{has_errors} :: \code{logical(1)}\cr
Flag which is \code{TRUE} if any error has been recorded during \code{$train()} or \code{$predict()}.
\item \code{hash} :: \code{character(1)}\cr
Hash (unique identifier) for this object.
\item \code{model} :: \code{any}\cr
Access the trained model of the \link{Learner}.
Only available after the learner has been trained.
\item \code{performance} :: named \code{numeric()}\cr
Access the scored performance scores as returned by the \link{Measure} stored in the \link{Task}.
\item \code{prediction} :: \link{Prediction}\cr
Access the individual predictions of the model stored in the \link{Learner}.
\item \code{seeds} :: \code{integer(3)}\cr
Named integer of random number generator seeds passed to \code{\link[=set.seed]{set.seed()}} prior to calling external code in \code{train()}, \code{predict()} or \code{score()}.
Names must match "train", "predict" and "score". Set to \code{NA} to disable seeding (default).
\item \code{state} :: \code{ordered(1)}\cr
Returns the state of the experiment as ordered factor with levels \code{"defined"},
\code{"trained"}, \code{"predicted"}, and \code{"scored"}.
\item \code{task} :: \link{Task}\cr
Access to the stored \link{Task}.
\item \code{test_set} :: (\code{integer()} | \code{character()})\cr
The row ids of the \link{Task} for the test set used in \code{$predict()}
Timings are \code{NA} if the respective step has not been performed yet.
\item \code{timings} :: named \code{numeric(3)}\cr
Stores the elapsed time for the steps \code{train()}, \code{predict()} and \code{score()} in seconds with up to millisecond accuracy (c.f. \code{proc.time()}).
\item \code{train_set} :: (\code{integer()} | \code{character()})\cr
The row ids of the \link{Task} for the training set used in \code{$train()}.
\item \code{validation_set} :: (\code{integer()} || \code{character()})\cr
The row ids of the validation set of the \link{Task}.
Validation sets are not yet completely integrated into the package.
}
}

\section{Methods}{

\itemize{
\item \code{train(row_ids = NULL, ctrl = list())}\cr
(\code{integer()} | \code{character()}, \code{list()}) -> \code{self}\cr
Fits the induced \link{Learner} on the \code{row_ids} of the \link{Task} and stores the model inside the \link{Learner} object.
If no \code{row_ids} are provided, trains the model on all rows of the \link{Task} with row role \code{"use"}.
The fitted model can be accessed via \code{$model}.
\item \code{predict(row_ids = NULL, newdata = NULL, ctrl = list())}\cr
(\code{integer()} | \code{character()}, \code{data.frame()}, \code{list()}) -> \code{self}\cr
Uses the previously fitted model to predict new observations.
New observations are either addressed as \code{row_ids} referencing rows in the stored task, or
as \code{data.frame()} via \code{newdata}.
The later fuses the new observations with the stored \link{Task}, and thereby mutates the Experiment.
To avoid any side effects, it is advised to clone the experiment first.
The resulting predictions are stored internally as an \link{Prediction} object and can be accessed via \code{$prediction}.
\item \code{score(measures = NULL, ctrl = list())}\cr
(list of \code{[Measure]}, \code{list()}) -> \code{self}\cr
Quantifies stored predictions using the list of \link{Measure} provided here,
defaulting to the default measures that come with the \link{Task}.
The performance values are stored internally and can be accessed via \code{$performance}.
\item \code{log(steps = c("train", "predict"))}\cr
\code{character(1)} -> \link{Log}\cr
Returns a \link{Log} for specified steps.
}
}

\section{Internal Data Storage}{

All data is stored in the slot \code{data} as named \code{list()}.
Directly accessing the elements is not recommended, but sometimes required, especially if you aim to extend \code{mlr3}.
The data object contains the following items:
\itemize{
\item \code{task} :: \link{Task}\cr
A clone of the \link{Task} which was provided during construction.
Also accessible via \code{e$task}.
\item \code{learner} :: \link{Learner}\cr
A clone of the \link{Learner} which was provided during construction.
Also accessible via \code{e$learner}.
\item \code{resampling} :: \link{Resampling}\cr
Is \code{NULL} prior to calling \code{$train()}.
If the experiment is constructed manually (i.e., not via \code{\link[=resample]{resample()}} or \code{\link[=benchmark]{benchmark()}}), a \link{ResamplingCustom} object is stored.
The combination of \code{resampling} and \code{iteration} (next item) is used to extract the training and test set indices.
These are directly accessible via \code{e$train_set} and \code{e$test_set}.
\item \code{iteration} :: \code{integer(1)}\cr
Refers to the iteration number of the stored \link{Resampling} object.
If the experiment is constructed manually, this is always \code{1}, as there is only one training-test split.
\item \code{train_log} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\link{Log} for the training step. May be \code{NULL} if no encapsulation has been enabled via \code{\link[=mlr_control]{mlr_control()}}.
\item \code{train_time} :: \code{numeric(1)}\cr
Elapsed time during train in seconds with up to millisecond accuracy (c.f. \code{proc.time()}).
\item \code{predict_log} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
\link{Log} for the predict step. May be \code{NULL} if no encapsulation has been enabled via \code{\link[=mlr_control]{mlr_control()}}.
\item \code{predict_time} :: \code{numeric(1)}\cr
Elapsed time during predict in seconds with up to millisecond accuracy (c.f. \code{proc.time()}).
\item \code{prediction} :: \link{Prediction}\cr
Prediction object as returned by the \link{Learner}'s \code{predict()} call.
\item \code{measures} :: \code{list()} of \link{Measure}\cr
Measures which where used for performance assessment.
\item \code{performance} :: named \code{numeric()}\cr
Aggregated scores returned by the measures, named with measure ids.
\item \code{score_time} :: \code{numeric(1)}\cr
Elapsed time during score in seconds with up to millisecond accuracy (c.f. \code{proc.time()})..
}
}

\examples{
e = Experiment$new(task = "iris", learner = "classif.rpart")
print(e)
e$state

e$train(row_ids = 1:120)
print(e)
e$state
e$model

e$predict(row_ids = 121:150)
print(e)
e$state
e$prediction

e$score()
print(e)
e$state
e$performance

e$train_set
e$test_set
}
\keyword{datasets}
