% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{Experiment}
\alias{Experiment}
\title{Experiment}
\format{\link{R6Class} object.}
\description{
Container object for machine learning experiments.
}
\section{Usage}{
\preformatted{# Construction
e = Experiment$new(task, learner, ...)

# Members
e$ctrl
e$data
e$has_errors
e$hash
e$learner
e$logs
e$model
e$performance
e$prediction
e$seeds
e$state
e$task
e$test_set
e$timings
e$train_set
e$validation_set

# Methods
e$predict(row_ids, newdata, ctrl = list())
e$score(ctrl = list())
e$train(row_ids, ctrl = list())
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\link{Task}): Task to conduct experiment on
\item \code{learner} (\link{Learner}): Learner to conduct experiment with.
\item \code{row_ids} (\code{integer()} | \code{character()}): Subset of the task's row ids to work on. Invalid row ids are silently ignored.
\item \code{newdata} (\link{data.frame}): New data to predict on. Will be appended to the task.
}
}

\section{Details}{

\itemize{
\item \code{$new()} initializes a new machine learning experiment which can grow in a stepwise fashion.
\item \code{$predict()} uses the previously fitted model to predict new observations.
The predictions are stored internally as an \link{Prediction} object and can be
accessed via \code{e$prediction} as \code{\link[=data.table]{data.table()}}.
\item \code{$score()} quantifies stored predictions using the task's \link{Measure} and stores the resulting
performance values. The performance can be accessed via \code{e$performance}.
\item \code{$train()} fits the induced \link{Learner} on the (subset of the) \code{task} and'stores the model in the \link{Learner}. The model can be accessed via \code{e$model}.
\item \code{$ctrl} (\link{list}). List of control settings passed to \code{$train()}, \code{$predict()} and \code{$score()}.
\item \code{$data} stores the internal representation of an Experiment as a \code{named list} with the following slots:
\itemize{
\item iteration (\code{integer(1)}). If the experiment is constructed manually, this is always 1.
\item learner (\link{Learner}).
\item measures (\code{list} of \link{Measure}). Measures used for performance assessment.
\item performance (\code{named numeric}). Performance values returned by the measures.
\item predict_log. \link{Log} for the predict step.
\item predict_time (\code{numeric(1)}). Elapsed time in microseconds.
\item prediction (\link{Prediction}).
\item resampling (\link{Resampling}). Is \code{NULL} prior to calling \code{$train()}. If the experiment is constructed manually (i.e., not via \code{\link[=resample]{resample()}} or \code{\link[=benchmark]{benchmark()}}, a \code{ResamplingCustom} object is stored here.
\item score_time (\code{numeric(1)}). Elapsed time in microseconds.
\item task (\link{Task}).
\item train_log: \link{Log} for the training step.
\item train_time (\code{numeric(1)}). Elapsed time in microseconds.
}
\item \code{$has_errors} (\code{logical(1)}). Whether the Experiment showed errors either during training or prediction.
\item \code{$hash} (\code{character(1)}). The hash of the experiment.
\item \code{$logs} (named \code{list(2)}) returns a list with names \code{train} and \code{predict}.
Both store an object of class \link{Log} if logging of the learner has been
enabled via \code{\link[=mlr_control]{mlr_control()}}, and are \code{NULL} if logging was disabled or the
respective step has not been performed yet.
\item \code{$seeds} (\code{named numeric(3)}) stores seeds which are set prior to calling external code in train/predict/score.
Names must match \code{"train"}, \code{"predict"} and \code{"score"}. Set to \code{NA} to not set a specific seed (default).
\item \code{$state} (\code{ordered(1)}) returns the state of the experiment: \code{"defined"},
\code{"trained"}, \code{"predicted"}, or \code{"scored"}.
\item \code{$task} and \code{$learner} can be used to access the \link{Task} and \link{Learner}.
\item \code{$timings} (named \code{numeric(3)}) holds the elapsed time for the steps
\code{train}, \code{predict} and \code{score} in seconds with up to millisecond accuracy
(c.f. \code{\link[=proc.time]{proc.time()}}). Timings are \code{NA} if the respective step has not been
performed yet.
\item \code{$train_set} and \code{$test_set} (\code{integer()} | \code{character()}) return the row
ids of the training set or test set, respectively.
\item \code{$validation_set} (\code{integer()} | \code{character()}) returns the row ids of the
validation set (see \link{Task}).
}
}

\examples{
e = Experiment$new(
  task = mlr_tasks$get("iris"),
  learner = mlr_learners$get("classif.rpart")
)
print(e)
e$state

e$train(row_ids = 1:120)
print(e)
e$state
e$model

e$predict(row_ids = 121:150)
print(e)
e$state
e$prediction

e$score()
print(e)
e$state
e$performance

e$train_set
e$test_set
}
