% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\docType{data}
\name{Experiment}
\alias{Experiment}
\title{Experiment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Experiment
}
\description{
Container object for machine learning experiments.
}
\examples{
e = Experiment$new(
  task = mlr_tasks$get("iris"),
  learner = mlr_learners$get("classif.rpart")
)
print(e)
e$train()
print(e)
e$predict()
print(e)
e$score()
print(e)
}
\keyword{datasets}
