% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\docType{data}
\name{Experiment}
\alias{Experiment}
\title{Experiment}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Container object for a machine learning experiment.
}
\section{Construction}{
\preformatted{Experiment$new(task = NULL, learner = NULL, ctrl = list())
}
\itemize{
\item \code{task} :: \link{Task}\cr
May be \code{NULL} during initialization, but is mandatory to train the Experiment.
Instead if a \link{Task} object, it is also possible to provide a key to retrieve a task from the \link{mlr_tasks} dictionary.
\item \code{learner} :: \link{Learner}\cr
May be \code{NULL} during initialization, but is mandatory to train the Experiment.
Instead if a \link{Learner} object, it is also possible to provide a key to retrieve a task from the \link{mlr_learners} dictionary.
\item \code{ctrl} :: named \code{list()}\cr
Control object, see \code{\link[=mlr_control]{mlr_control()}}.
}
}

\section{Fields}{

\itemize{
\item \code{data} :: named \code{list()}\cr
Internal data storage as a \code{named list} with the following slots:
\itemize{
\item \code{iteration} :: \code{integer(1)}\cr
Refers to the iteration number of the stored \link{Resampling} object.
If the experiment is constructed manually, this is always \code{1}, as there is only one training-test split.
\item \code{learner} :: \link{Learner}\cr
A clone of the \link{Learner} provided during construction.
\item \code{measures} :: \code{list()} of \link{Measure}\cr
Measures for performance assessment.
\item \code{performance} :: named \code{numeric()}\cr
Aggregated scores returned by the measures, named with measure ids.
\item \code{prediction} :: \link{Prediction}\cr
Prediction object as returned by the \link{Learner}'s \code{predict()} call.
\item \code{resampling} :: \link{Resampling}\cr
Is \code{NULL} prior to calling \code{$train()}.
If the experiment is constructed manually (i.e., not via \code{\link[=resample]{resample()}} or \code{\link[=benchmark]{benchmark()}}), a \link{ResamplingCustom} object is stored.
\item \code{task} :: \link{Task}\cr
A clone of the \link{Task} provided during construction.
\item \code{train_log} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Log for the training step.
\item \code{predict_log} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Log for the predict step.
\item \code{train_time} :: \code{numeric(1)}\cr
Elapsed time during train in seconds.
\item \code{predict_time} :: \code{numeric(1)}\cr
Elapsed time during predict in seconds.
\item \code{score_time} :: \code{numeric(1)}\cr
Elapsed time during score in seconds.
}
\item \code{ctrl} :: \code{list()}\cr
Control settings passed during initialization.
\item \code{has_errors} :: \code{logical(1)}\cr
Flag which is \code{TRUE} if any error has been recorded during \code{$train()} or \code{$predict()}.
\item \code{hash} :: \code{character(1)}\cr
Hash (unique identifier) for this object. This hash is cached.
\item \code{state} :: \code{ordered(1)}\cr
Returns the state of the experiment as ordered factor: "defined", "trained", "predicted", or "scored".
\item \code{train_set} :: (\code{integer()} | \code{character()})\cr
The row ids of the training set for \code{$train()}.
\item \code{test_set} :: (\code{integer()} | \code{character()})\cr
The row ids of the test set for \code{$predict()}
\item \code{learner} :: \link{Learner}\cr
Access the stored \link{Learner}.
\item \code{model} :: \code{any}\cr
Access the trained model of the \link{Learner}.
\item \code{performance} :: \code{numeric()}\cr
Access the scored performance scores as returned by the \link{Measure} stored in the \link{Task}.
\item \code{prediction} :: \link{Prediction}\cr
Access the individual predictions of the model stored in the \link{Learner}.
\item \code{seeds} :: \code{integer(3)}\cr
Named integer of random number generator seeds passed to \code{\link[=set.seed]{set.seed()}} prior to calling external code in \code{train()}, \code{predict()} or \code{score()}.
Names must match "train", "predict" and "score". Set to \code{NA} to disable seeding (default).
\item \code{task} :: \link{Task}\cr
Access to the stored \link{Task}.
\item \code{timings} :: \code{numeric(3)}\cr
Stores the elapsed time for the steps \code{train()}, \code{predict()} and \code{score()} in seconds with up to millisecond accuracy (c.f. \code{proc.time()}).
Timings are \code{NA} if the respective step has not been performed yet.
\item \code{validation_set} :: (\code{integer()} || \code{character()})\cr
The row ids of the validation set of the \link{Task}.
}
}

\section{Methods}{

\itemize{
\item \code{train(row_ids = NULL, ctrl = list())}\cr
(\code{integer()} | \code{character()}, \code{list()}) -> \code{self}\cr
Fits the induced \link{Learner} on the \code{row_ids} of the \link{Task} and stores the model inside the \link{Learner} object.
The model can be accessed via \code{$model}.
\item \code{predict(row_ids = NULL, newdata = NULL, ctrl = list())}\cr
(\code{integer()} | \code{character()}, \code{data.frame()}, \code{list()}) -> \code{self}\cr
Uses the previously trained model to predict new observations.
New observations are either addressed as \code{row_ids} of the stored task, or
you can pass new observations as \code{data.frame()}.
Note that predicting on new data fuses the new observations into the \link{Task} first, and thereby
mutates the Experiment. To avoid any side effects, it is advised to clone the experiment first.
The resulting predictions are stored internally as an \link{Prediction} object and can be
accessed via \code{$prediction}.
\item \code{score(measures = NULL, ctrl = list())}\cr
(list of \code{[Measure]}, \code{list()}) -> \code{self}\cr
Quantifies stored predictions using the list of \link{Measure} provided here,
defaulting to the measures provided in the \link{Task}.
The performance is stored internally and can be accessed via \code{$performance}.
\item \code{log(steps = c("train", "predict"))}\cr
\code{character(1)} -> \link{Log}\cr
Returns a \link{Log} for specified steps.
}
}

\examples{
e = Experiment$new(task = "iris", learner = "classif.rpart")
print(e)
e$state

e$train(row_ids = 1:120)
print(e)
e$state
e$model

e$predict(row_ids = 121:150)
print(e)
e$state
e$prediction

e$score()
print(e)
e$state
e$performance

e$train_set
e$test_set
}
\keyword{datasets}
