% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faggregate.R
\name{faggregate}
\alias{faggregate}
\title{Fast Aggregation of ResampleResults and BenchmarkResults}
\usage{
faggregate(obj, measure)
}
\arguments{
\item{obj}{(\link{ResampleResult} | \link{BenchmarkResult}).}

\item{measure}{(\link{Measure}).}
}
\description{
Aggregates a \link{BenchmarkResult} for a single simple measure.
Returns the aggregated score for each uhash.
}
\details{
This function is faster than \verb{$aggregate()} because it does not reassemble the resampling results.
It also does not require the task, learner, model or train set to be available.
It only works for simple measures.
}
