% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hash_resample_result}
\alias{hash_resample_result}
\title{Calculate the Hash for a ResampleResult}
\usage{
hash_resample_result(task, learner, resampling)
}
\arguments{
\item{task}{:: \link{Task}.}

\item{learner}{:: \link{Learner}.}

\item{resampling}{:: \link{Resampling}.}
}
\value{
(\code{character(1)}) hash value.
}
\description{
Internally used, exported for addon-packages.
}
\keyword{internal}
