% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HotStartStack.R
\name{HotStartStack}
\alias{HotStartStack}
\title{Stack for Hot Start Learners}
\description{
This class stores \link{Learner}s for hot starting other learners.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{stack}}{\code{\link[data.table:data.table]{data.table::data.table()}}\cr
Stores hot start learners.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{HotStartStack$new()}}
\item \href{#method-add}{\code{HotStartStack$add()}}
\item \href{#method-adaption_cost}{\code{HotStartStack$adaption_cost()}}
\item \href{#method-adaption_learner}{\code{HotStartStack$adaption_learner()}}
\item \href{#method-clone}{\code{HotStartStack$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotStartStack$new(learners)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learners}}{(List of \link{Learner}s).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
Add learners to hot start stack.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotStartStack$add(learners)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learners}}{(List of \link{Learner}s).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-adaption_cost"></a>}}
\if{latex}{\out{\hypertarget{method-adaption_cost}{}}}
\subsection{Method \code{adaption_cost()}}{
Calculates costs of hot starting \code{learner}. Returns \code{NA} if learner is
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotStartStack$adaption_cost(learner, task_hash, stack = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learner}}{\link{Learner}.}

\item{\code{task_hash}}{\link{Task}.}

\item{\code{stack}}{\code{\link[data.table:data.table]{data.table::data.table()}}\cr
Optionally, prefiltered stack. Must only contain learners of the same
type as \code{learner}. Quickens calculation of costs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-adaption_learner"></a>}}
\if{latex}{\out{\hypertarget{method-adaption_learner}{}}}
\subsection{Method \code{adaption_learner()}}{
Returns learner with the lowest cost of hot starting \code{learner}. If no
learner is found, returns \code{NULL}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotStartStack$adaption_learner(learner, task_hash)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{learner}}{\link{Learner}.}

\item{\code{task_hash}}{\link{Task}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{HotStartStack$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
