% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkgs.R
\name{install_pkgs}
\alias{install_pkgs}
\alias{extract_pkgs}
\alias{extract_pkgs.character}
\title{Install (missing) packages}
\usage{
install_pkgs(x, update = TRUE, force = FALSE)

extract_pkgs(x)

\method{extract_pkgs}{character}(x)
}
\arguments{
\item{x}{(any)\cr
Object with package information (or a list of such objects).}

\item{update}{(\code{logical(1)})\cr
Update packages which are already installed?}

\item{force}{(\code{logical(1)})\cr
Force installation, even if the package is already installed.}
}
\description{
Install required packages.
Extracts the package information from various objects, including
\link{TaskGenerator}, \link{Learner}, \link{Measure} and objects from
extension packages such as \CRANpkg{mlr3pipelines} or \CRANpkg{mlr3filters}.

If provided a list, the function is called recursively on all elements.
}
