% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner.R
\name{Learner}
\alias{Learner}
\title{Learner}
\description{
This is the abstract base class for learner objects like \link{LearnerClassif} and \link{LearnerRegr}.

Learners consist of the following parts:
\itemize{
\item Methods \code{train()} and \code{predict()} which call \code{train_internal()} and \code{predict_internal()}.
\item The fitted model, after calling \code{train()}.
\item A \link[paradox:ParamSet]{paradox::ParamSet} which stores meta-information about available hyperparameters, and also stores hyperparameter settings.
\item Meta-information about the requirements and capabilities of the learner.
}

Predefined learners are stored in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link{mlr_learners},
e.g. \code{\link[=mlr_learners_classif.rpart]{classif.rpart}} or \code{\link[=mlr_learners_regr.rpart]{regr.rpart}}.
A guide on how to extend \CRANpkg{mlr3} with custom learners can be found in the \href{https://mlr3book.mlr-org.com}{mlr3book}.
}
\section{Optional Extractors}{


Specific learner implementations are free to implement additional getters to ease the access of certain parts
of the model in the inherited subclasses.

For the following operations, extractors are standardized:
\itemize{
\item \code{importance(...)}: Returns the feature importance score as numeric vector.
The higher the score, the more important the variable.
The returned vector is named with feature names and sorted in decreasing order.
Note that the model might omit features it has not used at all.
The learner must be tagged with property \code{"importance"}.
To filter variables using the importance scores, use package \CRANpkg{mlr3filters}.
\item \code{selected_features(...)}: Returns a subset of selected features as \code{character()}.
The learner must be tagged with property \code{"selected_features"}.
\item \code{oob_error(...)}: Returns the out-of-bag error of the model as \code{numeric(1)}.
The learner must be tagged with property \code{"oob_error"}.
}
}

\section{Setting Hyperparameters}{


All information about hyperparameters is stored in the slot \code{param_set} which is a [paradox::ParamSet].
The printer gives an overview about the ids of available hyperparameters, their storage type, lower and upper bounds,
possible levels (for factors), default values and assigned values.
To set hyperparameters, assign a named list to the subslot \code{values}:\preformatted{lrn = lrn("classif.rpart")
lrn$param_set$values = list(minsplit = 3, cp = 0.01)
}

Note that this operation replaces all previously set hyperparameter values.
If you only intend to change one specific hyperparameter value and leave the others as-is, you can use the helper function [mlr3misc::insert_named()]:\preformatted{lrn$param_set$values = mlr3misc::insert_named(lrn$param_set$values, list(cp = 0.001))
}

If the learner has additional hyperparameters which are not encoded in the [ParamSet][paradox::ParamSet], you can easily extend the learner.
Here, we add a hyperparameter with id \code{"foo"} possible levels \code{"a"} and \code{"b"}:\preformatted{lrn$param_set$add(paradox::ParamFct$new("foo", levels = c("a", "b")))


[paradox::ParamSet]: R:paradox::ParamSet
[mlr3misc::insert_named()]: R:mlr3misc::insert_named()
[paradox::ParamSet]: R:paradox::ParamSet
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{\code{character(1)} :: Identifier of the learner.}

\item{\code{state}}{\code{NULL} | \code{named [list()]} :: Current (internal) state of the learner.
Contains all information learnt during \code{train()} and \code{predict()}.
Do not access elements from here directly.}

\item{\code{task_type}}{\code{character(1)} :: Stores the type of class this learner can operate on, e.g. \code{"classif"} or \code{"regr"}.
A complete list of task types is stored in \code{\link[=mlr_reflections]{mlr_reflections$task_types}}.}

\item{\code{predict_types}}{\code{character(0)} :: Stores the possible predict types the learner is capable of.
A complete list of candidate predict types, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$learner_predict_types}}.}

\item{\code{feature_types}}{\code{character()} :: Stores the feature types the learner can handle, e.g. \code{"logical"}, \code{"numeric"}, or \code{"factor"}.
A complete list of candidate feature types, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$task_feature_types}}.}

\item{\code{properties}}{\code{character()} :: Stores a set of properties/capabilities the learner has.
A complete list of candidate properties, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$learner_properties}}.}

\item{\code{data_formats}}{\code{character()} :: Vector of supported data formats which can be processed during \verb{$train()} and \verb{$predict()}.
Defaults to \code{"data.table"}.}

\item{\code{packages}}{\code{character()} :: Stores the names of required packages.}

\item{\code{predict_sets}}{\code{character()} :: Sets to predict on during \code{\link[=resample]{resample()}}/\code{\link[=benchmark]{benchmark()}}.
Creates and stores a separate \link{Prediction} object for each set.
The individual sets can be combined via getters in \link{ResampleResult}/\link{BenchmarkResult}, or \link{Measure}s can be set to operate on subsets of the calculated prediction sets.
Must be a non-empty subset of \verb{("train", "test")}.
Default is \code{"test"}.}

\item{\code{fallback}}{\code{mlr3::Learner} :: Learner which is fitted to impute predictions in case that either the model fitting or the prediction of the top learner is not successful.
Requires you to enable encapsulation, otherwise errors are not caught and the execution is terminated before the fallback learner kicks in.}

\item{\code{man}}{\code{character(1)} :: String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{model}}{\code{any} :: The fitted model. Only available after \verb{$train()} has been called.}

\item{\code{timings}}{\code{numeric(2)} :: Elapsed time in seconds for the steps \code{"train"} and \code{"predict"}.
Measured via \code{\link[mlr3misc:encapsulate]{mlr3misc::encapsulate()}}.}

\item{\code{log}}{\verb{[data.table::data.table()]} :: Returns the output (including warning and errors) as table with columns
\code{"stage"} (train or predict), \code{"class"} (output, warning, error) and
\code{"msg"} (\code{character()}).}

\item{\code{warnings}}{\code{character()} :: Returns the logged warnings as vector.}

\item{\code{errors}}{\code{character()} :: Returns the logged errors as vector.}

\item{\code{hash}}{\code{character(1)} :: Hash (unique identifier) for this object.}

\item{\code{predict_type}}{\code{character(1)} :: Stores the currently selected predict type. Must be an element of \code{l$predict_types}.}

\item{\code{param_set}}{\verb{[paradox::ParamSet]} :: Description of available hyperparameters and hyperparameter settings.}

\item{\code{encapsulate}}{\verb{named character()} :: How to call the code in \code{train_internal()} and \code{predict_internal()}.
Must be a named character vector with names \code{"train"} and \code{"predict"}.
Possible values are \code{"none"}, \code{"evaluate"} and \code{"callr"}.
See \code{\link[mlr3misc:encapsulate]{mlr3misc::encapsulate()}} for more details.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Learner$new()}}
\item \href{#method-help}{\code{Learner$help()}}
\item \href{#method-format}{\code{Learner$format()}}
\item \href{#method-print}{\code{Learner$print()}}
\item \href{#method-train}{\code{Learner$train()}}
\item \href{#method-predict}{\code{Learner$predict()}}
\item \href{#method-predict_newdata}{\code{Learner$predict_newdata()}}
\item \href{#method-clone}{\code{Learner$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
This object is typically constructed via a derived classes, e.g. \link{LearnerClassif} or \link{LearnerRegr}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$new(
  id,
  task_type,
  param_set = ParamSet$new(),
  predict_types = character(),
  feature_types = character(),
  properties = character(),
  data_formats = "data.table",
  packages = character(),
  man = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{\code{character(1)} :: Identifier of the learner.}

\item{\code{task_type}}{\code{character(1)} :: Stores the type of class this learner can operate on, e.g. \code{"classif"} or \code{"regr"}.
A complete list of task types is stored in \code{\link[=mlr_reflections]{mlr_reflections$task_types}}.}

\item{\code{param_set}}{\verb{[paradox::ParamSet]} :: Description of available hyperparameters and hyperparameter settings.}

\item{\code{predict_types}}{\code{character()} :: Stores the possible predict types the learner is capable of.
A complete list of candidate predict types, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$learner_predict_types}}.}

\item{\code{feature_types}}{\code{character()} :: Stores the feature types the learner can handle, e.g. \code{"logical"}, \code{"numeric"}, or \code{"factor"}.
A complete list of candidate feature types, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$task_feature_types}}.}

\item{\code{properties}}{\code{character()} :: Stores a set of properties/capabilities the learner has.
A complete list of candidate properties, grouped by task type, is stored in \code{\link[=mlr_reflections]{mlr_reflections$learner_properties}}.}

\item{\code{data_formats}}{\code{character()} :: Vector of supported data formats which can be processed during \verb{$train()} and \verb{$predict()}.
Defaults to \code{"data.table"}.}

\item{\code{packages}}{\code{character()} :: Stores the names of required packages.}

\item{\code{man}}{\code{character(1)} :: String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-help"></a>}}
\subsection{Method \code{help()}}{
Opens the corresponding help page referenced by \verb{$man}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$help()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format"></a>}}
\subsection{Method \code{format()}}{
Returns a string representing this learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Prints this learner.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
Train the learner on the row ids of the provided \link{Task}.
Mutates the learner by reference, i.e. stores the model alongside other objects in field \verb{$state}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$train(task, row_ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{The data to use for training.}

\item{\code{row_ids}}{A set of row ids to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
Uses the data stored during \verb{$train()} in \verb{$state} to create a new \link{Prediction} based on the provided \code{row_ids}
of the \code{task}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$predict(task, row_ids = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{The data to predict for.}

\item{\code{row_ids}}{A set of row ids to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The predicted data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_newdata"></a>}}
\subsection{Method \code{predict_newdata()}}{
Uses the model fitted during \verb{$train()} in to create a new \link{Prediction} based on the new data in \code{newdata}.
Object \code{task} is the task used during \verb{$train()} and required for conversions of \code{newdata}.
If the learner's \verb{$train()} method has been called, there is a (size reduced) version of the training task stored in the learner.
If the learner has been fitted via \code{\link[=resample]{resample()}} or \code{\link[=benchmark]{benchmark()}}, you need to pass the corresponding task stored
in the \link{ResampleResult} or \link{BenchmarkResult}, respectively.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$predict_newdata(newdata, task = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{The new data to predict values for.}

\item{\code{task}}{The task used in the \verb{$train()} method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The predicted data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Learner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
