% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner.R
\docType{data}
\name{Learner}
\alias{Learner}
\title{Class for Learners}
\format{\code{\link[=R6Class]{R6Class()}} object}
\usage{
Learner
}
\value{
[\code{\link[=Learner]{Learner()}}].
}
\description{
A \code{\link[R6:R6Class]{R6::R6Class()}} to construct learners.
}
\seealso{
Other Learner: \code{\link{LearnerClassif}},
  \code{\link{LearnerRegr}}
}
\concept{Learner}
\keyword{datasets}
