% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner.R
\name{Learner}
\alias{Learner}
\title{Abstract learner class}
\description{
Abstraction for learners.

Predefined learners are stored in \link{mlr_learners}.
}
\section{Usage}{
\preformatted{l = Learner$new(id, packages = character(0), par_set = ParamSet$new(), par_vals = list(), properties = character(0))
l$id
l$packages
l$par_set
l$par_vals
l$properties
l$predict_type
l$train(task)
l$predict(task, model)
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\link{Task}):
Task to train/predict on.
\item \code{model} (any):
Fitted model as returned by \code{train}.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{Learner}.

\code{$id} (\code{character(1)}) stores the identifier of the object.

\code{$packages} (\code{character(1)}) stores the names of required packages.

\code{$par_set()} (\link[paradox:ParamSet]{paradox::ParamSet}) describes the available hyperparameter and possible settings.

\code{$par_vals()} (\code{named list}) stores the list set hyperparameter values.

\code{$properties} (\code{character()}) is a set of tags which describe the properties of the learner.

\code{$predict_type} (\code{character(1)}) stores the predict type of the learner,
e.g. "response" or "probability" for classification learners of class \link{LearnerClassif}.

\code{$train()} takes a task and returns a model fitted on all observations.

\code{$predict()} takes a task and the model fitted in \code{$train()} to return predicted labels.
}

\seealso{
Other Learner: \code{\link{LearnerClassif}},
  \code{\link{LearnerRegr}}
}
\concept{Learner}
\keyword{internal}
