% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassif.R
\name{LearnerClassif}
\alias{LearnerClassif}
\title{Classification Learner}
\arguments{
\item{id}{Identifier for the learner.}

\item{param_Set}{Set of hyperparameters.}

\item{predict_types}{Supported predict types. Must be a subset of \code{\link[=mlr_reflections]{mlr_reflections$learner_predict_types}}.
Possible values for \code{predict_types} are passed to and converted by \link{PredictionClassif}:
\itemize{
\item \code{"response"}: Predicts a class label for each observation in the test set.
\item \code{"prob"}: Predicts the posterior probability for each class for each observation in the test set.
}}

\item{feature_types}{Feature types the learner operates on. Must be a subset of \code{\link[=mlr_reflections]{mlr_reflections$task_feature_types}}.}

\item{properties}{Set of properties of the learner. Must be a subset of \code{\link[=mlr_reflections]{mlr_reflections$learner_properties}}.
The following properties are currently standardized and understood by learners in \CRANpkg{mlr3}:
\itemize{
\item \code{"missings"}: The learner can handle missing values in the data.
\item \code{"weights"}: The learner supports observation weights.
\item \code{"importance"}: The learner supports extraction of importance scores, i.e. comes with a \code{importance()} extractor function (see section on optional extractors).
\item \code{"selected_features"}: The learner supports extraction of the set of selected features, i.e. comes with a \code{selected_features()} extractor function (see section on optional extractors).
\item \code{"oob_error"}: The learner supports extraction of estimated out of bag error, i.e. comes with a \code{oob_error()} extractor function (see section on optional extractors).
}}

\item{data_formats}{Vector of supported data formats which can be processed during \verb{$train()} and \verb{$predict()}. Defaults to \code{"data.table"}.}

\item{packages}{Set of required packages. Note that these packages will be loaded via \code{\link[=requireNamespace]{requireNamespace()}}, and are not attached.}

\item{man}{String in the format \verb{[pkg]::[topic]} pointing to a manual page for this object.}
}
\value{
An object representing the learner.
}
\description{
This Learner specializes \link{Learner} for classification problems.
Many predefined learners can be found in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link{mlr_learners} after loading the \CRANpkg{mlr3learners} package.

For a description of the arguments, see \link{Learner}.
\code{task_type} is set to \code{"classif"}.

Additional learner properties include:
\itemize{
\item \code{"twoclass"}: The learner works on binary classification problems.
\item \code{"multiclass"}: The learner works on multiclass classification problems.
}
}
\examples{
# get all classification learners from mlr_learners:
lrns = mlr_learners$mget(mlr_learners$keys("^classif"))
names(lrns)

# get a specific learner from mlr_learners:
lrn = lrn("classif.rpart")
print(lrn)

# train the learner:
task = tsk("iris")
lrn$train(task, 1:120)

# predict on new observations:
lrn$predict(task, 121:150)$confusion

}
\seealso{
Example classification learners: \link{LearnerClassifRpart}
}
