% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassif.R
\name{LearnerClassif}
\alias{LearnerClassif}
\title{Classification Learner}
\format{\link{R6Class} object}
\description{
This Learner specializes \link{Learner} for classification problems.
}
\section{Usage}{

See \link{Learner}.
}

\section{Details}{

\itemize{
\item \code{$task_type} is \code{"classif"}.
}
}

\seealso{
Other Learner: \code{\link{LearnerRegr}},
  \code{\link{Learner}}, \code{\link{mlr_learners}}
}
\concept{Learner}
