% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassif.R
\name{LearnerClassif}
\alias{LearnerClassif}
\title{Classification Learner}
\format{\link{R6Class} object}
\description{
This Learner specializes \link{Learner} for classification problems.
}
\section{Usage}{

Inherits from \link{Learner}.
}

\section{Details}{

\itemize{
\item \code{$task_type} is \code{"classif"}.
}
}

\examples{
# get all classification learners from mlr_learners:
lrns = mlr_learners$mget(mlr_learners$ids("^classif"))
names(lrns)

# get a specific learner from mlr_learners:
lrn = mlr_learners$get("classif.rpart")
print(lrn)
}
\seealso{
Other Learner: \code{\link{LearnerRegr}},
  \code{\link{Learner}}, \code{\link{mlr_learners}}
}
\concept{Learner}
