% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassif.R
\name{LearnerClassif}
\alias{LearnerClassif}
\title{Class for classification learners}
\description{
Class for classification learners. Inherits from \link{Learner}.
Predict type is set to "response" by default.

Instantiated learners can be retrieved from \link{mlr_learners}.
}
\section{Usage}{
\preformatted{l = LearnerClassif$new()
}
}

\seealso{
Other Learner: \code{\link{LearnerRegr}},
  \code{\link{Learner}}
}
\concept{Learner}
