% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassif.R
\docType{data}
\name{LearnerClassif}
\alias{LearnerClassif}
\title{Class for Classif Learners}
\format{\code{\link[=R6Class]{R6Class()}} object}
\usage{
LearnerClassif
}
\value{
\code{\link[=Learner]{Learner()}}.
}
\description{
A \code{\link[R6:R6Class]{R6::R6Class()}} to construct learners.
}
\seealso{
Other Learner: \code{\link{LearnerRegr}}
}
\concept{Learner}
\keyword{datasets}
