% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegr.R
\name{LearnerRegr}
\alias{LearnerRegr}
\title{Class for regression learners}
\description{
Class for regression learners. Inherits from \link{Learner}.
Predict type is set to "response" by default.

Instantiated learners can be retrieved from \link{mlr_learners}.
}
\section{Usage}{
\preformatted{l = LearnerRegr$new()
}
}

\seealso{
Other Learner: \code{\link{LearnerClassif}},
  \code{\link{Learner}}
}
\concept{Learner}
