% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegr.R
\docType{data}
\name{LearnerRegr}
\alias{LearnerRegr}
\title{Class for Regr Learners}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LearnerRegr
}
\value{
\code{\link[=Learner]{Learner()}}.
}
\description{
A \code{\link[R6:R6Class]{R6::R6Class()}} to construct learners.
}
\seealso{
Other Learner: \code{\link{LearnerClassif}},
  \code{\link{Learner}}
}
\concept{Learner}
\keyword{datasets}
