% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegr.R
\name{LearnerRegr}
\alias{LearnerRegr}
\title{Regression Learner}
\format{\link{R6Class} object}
\description{
This Learner specializes \link{Learner} for regression problems.
}
\section{Usage}{

See \link{Learner}.
}

\section{Details}{

\itemize{
\item \code{$task_type} is \code{"regr"}.
}
}

\seealso{
Other Learner: \code{\link{LearnerClassif}},
  \code{\link{Learner}}, \code{\link{mlr_learners}}
}
\concept{Learner}
