% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrRpart.R
\name{LearnerRegrRpart}
\alias{LearnerRegrRpart}
\alias{mlr_learners_regr.rpart}
\title{Regression Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\description{
A \link{LearnerRegr} for a regression tree implemented in \code{\link[rpart:rpart]{rpart::rpart()}} in package \CRANpkg{rpart}.
Parameter \code{xval} is set to 0 in order to save some computation time.
}
\section{Construction}{
\preformatted{LearnerRegrRpart$new()
mlr_learners$get("regr.rpart")
lrn("regr.rpart")
}
}

\references{
\cite{mlr3}{breiman_2017}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Learner]{Learners}: \link{mlr_learners}

\code{as.data.table(mlr_learners)} for a complete table of all (also dynamically created) \link{Learner} implementations.
}
