% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrRpart.R
\docType{data}
\name{LearnerRegrRpart}
\alias{LearnerRegrRpart}
\alias{mlr_learners_regr.rpart}
\title{Regression Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrRpart
}
\description{
A \link{LearnerRegr} for a regression tree implemented in \code{\link[rpart:rpart]{rpart::rpart()}} in package \CRANpkg{rpart}.
Parameter \code{xval} is set to 0 in order to save some computation time.
}
\references{
Breiman, L. (1984).
Classification and Regression Trees.
New York: Routledge.
\doi{10.1201/9781315139470}.
}
\keyword{datasets}
