% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log.R
\name{Log}
\alias{Log}
\title{Learner Output Log}
\format{\link{R6Class} object}
\description{
Object which stores the output of the \code{train} or \code{predict} step of an \link{Experiment}.
}
\section{Usage}{
\preformatted{# Construction
l = Log$new(log = NULL)

# Members
l$errors
l$log
l$warnings

# Methods
l$format()
l$has_condition(cl)
l$print()
}
}

\section{Arguments}{

\itemize{
\item \code{log} (\code{data.table::data.table()}):
\code{data.table()} with columns \code{class} (\code{character()}) and \code{message} (\code{character()}).
\item \code{cl} (\code{character(1)}):
Class of a condition. One of "output", "warning", or "error". #FIXME: missing "message"?
}
}

\section{Details}{

\itemize{
\item \code{$errors} (\code{character()}) returns all lines which are errors.
\item \code{$warnings} (\code{character()}) returns all lines which are warnings.
\item \code{$has_condition(cl)} returns \code{TRUE} if at least on message of class \code{cl} is
logged. Possible conditions are "output", "message", "warning", and "error".
\item \code{$new(log)} parses the object returned by \link[evaluate:evaluate]{evaluate::evaluate} and creates
a new \link{Log}.
\item \code{format()} and \code{print()} are for formatting and printing via \link{format} or
\link{print}, respectively.
}
}

\examples{
# Create a simple experiment and extract the train log:
task = mlr_tasks$get("sonar")
learner = mlr_learners$get("classif.debug")
learner$param_vals = list(message_train = TRUE, error_train = TRUE)
learner$fallback = mlr_learners$get("classif.featureless")
e = Experiment$new(task, learner)
e$train(ctrl = list(encapsulate_train = "evaluate"))
l = e$logs$train

l$has_condition("error")
print(l)
}
\references{
\href{https://mlr3.mlr-org.com/reference/Log.html}{HTML help page}
}
