% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log.R
\docType{data}
\name{Log}
\alias{Log}
\title{Learner Output Log}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
Object which stores the text output of the \code{train} or \code{predict} step of an \link{Experiment}.
}
\section{Construction}{
\preformatted{l = Log$new(log = NULL)
}
\itemize{
\item \code{log} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table with at least the columns "class" (allowed values are "output", "warning", and "error")
and "msg" (\code{character()}).
A third column "context" is optional, and currently only used for printing.
}
}

\section{Fields}{

\itemize{
\item \code{warnings} :: \code{character(1)}\cr
Vector of all messages of class "warning".
\item \code{errors} :: \code{character(1)}\cr
Vector of all messages of class "error".
}
}

\section{Methods}{

\itemize{
\item \code{has_condition(cl)}\cr
\code{character(1)} -> \code{logical(1)}\cr
Returns \code{TRUE} if at least one message with condition \code{cl} has been recorded.
}
}

\examples{
# Create a simple experiment and extract the train log:
task = mlr_tasks$get("sonar")
learner = mlr_learners$get("classif.debug",
  param_vals = list(message_train = TRUE, error_train = TRUE))
e = Experiment$new(task, learner)
e$train(ctrl = list(encapsulate_train = "evaluate"))
l = e$log()

print(l)
l$has_condition("error")
}
\keyword{datasets}
