% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\docType{data}
\name{Measure}
\alias{Measure}
\title{Measure Class}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
This is the abstract base class for measures like \link{MeasureClassif} and \link{MeasureRegr}.
Predefined measures are stored in \link{mlr_measures}.
}
\section{Construction}{
\preformatted{m = Measure$new(id, task_type, range, minimize, predict_type = "response",
     task_properties = character(0L), na_score = FALSE, packages = character(0L))
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the measure.
\item \code{task_type} :: \code{character(1)}\cr
Type of the task the measure can operator on. E.g., \code{\"classif\"} or \code{\"regr\"}.
\item \code{range} :: \code{numeric(2)}\cr
Feasible range for this measure as \code{c(lower_bound, upper_bound)}.
\item \code{minimize} :: \code{logical(1)}\cr
Set to \code{TRUE} if good predictions correspond to small values.
\item \code{aggregator} :: \code{function()}\cr
Function to aggregate individual performance values.
If \code{NULL}, defaults to \code{\link[base:mean]{base::mean()}}.
\item \code{predict_type} :: \code{character(1)}\cr
Required predict type of the \link{Learner}.
\item \code{task_properties} :: \code{character()}\cr
Required task properties, see \link{Task}.
\item \code{na_score} :: \code{logical(1)}\cr
Is the measure expected to return \code{NA} in some cases? Default is \code{FALSE}.
\item \code{packages} :: \code{character()}\cr
Set of required packages.
Note that these packages will be loaded via \code{\link[=requireNamespace]{requireNamespace()}}, and are not attached.
}
}

\section{Fields}{

\itemize{
\item \code{id} :: \code{character(1)}\cr
Stores the identifier of the measure.
\item \code{minimize} :: \code{logical(1)}\cr
Is \code{TRUE} if the best value is reached via minimization and \code{FALSE} by maximization.
\item \code{packages} :: \code{character()}\cr
Stores the names of required packages.
\item \code{range} :: \code{numeric(2)}\cr
Stores the feasible range of the measure.
\item \code{task_type} :: \code{character(1)}\cr
Stores the required type of the \link{Task}.
\item \code{task_properties} :: \code{character()}\cr
Stores required properties of the \link{Task}.
}
}

\section{Methods}{

\itemize{
\item \code{aggregate(rr)}\cr
\link{ResampleResult} -> \code{numeric(1)}\cr
Aggregates multiple performance scores into a single score using the \code{aggregate} function of the measure.
Operates on a \link{ResampleResult} as returned by \link{resample}.
\item \code{calculate(experiment = NULL, prediction = experiment$prediction)}\cr
(\link{Experiment}, \link{Prediction}) -> \code{numeric(1)}\cr
Takes an \link{Experiment} and a \link{Prediction} (defaults to prediction stored in experiment), and calculates a numeric score.
}
}

\seealso{
Other Measure: \code{\link{MeasureClassif}},
  \code{\link{MeasureRegr}}, \code{\link{mlr_measures}}
}
\concept{Measure}
\keyword{datasets}
