% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{Measure}
\alias{Measure}
\title{Measure Class}
\format{\link{R6Class} object.}
\description{
Predefined measures are stored in \link{mlr_measures}.
}
\section{Usage}{
\preformatted{# Construction
m = Measure$new(id, range, minimize, predict_type = "response", task_properties = character(0L), packages = character(0L))
m = MeasureClassif$new(id, range, minimize, predict_type = "response", task_properties = character(0L), packages = character(0L))
m = MeasureRegr$new(id, range, minimize, predict_type = "response", task_properties = character(0L), packages = character(0L))

# Members
m$id
m$minimize
m$packages
m$predict_type
m$range
m$task_properties
m$task_type

# Methods
m$aggregate(rr)
m$calculate(e)
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}): Identifier for this object.
\item \code{range} (\code{numeric(2)}): Feasible range for this measure as \code{c(lower_bound, upper_bound)}.
\item \code{minimize} (\code{logical(1)}): Set to \code{TRUE} if good predictions correspond to small values.
\item \code{predict_type} (\code{character(1)}): Required predict type of the \link{Learner}.
\item \code{task_properties} (\code{character()}): Required task properties, see \link{Task}.
\item \code{packages} (\code{character()}): Set of required packages.
\item \code{e} (\link{Experiment}): Experiment to work on.
\item \code{rr} (\link{ResampleResult}): Performance object returned by \link{resample} to be aggregated.
}
}

\section{Details}{

\itemize{
\item \code{$id} (\code{character(1)}) stores the identifier of the object.
\item \code{$minimize} (\code{logical(1)}) indicates if the good values are reached via minimization.
\item \code{$packages} (\code{character()}) stores the set of required packages.
\item \code{$range} (\code{numeric(2)}) stores the numeric range of feasible measure values.
\item \code{$task_properties} (\code{character()}) defines a set of required task properties.
\item \code{$task_type} (\code{character()}) stores the class names of tasks this measure can operate on.
\item \code{$aggregate()} (\code{function(rr)}) aggregates multiple performance measures using the \code{aggregate} function. Operates on a \link{ResampleResult} as returned by \link{resample}.
\item \code{$calculate()} (\code{function(e)}) does the actual work.
\item \code{$new()} creates a new object of class \link{Measure}.
}
}

\examples{
mlr_measures$get("classif.mmce")
}
\seealso{
Other Measure: \code{\link{MeasureClassif}},
  \code{\link{MeasureRegr}}, \code{\link{mlr_measures}}
}
\concept{Measure}
