% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{Measure}
\alias{Measure}
\title{Abstract measure class}
\description{
Abstraction for performance measures.
}
\section{Usage}{
\preformatted{m = Measure$new()
m$id
m$packages
m$task_types
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{Measure}.
Predefined learners are stored in \link{mlr_measures}.

\code{$id} (\code{character(1)}) stores the identifier of the object.

\code{$packages} (\code{character(1)}) stores the names of required packages.

\code{$task_types} (\code{character}) stores the class names of tasks this measure can operate on.
}

\concept{Measure}
\keyword{internal}
