% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassif.R
\name{MeasureClassif}
\alias{MeasureClassif}
\title{Classification Measure}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Measure}.}
\description{
This task specializes \link{Measure} for classification problems.
}
\section{Usage}{

See \link{Measure}.
}

\section{Details}{

\code{$task_type} is \code{"classif"}.
}

\seealso{
Other Measure: \code{\link{MeasureRegr}},
  \code{\link{Measure}}, \code{\link{mlr_measures}}
}
\concept{Measure}
