% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifACC.R
\docType{data}
\name{MeasureClassifACC}
\alias{MeasureClassifACC}
\alias{mlr_measures_classif.acc}
\title{Accuracy Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\description{
Calls \code{\link[Metrics:accuracy]{Metrics::accuracy()}}.
}
\section{Construction}{
\preformatted{MeasureClassifACC$new()
mlr_measures$get("classif.acc")
msr("classif.acc")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
}
\keyword{datasets}
