% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifAUC.R
\docType{data}
\name{MeasureClassifAUC}
\alias{MeasureClassifAUC}
\alias{mlr_measures_classif.auc}
\title{Area Under the Curve Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\description{
Calls \code{\link[Metrics:auc]{Metrics::auc()}}.
}
\section{Construction}{
\preformatted{MeasureClassifAUC$new()
mlr_measures$get("classif.auc")
msr("classif.auc")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
}
\keyword{datasets}
