% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifFScore.R
\docType{data}
\name{MeasureClassifFScore}
\alias{MeasureClassifFScore}
\alias{mlr_measures_classif.f_score}
\title{F-score Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\description{
Calls \code{\link[Metrics:fbeta_score]{Metrics::fbeta_score()}}.
Argument \code{beta} defaults to \code{1}.
}
\section{Construction}{
\preformatted{MeasureClassifFScore$new(beta = 1)
mlr_measures$get("classif.f_score")
msr("classif.f_score")
}
\itemize{
\item \code{beta} :: \code{numeric(1)}\cr
Non-negative parameter balancing precision and recall.
Passed down to \code{\link[Metrics:fbeta_score]{Metrics::fbeta_score()}}
}
}

\references{
Sasaki, Yutaka (2007).
The truth of the F-measure.
Teach Tutor mater 1.5: 1-5.
\url{https://www.cs.odu.edu/~mukka/cs795sum09dm/Lecturenotes/Day3/F-measure-YS-26Oct07.pdf}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
}
\keyword{datasets}
