% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureOOBError.R
\name{MeasureOOBError}
\alias{MeasureOOBError}
\alias{mlr_measures_oob_error}
\title{Out-of-bag Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\description{
Returns the out-of-bag error of the \link{Learner} for learners that support it
(learners with property \code{"oob_error"}).
Returns \code{NA} for unsupported learners.
}
\section{Construction}{
\preformatted{MeasureOOBError$new()
mlr_measures$get("oob_error")
msr("oob_error")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}

\code{as.data.table(mlr_measures)} for a complete table of all (also dynamically created) \link{Measure} implementations.
}
