% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegr.R
\docType{data}
\name{MeasureRegr}
\alias{MeasureRegr}
\title{Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Measure}.}
\description{
This measure specializes \link{Measure} for regression problems.
Predefined measures can be found in the \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} \link{mlr_measures}.
}
\section{Construction}{
\preformatted{m = MeasureRegr$new(id, range, minimize = NA, average = "macro", aggregator = NULL, properties = character(), predict_type = "response",
    predict_sets = "test", task_properties = character(), packages = character())
}

For a description of the arguments, see \link{Measure}.
The \code{task_type} is set to \code{"regr"}.
Possible values for \code{predict_type} are \code{"response"} and \code{"se"}.
}

\section{Fields}{

See \link{Measure}.
}

\section{Methods}{

See \link{Measure}.
}

\seealso{
Example regression measures: \code{\link[=mlr_measures_regr.mse]{regr.mse}}

Other Measure: \code{\link{MeasureClassif}},
  \code{\link{Measure}}, \code{\link{mlr_measures}}
}
\concept{Measure}
\keyword{datasets}
