% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegr.R
\name{MeasureRegr}
\alias{MeasureRegr}
\title{Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Measure}.}
\description{
This task specializes \link{Measure} for regression problems.
}
\section{Usage}{

See \link{Measure}.
}

\section{Details}{

\code{$task_type} is \code{"regr"}.
}

\seealso{
Other Measure: \code{\link{MeasureClassif}},
  \code{\link{Measure}}, \code{\link{mlr_measures}}
}
\concept{Measure}
