% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegr.R
\name{MeasureRegr}
\alias{MeasureRegr}
\title{Regression Measure}
\description{
This task specializes \link{Measure} for regression problems.
}
\section{Usage}{

See \link{Measure}.
}

\section{Details}{

\itemize{
\item \code{$task_type} is \code{"regr"}.
}
}

\seealso{
Other Measure: \code{\link{MeasureClassif}},
  \code{\link{Measure}}, \code{\link{mlr_measures}}
}
\concept{Measure}
