% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegrMAE.R
\docType{data}
\name{MeasureRegrMAE}
\alias{MeasureRegrMAE}
\alias{mlr_measures_regr.mae}
\title{Absolute Errors Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{MeasureClassif}.}
\usage{
MeasureRegrMAE
}
\description{
Calls \code{\link[Metrics:mae]{Metrics::mae()}}.
}
\keyword{datasets}
