% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegrMSE.R
\docType{data}
\name{MeasureRegrMSE}
\alias{MeasureRegrMSE}
\alias{mlr_measures_regr.mse}
\title{Mean Squared Error Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{MeasureClassif}.}
\usage{
MeasureRegrMSE
}
\description{
Calls \code{\link[Metrics:mse]{Metrics::mse()}}.
}
\keyword{datasets}
