% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureRegrRMSE.R
\docType{data}
\name{MeasureRegrRMSE}
\alias{MeasureRegrRMSE}
\alias{mlr_measures_regr.rmse}
\title{Root Mean Squared Error Regression Measure}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{MeasureClassif}.}
\description{
Calls \code{\link[Metrics:rmse]{Metrics::rmse()}}.
}
\section{Construction}{
\preformatted{MeasureRegrRMSE$new()
mlr_measures$get("regr.rmse")
msr("regr.rmse")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
}
\keyword{datasets}
