% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3-package}
\alias{mlr3}
\alias{mlr3-package}
\title{mlr3: Machine Learning in R - Next Generation}
\description{
Machine Learning in R. Next Generation.
}
\section{Options}{

Available package options are documented in \code{\link[=mlr_options]{mlr_options()}}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/mlr3}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Bernd Bischl \email{bernd_bischl@gmx.de}
  \item Jakob Richter \email{jakob1richter@gmail.com}
}

}
