% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3-package}
\alias{mlr3}
\alias{mlr3-package}
\title{mlr3: Machine Learning in R - Next Generation}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Efficient, object-oriented programming on the building blocks of
    machine learning. Provides 'R6' objects for tasks, learners, resamplings,
    and measures. The package is geared towards scalability and larger datasets
    by supporting parallelization and out-of-memory data-backends like
    databases. While 'mlr3' focuses on the core computational operations,
    add-on packages provide additional functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (0000-0001-6002-6980)
  \item Jakob Richter \email{jakob1richter@gmail.com} (0000-0003-4481-5554)
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)
  \item Martin Binder \email{mlr.developer@mb706.com}
}

Other contributors:
\itemize{
  \item Giuseppe Casalicchio \email{giuseppe.casalicchio@stat.uni-muenchen.de} (0000-0001-5324-5966) [contributor]
  \item Stefan Coors \email{mail@stefancoors.de} (0000-0002-7465-2146) [contributor]
  \item Quay Au \email{quayau@gmail.com} (0000-0002-5252-8902) [contributor]
}

}
