% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mlr_assertions}
\alias{mlr_assertions}
\alias{assert_backend}
\alias{assert_experiment}
\alias{assert_task}
\alias{assert_tasks}
\alias{assert_learner}
\alias{assert_learners}
\alias{assert_measure}
\alias{assert_measures}
\alias{assert_resampling}
\alias{assert_resamplings}
\alias{assert_resample_result}
\alias{assert_benchmark_result}
\alias{assert_param_set}
\alias{assert_id}
\alias{assert_ids}
\title{Assertion for mlr3 Objects}
\usage{
assert_backend(b, .var.name = vname(b))

assert_experiment(e, .var.name = vname(e))

assert_task(task, feature_types = NULL, task_properties = NULL,
  clone = FALSE, allow_empty_rows = FALSE, minimum_nrows = NULL,
  allow_empty_cols = FALSE, minimum_ncols = NULL)

assert_tasks(tasks, feature_types = NULL, task_properties = NULL,
  clone = FALSE)

assert_learner(learner, task = NULL, properties = character(0L),
  clone = FALSE)

assert_learners(learners, task = NULL, properties = character(0L),
  clone = FALSE)

assert_measure(measure, task = NULL, predict_types = NULL,
  clone = FALSE)

assert_measures(measures, task = NULL, predict_types = NULL,
  clone = FALSE)

assert_resampling(resampling, instantiated = NULL, clone = FALSE)

assert_resamplings(resamplings, instantiated = NULL, clone = FALSE)

assert_resample_result(resample_result,
  .var.name = vname(resample_result))

assert_benchmark_result(bmr, .var.name = vname(bmr))

assert_param_set(param_set, .var.name = vname(param_set))

assert_id(id, .var.name = vname(id))

assert_ids(ids, .var.name = vname(ids))
}
\arguments{
\item{b}{:: \link{DataBackend}.}

\item{e}{:: \link{Experiment}.}

\item{task}{:: \link{Task}.}

\item{feature_types}{:: \code{character()}\cr
Set of allowed feature types.}

\item{task_properties}{:: \code{character()}\cr
Set of required task properties.}

\item{allow_empty_rows}{:: \code{logical()}\cr
Is a task without observations allowed?}

\item{minimum_nrows}{:: \code{integer()}\cr
Is a minimum amount of observations required?}

\item{tasks}{:: list of \link{Task}.}

\item{learner}{:: \link{Learner}.}

\item{learners}{:: list of \link{Learner}.}

\item{measure}{:: \link{Measure}.}

\item{predict_types}{:: \code{character()}\cr
Vector of predict types provided by the \link{Experiment} or \link{Learner}.}

\item{measures}{:: list of \link{Measure}.}

\item{resampling}{:: \link{Resampling}.}

\item{resamplings}{:: list of \link{Resampling}.}

\item{resample_result}{:: \link{ResampleResult}.}

\item{bmr}{:: \link{BenchmarkResult}.}

\item{param_set}{:: \link[paradox:ParamSet]{paradox::ParamSet}.}

\item{id}{:: \code{character(1)}.}

\item{allow_empty_rows}{:: \code{logical()}\cr
Is a task without features allowed?}

\item{minimum_cols}{:: \code{integer()}\cr
Is a minimum amount of features required?}

\item{id}{:: \code{character(1)}.}
}
\description{
Functions intended to be used in packages extending \pkg{mlr3}.
}
\keyword{internal}
