% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{mlr_assertions}
\alias{mlr_assertions}
\alias{assert_backend}
\alias{assert_experiment}
\alias{assert_task}
\alias{assert_task_type}
\alias{assert_learner}
\alias{assert_measure}
\alias{assert_measures}
\alias{assert_resampling}
\alias{assert_resample_result}
\alias{assert_benchmark_result}
\alias{assert_param_set}
\alias{assert_param_vals}
\title{Assertion for mlr3 Objects}
\usage{
assert_backend(b)

assert_experiment(e)

assert_task(task)

assert_task_type(object, task = NULL)

assert_learner(learner, task = NULL)

assert_measure(measure, task = NULL, prediction = NULL)

assert_measures(measures, task = NULL, prediction = NULL)

assert_resampling(resampling)

assert_resample_result(resample_result)

assert_benchmark_result(bmr)

assert_param_set(param_set)

assert_param_vals(param_vals, param_set)
}
\arguments{
\item{b}{(\link{DataBackend}):\cr
A data backend to be checked.}

\item{e}{(\link{Experiment}):\cr
An experiment to be checked.}

\item{task}{(\link{Task}):\cr
An task to be checked.}

\item{object}{(\link{Learner} | \link{Filter}):\cr
A filter or learner to be checked}

\item{learner}{(\link{Learner}):\cr
A learner to be checked.}

\item{measure}{(\link{Measure}):\cr
A measure to be checked.}

\item{measures}{(list of \link{Measure}).}

\item{resampling}{(\link{Resampling}):\cr
A resampling object to be checked.}

\item{resample_result}{(\link{ResampleResult}):\cr
A resample result object to be checked.}

\item{bmr}{(\link{BenchmarkResult}):\cr
A benchmark object to be checked.}

\item{param_set}{(\link[paradox:ParamSet]{paradox::ParamSet}):\cr
A param set to be checked.}

\item{param_vals}{(named list):\cr
A param vals object be checked.}

\item{task}{(\link{Task}):\cr
An task to be checked.}
}
\description{
Functions intended to be used in packages extending \pkg{mlr3}.
}
\keyword{internal}
