% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_assertions.R
\name{mlr_assertions}
\alias{mlr_assertions}
\alias{assert_backend}
\alias{assert_experiment}
\alias{assert_task}
\alias{assert_learner}
\alias{assert_measure}
\alias{assert_measures}
\alias{assert_resampling}
\alias{assert_resample_result}
\alias{assert_benchmark_result}
\alias{assert_param_set}
\alias{assert_param_vals}
\title{Assertion for mlr3 objects}
\usage{
assert_backend(b)

assert_experiment(e)

assert_task(task)

assert_learner(learner, task = NULL)

assert_measure(measure, task = NULL, learner = NULL)

assert_measures(measures, task = NULL, learner = NULL)

assert_resampling(resampling)

assert_resample_result(resample_result)

assert_benchmark_result(bmr)

assert_param_set(param_set)

assert_param_vals(param_vals, param_set)
}
\arguments{
\item{b}{[\link{DataBackend}].}

\item{e}{[\link{Experiment}].}

\item{task}{[\link{Task}].}

\item{learner}{[\link{Learner}].}

\item{measure}{[\link{Measure}].}

\item{measures}{[\code{list} of \link{Measure}].}

\item{resampling}{[\link{Resampling}].}

\item{resample_result}{[\link{ResampleResult}].}

\item{bmr}{[\link{BenchmarkResult}].}

\item{param_set}{[\code{\link[paradox:ParamSet]{paradox::ParamSet()}}].}

\item{param_vals}{[\code{named list}].}
}
\description{
Functions intended to be used in packages extending \pkg{mlr3}.
}
\keyword{internal}
