% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_control.R
\name{mlr_control}
\alias{mlr_control}
\title{Execution Control Object}
\usage{
mlr_control(...)
}
\arguments{
\item{...}{:: \code{any}\cr
Named arguments to overwrite the defaults / options.
Settings may be provided in a \code{name = value} fashion, or by providing a single named \code{list()}.}
}
\value{
(named \code{list()}) of all settings.
}
\description{
This function creates a named list of settings which control the execution of a \link{Learner}.
\itemize{
\item \code{store_models} (\code{logical(1))}:\cr
If \code{FALSE} (default), models fitted during \code{\link[=resample]{resample()}} and \code{\link[=benchmark]{benchmark()}} are discarded.
Note that you will be unable to predict on new data or extract additional information like
variable importance from the learner.
}

\strong{Defaults}\preformatted{$store_models
TRUE

$log_threshold
400
}
}
\examples{
# get a list of the defaults
mlr_control()
}
