% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_generators.R
\name{mlr_generators}
\alias{mlr_generators}
\title{Dictionary of Task Generators}
\description{
A simple \link{Dictionary} storing generator functions returning a \link{Task}.
}
\section{Methods}{



\itemize{
\item \code{get(key, ...)}\cr
(\code{character(1)}, ...) -> \code{any}\cr
Retrieves object with key \code{key} from the dictionary.
Additional arguments are passed to the stored object during construction.
\item \code{mget(keys, ...)}\cr
(\code{character()}, ...) -> named \code{list()}\cr
Retrieves objects with keys \code{keys} from the dictionary, returns them in a list named with \code{keys}.
Additional arguments are passed to the stored objects during construction.
\item \code{has(keys)}\cr
\code{character()} -> \code{logical()}\cr
Returns a logical vector with \code{TRUE} at its i-th position, if the i-th key exists.
\item \code{keys(pattern)}\cr
\code{character(1)} -> \code{character()}\cr
Returns all keys which comply to the regular expression \code{pattern}.
\item \code{add(key, value, ..., required_args = character())}\cr
(\code{character(1)}, \code{any}, ..., \code{character()}) -> \code{self}\cr
Adds object \code{value} to the dictionary with key \code{key}, potentially overwriting a previously stored item.
Additional arguments in \code{...} are used as default arguments for \code{value} during construction.
If the object is not constructable without additional arguments, their names should be listed in \code{required_args}.
\item \code{remove(key)}\cr
\code{character()} -> \code{self}\cr
Removes object with key \code{key} from the dictionary.
\item \code{required_args()}\cr
() -> named \code{list()}\cr
Returns character vectors of required arguments for all items.
}
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link{Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{data.table()} with fields "key" and "packages" as columns.
}
}

\examples{
g = mlr_generators$get("smiley")
task = g$generate(10)
print(task)
task$data()
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_learners}}, \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}

Other Task: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{TaskSupervised}},
  \code{\link{Task}}, \code{\link{mlr_tasks}}
}
\concept{Dictionary}
\concept{Generator}
\concept{Task}
