% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{mlr_learner_conditions}
\alias{mlr_learner_conditions}
\alias{error_learner}
\alias{condition_learner}
\alias{error_learner_train}
\alias{condition_learner_train}
\alias{error_learner_predict}
\alias{condition_learner_predict}
\title{Learner Classes}
\usage{
error_learner(msg, ..., class = NULL)

condition_learner(msg, ..., class = NULL)

error_learner_train(msg, ..., class = NULL)

condition_learner_train(msg, ..., class = NULL)

error_learner_predict(msg, ..., class = NULL)

condition_learner_predict(msg, ..., class = NULL)
}
\arguments{
\item{msg}{(\code{character(1)})\cr
Message to be printed.}

\item{...}{(\code{any})\cr
Additional arguments to be passed to \code{\link[mlr3misc:printf]{mlr3misc::stopf()}}.}

\item{class}{(\code{character()})\cr
Additional classes to be added to the error.}
}
\description{
Learner Classes
}
