% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners.R
\name{mlr_learners}
\alias{mlr_learners}
\title{Dictionary of Learners}
\format{\link[R6:R6Class]{R6::R6Class} object}
\description{
A simple \link{Dictionary} storing objects of class \link{Learner}.
Each learner has an associated help page, see \code{mlr_learners_[id]}.
}
\section{Methods}{



\itemize{
\item \code{get(key, ...)}\cr
(\code{character(1)}, ...) -> \code{any}\cr
Retrieves object with key \code{key} from the dictionary.
Additional arguments are passed to the stored object during construction.
\item \code{mget(keys, ...)}\cr
(\code{character()}, ...) -> named \code{list()}\cr
Retrieves objects with keys \code{keys} from the dictionary, returns them in a list named with \code{keys}.
Additional arguments are passed to the stored objects during construction.
\item \code{has(keys)}\cr
\code{character()} -> \code{logical()}\cr
Returns a logical vector with \code{TRUE} at its i-th position, if the i-th key exists.
\item \code{keys(pattern)}\cr
\code{character(1)} -> \code{character()}\cr
Returns all keys which comply to the regular expression \code{pattern}.
\item \code{add(key, value, ..., required_args = character())}\cr
(\code{character(1)}, \code{any}, ..., \code{character()}) -> \code{self}\cr
Adds object \code{value} to the dictionary with key \code{key}, potentially overwriting a previously stored item.
Additional arguments in \code{...} are used as default arguments for \code{value} during construction.
If the object is not constructable without additional arguments, their names should be listed in \code{required_args}.
\item \code{remove(key)}\cr
\code{character()} -> \code{self}\cr
Removes object with key \code{key} from the dictionary.
\item \code{required_args()}\cr
() -> named \code{list()}\cr
Returns character vectors of required arguments for all items.
}
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link{Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{data.table()} with fields "key", "feature_types", "packages",
"properties" and "predict_types" as columns.
}
}

\examples{
as.data.table(mlr_learners)
mlr_learners$get("classif.featureless")
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_generators}}, \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}

Other Learner: \code{\link{LearnerClassif}},
  \code{\link{LearnerRegr}}, \code{\link{Learner}}
}
\concept{Dictionary}
\concept{Learner}
