% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners.R
\name{mlr_learners}
\alias{mlr_learners}
\title{Dictionary of Learners}
\description{
A simple \link{Dictionary} storing objects of class \link{Learner}.
Each learner has an associated help page, see \code{mlr_learners_[id]}.
}
\section{Usage}{

See \link{Dictionary}.
}

\examples{
mlr_learners$ids()
as.data.table(mlr_learners)
mlr_learners$get("classif.featureless")
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}

Other Learner: \code{\link{LearnerClassif}},
  \code{\link{LearnerRegr}}, \code{\link{Learner}}
}
\concept{Dictionary}
\concept{Learner}
