% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners.R
\name{mlr_learners}
\alias{mlr_learners}
\title{Dictionary of Learners}
\description{
Dictionary of Learners
}
\section{Usage}{

See \link{Dictionary}.
}

\examples{
mlr_learners$keys()
as.data.table(mlr_learners)
mlr_learners$get("classif.dummy")
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}

Other Learner: \code{\link{Learner}}
}
\concept{Dictionary}
\concept{Learner}
