% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners.R
\docType{class}
\name{mlr_learners}
\alias{mlr_learners}
\title{Registered Learners}
\format{\code{\link[=R6Class]{R6Class()}} object}
\usage{
mlr_learners
}
\description{
\code{mlr_learners} is a \code{\link[=Dictionary]{Dictionary()}} used to manage learners.
}
\examples{
mlr_learners$ids
mlr_learners$get("classif.dummy")
}
\keyword{datasets}
