% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_rpart.R
\docType{data}
\name{mlr_learners_classif.rpart}
\alias{mlr_learners_classif.rpart}
\alias{LearnerClassifRpart}
\title{Classification Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifRpart
}
\description{
A \link{LearnerClassif} for a classification tree implemented in \link[rpart:rpart]{rpart::rpart}.
}
\keyword{datasets}
