% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifRpart.R
\name{mlr_learners_classif.rpart}
\alias{mlr_learners_classif.rpart}
\alias{LearnerClassifRpart}
\title{Classification Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\description{
A \link{LearnerClassif} for a classification tree implemented in \code{\link[rpart:rpart]{rpart::rpart()}} in package \CRANpkg{rpart}.
Parameter \code{xval} is set to 0 in order to save some computation time.
}
\section{Construction}{
\preformatted{LearnerClassifRpart$new()
mlr_learners$get("classif.rpart")
lrn("classif.rpart")
}
}

\references{
\cite{mlr3}{breiman_1984}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Learner]{Learners}: \link{mlr_learners}

\code{as.data.table(mlr_learners)} for a complete table of all (also dynamically created) \link{Learner} implementations.
}
