% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_classif_rpart.R
\docType{data}
\name{mlr_learners_classif_rpart}
\alias{mlr_learners_classif_rpart}
\alias{LearnerClassifRpart}
\title{Classification Tree Learner}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifRpart
}
\description{
A learner for a classification tree implemented in \code{\link[rpart:rpart]{rpart::rpart()}}.
}
\keyword{datasets}
