% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_learners_regr_rpart.R
\docType{data}
\name{mlr_learners_regr.rpart}
\alias{mlr_learners_regr.rpart}
\alias{LearnerRegrRpart}
\title{Regression Tree Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrRpart
}
\description{
A \link{LearnerClassif} for a regression tree implemented in \code{\link[rpart:rpart]{rpart::rpart()}} in package \CRANpkg{rpart}.
}
\references{
Breiman, L. (1984).
Classification and Regression Trees.
New York: Routledge.
\url{https://doi.org/10.1201/9781315139470}
}
\keyword{datasets}
