% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures.R
\docType{class}
\name{mlr_measures}
\alias{mlr_measures}
\title{Registered Measures}
\format{\code{\link{R6Class}} object}
\usage{
mlr_measures
}
\description{
\code{mlr_measures} is a \code{\link{Dictionary}} used to manage performance measures.
}
\keyword{datasets}
