% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures.R
\docType{data}
\name{mlr_measures}
\alias{mlr_measures}
\title{Dictionary of Performance Measures}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.}
\description{
A simple \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} storing objects of class \link{Measure}.
Each measure has an associated help page, see \code{mlr_measures_[id]}.

This dictionary can get populated with additional measures by add-on packages.

For a more convenient way to retrieve and construct measures, see \code{\link[=msr]{msr()}}/\code{\link[=msrs]{msrs()}}.
}
\section{Methods}{

See \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link[mlr3misc:Dictionary]{mlr3misc::Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{\link[data.table:data.table]{data.table::data.table()}} with fields "key", "task_type", "predict_type",
and "packages" as columns.
}
}

\examples{
as.data.table(mlr_measures)
mlr_measures$get("classif.ce")
msr("regr.mse")
}
\seealso{
Sugar functions: \code{\link[=msr]{msr()}}, \code{\link[=msrs]{msrs()}}

Implementation of most measures: \CRANpkg{mlr3measures}

Other Dictionary: 
\code{\link{mlr_learners}},
\code{\link{mlr_resamplings}},
\code{\link{mlr_task_generators}},
\code{\link{mlr_tasks}}

Other Measure: 
\code{\link{MeasureClassif}},
\code{\link{MeasureRegr}},
\code{\link{Measure}},
\code{\link{mlr_measures_classif.costs}},
\code{\link{mlr_measures_debug}},
\code{\link{mlr_measures_elapsed_time}},
\code{\link{mlr_measures_oob_error}},
\code{\link{mlr_measures_selected_features}}
}
\concept{Dictionary}
\concept{Measure}
\keyword{datasets}
