% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures.R
\name{mlr_measures}
\alias{mlr_measures}
\title{Dictionary of Performance Measures}
\format{\link{R6Class} object}
\description{
A simple \link{Dictionary} storing objects of class \link{Measure}.
Each measure has an associated help page, see \code{mlr_measures_[id]}.
}
\section{Usage}{

See \link{Dictionary}.
}

\examples{
mlr_measures$ids()
as.data.table(mlr_measures)
mlr_measures$get("mmce")
}
\references{
\href{https://mlr3.mlr-org.com/reference/mlr_measures.html}{HTML help page}
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_learners}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}

Other Measure: \code{\link{MeasureClassif}},
  \code{\link{MeasureRegr}}, \code{\link{Measure}}
}
\concept{Dictionary}
\concept{Measure}
