% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureBestValidScore.R
\name{mlr_measures_best_valid_score}
\alias{mlr_measures_best_valid_score}
\alias{MeasureBestValidScore}
\title{Measure Best Validation Score}
\description{
Returns the selected best internal validation score of the \link{Learner} for learners with property \code{"validation"}.
Returns \code{NA} for unsupported learners, when no validation was done, or when the selected id was not found.
The \code{id} of this measure is set to the value of \code{select} if provided.

Note that not every learner with the \code{"validation"} property tracks the best validation score:
this requires the learner's \verb{$.extract_internal_valid_scores()} method to support the \code{which} argument
(see \code{\link{Learner}}, section \emph{Implementing Validation}).
For learners that do not, this measure returns \code{NA}.

While \code{\link[=mlr_measures_internal_valid_score]{msr("internal_valid_score")}} reports the validation score of the
\emph{final} model, this measure reports the \emph{best} validation score observed during training.
For learners that internally tune a hyperparameter such as the number of boosting rounds or epochs, these two
can differ: the internally tuned values (see \verb{$internal_tuned_values}) usually correspond to the iteration with
the best validation score, whereas the final model is the one after the last iteration.
Which of the two is the appropriate tuning measure depends on the learner:
if the model that is used for prediction is the one from the best iteration, use this measure,
otherwise use \code{\link[=mlr_measures_internal_valid_score]{msr("internal_valid_score")}}.
}
\section{Dictionary}{

This \link{Measure} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_measures} or with the associated sugar function \code{\link[=msr]{msr()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_measures$get("best_valid_score")
msr("best_valid_score")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{NA}
\item Range: \eqn{(-\infty, \infty)}{(-Inf, Inf)}
\item Minimize: NA
\item Average: macro
\item Required Prediction: \dQuote{NA}
\item Required Packages: \CRANpkg{mlr3}
}
}

\section{Parameters}{

Empty ParamSet
}

\examples{
rr = resample(tsk("iris"), lrn("classif.debug", validate = 0.3), rsmp("holdout"))
rr$score(msr("best_valid_score", select = "acc"))
}
\seealso{
\itemize{
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}:
\url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html#sec-eval}
\item Package \CRANpkg{mlr3measures} for the scoring functions.
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Measure]{Measures}: \link{mlr_measures}
\code{as.data.table(mlr_measures)} for a table of available \link[=Measure]{Measures} in the running session (depending on the loaded packages).
\item Extension packages for additional task types:
\itemize{
\item \CRANpkg{mlr3proba} for probabilistic supervised regression and survival analysis.
\item \CRANpkg{mlr3cluster} for unsupervised clustering.
}
}

Other Measure:
\code{\link{Measure}},
\code{\link{MeasureClassif}},
\code{\link{MeasureRegr}},
\code{\link{MeasureSimilarity}},
\code{\link{mlr_measures}},
\code{\link{mlr_measures_aic}},
\code{\link{mlr_measures_bic}},
\code{\link{mlr_measures_classif.costs}},
\code{\link{mlr_measures_debug_classif}},
\code{\link{mlr_measures_elapsed_time}},
\code{\link{mlr_measures_internal_valid_score}},
\code{\link{mlr_measures_oob_error}},
\code{\link{mlr_measures_regr.pinball}},
\code{\link{mlr_measures_regr.rqr}},
\code{\link{mlr_measures_regr.rsq}},
\code{\link{mlr_measures_selected_features}}
}
\concept{Measure}
\section{Super class}{
\code{\link[mlr3:Measure]{Measure}} -> \code{MeasureBestValidScore}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-MeasureBestValidScore-initialize}{\code{MeasureBestValidScore$new()}}
    \item \href{#method-MeasureBestValidScore-clone}{\code{MeasureBestValidScore$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>Measure$aggregate()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>Measure$format()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>Measure$help()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="obs_loss"><a href='../../mlr3/html/Measure.html#method-Measure-obs_loss'><code>Measure$obs_loss()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>Measure$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>Measure$score()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureBestValidScore-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureBestValidScore-initialize}{}}}
\subsection{\code{MeasureBestValidScore$new()}}{
  Creates a new instance of this \link[R6:R6Class]{R6} class.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MeasureBestValidScore$new(select = NULL, minimize = NA)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{select}}{(\code{character(1)})\cr
Which of the best validation scores to select.
Which scores are available depends on the learner and its configuration.
By default, the first score is chosen.}
      \item{\code{minimize}}{(\code{logical(1)})\cr
Whether smaller values are better.
Must be set to use for tuning.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureBestValidScore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureBestValidScore-clone}{}}}
\subsection{\code{MeasureBestValidScore$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{MeasureBestValidScore$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
