% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_classif_acc.R
\docType{data}
\name{mlr_measures_classif.acc}
\alias{mlr_measures_classif.acc}
\alias{MeasureClassifACC}
\title{Accuracy Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifACC
}
\description{
Calls \link[Metrics:accuracy]{Metrics::accuracy}.
}
\keyword{datasets}
