% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_classif_auc.R
\docType{data}
\name{mlr_measures_classif.auc}
\alias{mlr_measures_classif.auc}
\alias{MeasureClassifAUC}
\title{Area Under the Curve Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifAUC
}
\description{
Calls \code{\link[Metrics:auc]{Metrics::auc()}}.
}
\keyword{datasets}
