% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifCE.R
\docType{data}
\name{mlr_measures_classif.ce}
\alias{mlr_measures_classif.ce}
\alias{MeasureClassifCE}
\title{Classification Error Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifCE
}
\description{
Calls \code{\link[Metrics:ce]{Metrics::ce()}}.
}
\keyword{datasets}
