% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_classif_costs.R
\docType{data}
\name{mlr_measures_classif.costs}
\alias{mlr_measures_classif.costs}
\alias{MeasureClassifCosts}
\title{Cost-sensitive Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifCosts
}
\description{
The cost matrix is stored as slot "costs" and must be set manually.
}
\keyword{datasets}
