% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_classif_ppv.R
\docType{data}
\name{mlr_measures_classif.ppv}
\alias{mlr_measures_classif.ppv}
\alias{MeasureClassifPPV}
\title{Positive Predictive Value Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifPPV
}
\description{
Defined as: TP / (TP + FP). Also called precision. If the denominator is 0, PPV is set to be either 1 or 0
depending on whether the highest probability prediction is positive (1) or negative (0).
}
\keyword{datasets}
