% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_classif_tpr.R
\docType{data}
\name{mlr_measures_classif.tpr}
\alias{mlr_measures_classif.tpr}
\alias{MeasureClassifTPR}
\alias{mlr_measures_classif.recall}
\title{True Positive Rate Classification Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{MeasureClassif}.}
\usage{
MeasureClassifTPR
}
\description{
Defined as: TP / (TP + FN).
Also called hit rate or recall or sensitivity.
}
\keyword{datasets}
