% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_elapsed_time.R
\docType{data}
\name{mlr_measures_elapsed_time}
\alias{mlr_measures_elapsed_time}
\alias{MeasureElapsedTime}
\title{Elapsed Time Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\usage{
MeasureElapsedTime
}
\description{
Measures the elapsed time during train, predict, or both.
}
\keyword{datasets}
