% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_elapsed_time.R
\docType{data}
\name{mlr_measures_elapsed_time}
\alias{mlr_measures_elapsed_time}
\alias{MeasureElapsedTime}
\alias{MeasureTimeTrain}
\alias{MeasureTimePredict}
\alias{MeasureTimeBoth}
\title{Elapsed Time Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\usage{
MeasureElapsedTime

MeasureTimeTrain

MeasureTimePredict

MeasureTimeBoth
}
\description{
Measures the elapsed time during train, predict, or both.
}
\references{
\href{https://mlr3.mlr-org.com/reference/mlr_measures_elapsed_time.html}{HTML help page}
}
\keyword{datasets}
