% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_measures_selected_features.R
\docType{data}
\name{mlr_measures_selected_features}
\alias{mlr_measures_selected_features}
\alias{MeasureSelectedFeatures}
\title{Selected Features Measure}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Measure}.}
\usage{
MeasureSelectedFeatures
}
\description{
Measures the number of selected features.
If field \code{normalize} is set to \code{TRUE}, divides the number of features by the total number of available features.
}
\keyword{datasets}
