% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_reflections.R
\docType{data}
\name{mlr_reflections}
\alias{mlr_reflections}
\title{Reflections for mlr3}
\format{\link{environment}.}
\usage{
mlr_reflections
}
\description{
Environment which stores various information for reflections.
Can be modified by third-party packages, e.g. by adding information about new task types.
}
\examples{
ls.str(mlr_reflections)
}
\keyword{internal}
