% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflections.R
\docType{data}
\name{mlr_reflections}
\alias{mlr_reflections}
\title{Reflections for mlr3}
\format{An object of class \code{environment} of length 11.}
\usage{
mlr_reflections
}
\description{
Environment which stores various information for reflections.
}
\keyword{internal}
