% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings.R
\docType{class}
\name{mlr_resamplings}
\alias{mlr_resamplings}
\title{Registered Resampling Methods}
\format{\code{\link{R6Class}} object}
\usage{
mlr_resamplings
}
\description{
\code{Resamplings} is a \code{\link{Dictionary}} used to manage resampling methods.
}
\keyword{datasets}
