% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings.R
\name{mlr_resamplings}
\alias{mlr_resamplings}
\title{Dictionary of Resampling Strategies}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
A simple \link{Dictionary} storing objects of class \link{Resampling}.
Each resampling has an associated help page, see \code{mlr_resamplings_[id]}.
}
\section{Methods}{

See \link{Dictionary}.
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link{Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{data.table()} with fields "key", "hyperpars", and "iters"
as columns.
}
}

\examples{
as.data.table(mlr_resamplings)
mlr_resamplings$get("cv")
}
\seealso{
Example resamplings:
\itemize{
\item \code{\link[=mlr_resamplings_cv]{cv}}.
\item \code{\link[=mlr_resamplings_bootstrap]{bootstrap}}.
}

Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_generators}}, \code{\link{mlr_learners}},
  \code{\link{mlr_measures}}, \code{\link{mlr_tasks}}

Other Resampling: \code{\link{Resampling}}
}
\concept{Dictionary}
\concept{Resampling}
