% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings.R
\name{mlr_resamplings}
\alias{mlr_resamplings}
\title{Dictionary of registered resampling strategies}
\description{
Dictionary of registered resampling strategies
}
\section{Usage}{

See \link{Dictionary}.
}

\examples{
mlr_resamplings$ids
mlr_resamplings$get("cv")
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_learners}}, \code{\link{mlr_measures}},
  \code{\link{mlr_tasks}}
}
\concept{Dictionary}
