% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings_bootstrap.R
\docType{data}
\name{mlr_resamplings_bootstrap}
\alias{mlr_resamplings_bootstrap}
\alias{ResamplingBootstrap}
\title{Bootstrap Resampling}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Resampling}.}
\usage{
ResamplingBootstrap
}
\description{
Simple Bootstrap sampling.
You can control the number of bootstrap iterations (\code{repeats})
and the number of observations to draw per iteration (\code{ratio}).
}
\examples{
# Create a task with 10 observations
task = mlr_tasks$get("iris")
task$filter(1:10)

# Instantiate Resampling
rb = ResamplingBootstrap$new()
rb$param_vals = list(repeats = 2, ratio = 1)
rb$instantiate(task)

# Individual sets:
rb$train_set(1)
rb$test_set(1)
intersect(rb$train_set(1), rb$test_set(1))

# Internal storage:
rb$instance$M # Matrix of counts
}
\keyword{datasets}
