% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings_custom.R
\docType{data}
\name{mlr_resamplings_custom}
\alias{mlr_resamplings_custom}
\alias{ResamplingCustom}
\title{Custom Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\usage{
ResamplingCustom
}
\description{
A custom resampling class where the training and test indices
can be set manually.
}
\examples{
# Create a task with 10 observations
task = mlr_tasks$get("iris")
task$filter(1:10)

# Instantiate Resampling
rc = mlr_resamplings$get("custom")
train_sets = list(1:5, 5:10)
test_sets = list(5:10, 1:5)
rc$instantiate(task, train_sets, test_sets)

rc$train_set(1)
rc$test_set(1)
}
\keyword{datasets}
