% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings_cv.R
\docType{data}
\name{mlr_resamplings_cv}
\alias{mlr_resamplings_cv}
\alias{ResamplingCV}
\title{Cross Validation Resampling}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Resampling}.}
\usage{
ResamplingCV
}
\description{
Cross validation with \code{folds} folds (default: 10).
}
\examples{
# Create a task with 10 observations
task = mlr_tasks$get("iris")
task$filter(1:10)

# Instantiate Resampling
rcv = ResamplingCV$new()
rcv$param_vals = list(folds = 3)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)
rcv$test_set(1)
intersect(rcv$train_set(1), rcv$test_set(1))

# Internal storage:
rcv$instance # table
}
\keyword{datasets}
