% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings_holdout.R
\docType{data}
\name{mlr_resamplings_holdout}
\alias{mlr_resamplings_holdout}
\alias{ResamplingHoldout}
\title{Holdout Resampling}
\format{\link{R6Class} object}
\usage{
ResamplingHoldout
}
\description{
Simple holdout: A single split into training and test.
Parameter \code{ratio} determines the ratio of observation in the train set.
}
\examples{
# Create a task with 10 observations
task = mlr_tasks$get("iris")
task$filter(1:10)

# Instantiate Resampling
rho = mlr_resamplings$get("holdout")
rho$instantiate(task)

# Individual sets:
rho$train_set(1)
rho$test_set(1)
intersect(rho$train_set(1), rho$test_set(1))

# Internal storage:
rho$instance # simple list
}
\references{
\href{https://mlr3.mlr-org.com/reference/mlr_resamplings_holdout.html}{HTML help page}
}
\keyword{datasets}
