% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_resamplings_subsampling.R
\docType{data}
\name{mlr_resamplings_subsampling}
\alias{mlr_resamplings_subsampling}
\alias{ResamplingSubsampling}
\title{Subsampling Resampling}
\format{\code{\link[R6:R6Class]{R6::R6Class()}} inheriting from \link{Resampling}.}
\usage{
ResamplingSubsampling
}
\description{
\code{repeats} times repeated splits into training and test set with a \code{ratio} ratio of training observations.
}
\examples{
# Create a task with 10 observations
task = mlr_tasks$get("iris")
task$filter(1:10)

# Instantiate Resampling
rss = ResamplingSubsampling$new()
rss$param_vals = list(repeats = 2, ratio = 0.5)
rss$instantiate(task)

# Individual sets:
rss$train_set(1)
rss$test_set(1)
intersect(rss$train_set(1), rss$test_set(1))

# Internal storage:
rss$instance$train # list of bit vectors
}
\keyword{datasets}
