% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_sugar.R
\name{mlr_sugar}
\alias{mlr_sugar}
\alias{tsk}
\alias{tgen}
\alias{lrn}
\alias{rsmp}
\alias{msr}
\title{Syntactic Sugar for Object Construction}
\usage{
tsk(.key, ...)

tgen(.key, ...)

lrn(.key, ...)

rsmp(.key, ...)

msr(.key, ...)
}
\arguments{
\item{.key}{:: \code{character(1)}\cr
Key passed to the respective \link[mlr3misc:Dictionary]{mlr3misc::Dictionary} to retrieve the object.}

\item{...}{:: named \code{list()}\cr
Named arguments passed to the constructor, to be set as parameters in the \link[paradox:ParamSet]{paradox::ParamSet}, or to be set as public field.
See \code{\link[mlr3misc:dictionary_sugar]{mlr3misc::dictionary_sugar()}} for more details.}
}
\value{
\link[R6:R6Class]{R6::R6Class} of the respective type.
}
\description{
Currently, there are three mechanisms in place to ease the construction of objects.
\enumerate{
\item Instead of creating and passing an object manually, it is often possible to pass a character vector.
The elements of this vector are used as keys to retrieve objects form the respective dictionaries (c.f. \code{\link[mlr3misc:dictionary_sugar]{mlr3misc::dictionary_sugar()}}).
E.g., \code{resample(task = "iris", learner = "classif.rpart", resampling = "cv3")} automatically fetches the following:
\itemize{
\item The \link[=mlr_tasks_iris]{iris task} from \link{mlr_tasks}.
\item A \link[=mlr_learners_classif.rpart]{classification decision tree} from \link{mlr_learners}.
\item A \link[=mlr_resamplings_cv3]{3-fold cross validation} from \link{mlr_resamplings}.
}
\item Functions to retrieve objects, set hyperparameters and assign to fields in one go (c.f. \code{\link[mlr3misc:dictionary_sugar]{mlr3misc::dictionary_sugar()}}):
\itemize{
\item \code{tsk()} for a \link{Task} from \link{mlr_tasks}.
\item \code{tgen()} for a \link{TaskGenerator} from \link{mlr_task_generators}.
\item \code{lrn()} for a \link{Learner} from \link{mlr_learners}.
\item \code{rsmp()} for a \link{Resampling} from \link{mlr_resamplings}.
\item \code{msr()} for a \link{Measure} from \link{mlr_measures}.
}
\item Each task type has an associated default \link{Measure} which is used if no measure is provided (i.e., the measure is \code{NULL}).
This is done internally via call to \code{\link[=default_measures]{default_measures()}}.
}
}
\examples{
task = tsk("iris", id = "iris2")
print(task)

learner = lrn("classif.rpart", cp = 0.1, predict_type = "prob")
print(learner)
}
