% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_task_generators.R
\docType{data}
\name{mlr_task_generators}
\alias{mlr_task_generators}
\title{Dictionary of Task Generators}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link[mlr3misc:Dictionary]{mlr3misc::Dictionary}.}
\description{
A simple \link{Dictionary} storing generator functions returning a \link{Task}.
}
\section{Methods}{

See \link{Dictionary}.
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link{Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{\link[data.table:data.table]{data.table::data.table()}} with fields "key" and "packages" as columns.
}
}

\examples{
g = mlr_task_generators$get("smiley")
task = g$generate(10)
print(task)
task$data()
}
\seealso{
Example generators: \code{\link[=mlr_task_generators_xor]{xor}}

Other Dictionary: \code{\link{mlr_learners}},
  \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}, \code{\link{mlr_tasks}}

Other TaskGenerator: \code{\link{TaskGenerator}}
}
\concept{Dictionary}
\concept{TaskGenerator}
\keyword{datasets}
