% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGenerator2DNormals.R
\name{mlr_task_generators_2dnormals}
\alias{mlr_task_generators_2dnormals}
\alias{TaskGenerator2DNormals}
\title{2D Normals Classification Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskGenerator}.}
\description{
A \link{TaskGenerator} for the 2d normals task in \code{\link[mlbench:mlbench.2dnormals]{mlbench::mlbench.2dnormals()}}.
}
\section{Construction}{
\preformatted{TaskGenerator2DNormals$new()
mlr_task_generators$get("2dnormals")
tgen("2dnormals")
}
}

\examples{
tgen("2dnormals")$generate(10)$data()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=TaskGenerator]{TaskGenerators}: \link{mlr_task_generators}

\code{as.data.table(mlr_resamplings)} for a complete table of all (also dynamically created) \link{TaskGenerator} implementations.
}
