% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGenerator2DNormals.R
\name{mlr_task_generators_2dnormals}
\alias{mlr_task_generators_2dnormals}
\alias{TaskGenerator2DNormals}
\title{2D Normals Classification Task Generator}
\description{
A \link{TaskGenerator} for the 2d normals task in \code{\link[mlbench:mlbench.2dnormals]{mlbench::mlbench.2dnormals()}}.
}
\section{Dictionary}{

This \link{TaskGenerator} can be instantiated via the \link[mlr3misc:Dictionary]{dictionary} \link{mlr_task_generators} or with the associated sugar function \code{\link[=tgen]{tgen()}}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_task_generators$get("2dnormals")
tgen("2dnormals")
}\if{html}{\out{</div>}}
}

\section{Parameters}{

, |Id |Type    |Default |Range                       |, |:--|:-------|:-------|:---------------------------|, |cl |integer |-       |\eqn{[2, \infty)}{[2, Inf)} |, |r  |numeric |-       |\eqn{[1, \infty)}{[1, Inf)} |, |sd |numeric |-       |\eqn{[0, \infty)}{[0, Inf)} |
}

\examples{
generator = tgen("2dnormals")
plot(generator, n = 200)

task = generator$generate(200)
str(task$data())
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=TaskGenerator]{TaskGenerators}: \link{mlr_task_generators}
\item \code{as.data.table(mlr_task_generators)}  for a table of available \link[=TaskGenerator]{TaskGenerators} in the running session (depending on the loaded packages).
\item Extension packages for additional task types:
\itemize{
\item \CRANpkg{mlr3proba} for probabilistic supervised regression and survival analysis.
\item \CRANpkg{mlr3cluster} for unsupervised clustering.
}
}

Other TaskGenerator: 
\code{\link{TaskGenerator}},
\code{\link{mlr_task_generators_cassini}},
\code{\link{mlr_task_generators_circle}},
\code{\link{mlr_task_generators_friedman1}},
\code{\link{mlr_task_generators_moons}},
\code{\link{mlr_task_generators_simplex}},
\code{\link{mlr_task_generators_smiley}},
\code{\link{mlr_task_generators_spirals}},
\code{\link{mlr_task_generators_xor}},
\code{\link{mlr_task_generators}}
}
\concept{TaskGenerator}
\section{Super class}{
\code{\link[mlr3:TaskGenerator]{mlr3::TaskGenerator}} -> \code{TaskGenerator2DNormals}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskGenerator2DNormals-new}{\code{TaskGenerator2DNormals$new()}}
\item \href{#method-TaskGenerator2DNormals-plot}{\code{TaskGenerator2DNormals$plot()}}
\item \href{#method-TaskGenerator2DNormals-clone}{\code{TaskGenerator2DNormals$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="format"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-format'><code>mlr3::TaskGenerator$format()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="generate"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-generate'><code>mlr3::TaskGenerator$generate()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="TaskGenerator" data-id="print"><a href='../../mlr3/html/TaskGenerator.html#method-TaskGenerator-print'><code>mlr3::TaskGenerator$print()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGenerator2DNormals-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGenerator2DNormals-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGenerator2DNormals$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGenerator2DNormals-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGenerator2DNormals-plot}{}}}
\subsection{Method \code{plot()}}{
Creates a simple plot of generated data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGenerator2DNormals$plot(n = 200L, pch = 19L, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{integer(1)})\cr
Number of samples to draw for the plot. Default is 200.}

\item{\code{pch}}{(\code{integer(1)})\cr
Point char. Passed to \code{\link[=plot]{plot()}}.}

\item{\code{...}}{(any)\cr
Additional arguments passed to \code{\link[=plot]{plot()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskGenerator2DNormals-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskGenerator2DNormals-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskGenerator2DNormals$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
