% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_task_generators_smiley.R
\docType{data}
\name{mlr_task_generators_smiley}
\alias{mlr_task_generators_smiley}
\alias{TaskGeneratorSmiley}
\title{Smiley Classification Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskGenerator}.}
\usage{
TaskGeneratorSmiley
}
\description{
A \link{TaskGenerator} for the smiley task in \code{\link[mlbench:mlbench.smiley]{mlbench::mlbench.smiley()}}.
}
\keyword{datasets}
