% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGeneratorSmiley.R
\name{mlr_task_generators_smiley}
\alias{mlr_task_generators_smiley}
\alias{TaskGeneratorSmiley}
\title{Smiley Classification Task Generator}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskGenerator}.}
\description{
A \link{TaskGenerator} for the smiley task in \code{\link[mlbench:mlbench.smiley]{mlbench::mlbench.smiley()}}.
}
\section{Construction}{
\preformatted{TaskGeneratorSmiley$new()
mlr_task_generators$get("smiley")
tgen("smiley")
}
}

\examples{
tgen("smiley")$generate(10)$data()
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=TaskGenerator]{TaskGenerators}: \link{mlr_task_generators}

\code{as.data.table(mlr_resamplings)} for a complete table of all (also dynamically created) \link{TaskGenerator} implementations.

Other TaskGenerator: 
\code{\link{TaskGenerator}},
\code{\link{mlr_task_generators_2dnormals}},
\code{\link{mlr_task_generators_friedman1}},
\code{\link{mlr_task_generators_xor}},
\code{\link{mlr_task_generators}}
}
\concept{TaskGenerator}
