% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks.R
\docType{class}
\name{mlr_tasks}
\alias{mlr_tasks}
\title{Dictionary of defined Tasks}
\format{\code{\link{R6Class}} object}
\usage{
mlr_tasks
}
\description{
\code{Tasks} is a \code{\link{Dictionary}} used to manage tasks.
}
\examples{
# List task ids:
mlr_tasks$ids

# Get a briew summary:
as.data.frame(mlr_tasks)

# Retrieve a specific task:
mlr_tasks$get("iris")

# Add a new task, based on a subset of iris:
data = iris
data$Species = ifelse(data$Species == "setosa", "1", "0")
task = TaskClassif$new("iris.binary", data = data, target = "Species")
task$classes_n
mlr_tasks$add(task)
mlr_tasks$remove("iris.binary")
}
\keyword{datasets}
