% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks.R
\name{mlr_tasks}
\alias{mlr_tasks}
\title{Dictionary of Tasks}
\description{
A simple \link{Dictionary} storing objects of class \link{Task}.
Each task has an associated help page, see \code{mlr_tasks_[id]}.
}
\section{Methods}{



\itemize{
\item \code{get(key, ...)}\cr
(\code{character(1)}, ...) -> \code{any}\cr
Retrieves object with key \code{key} from the dictionary.
Additional arguments are passed to the stored object during construction.
\item \code{mget(keys, ...)}\cr
(\code{character()}, ...) -> named \code{list()}\cr
Retrieves objects with keys \code{keys} from the dictionary, returns them in a list named with \code{keys}.
Additional arguments are passed to the stored objects during construction.
\item \code{has(keys)}\cr
\code{character()} -> \code{logical()}\cr
Returns a logical vector with \code{TRUE} at its i-th position, if the i-th key exists.
\item \code{keys(pattern)}\cr
\code{character(1)} -> \code{character()}\cr
Returns all keys which comply to the regular expression \code{pattern}.
\item \code{add(key, value, ..., required_args = character())}\cr
(\code{character(1)}, \code{any}, ..., \code{character()}) -> \code{self}\cr
Adds object \code{value} to the dictionary with key \code{key}, potentially overwriting a previously stored item.
Additional arguments in \code{...} are used as default arguments for \code{value} during construction.
If the object is not constructable without additional arguments, the require argument names should be provided in \code{required_args}.
\item \code{remove(key)}\cr
\code{character()} -> \code{self}\cr
Removes object with key \code{key} from the dictionary.
\item \code{required_args(key)}\cr
(\code{character(1)}) -> \code{character()}\cr
Returns the names of arguments required to construct the object.
}
}

\section{S3 methods}{

\itemize{
\item \code{as.data.table(dict)}\cr
\link{Dictionary} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a \code{data.table()} with fields "key", "task_type", "measures", "nrow", "ncol" and
the number of features of type "lgl", "int", "dbl", "chr", "fct" and "ord" as columns.
}
}

\examples{
as.data.table(mlr_tasks)
mlr_tasks$get("iris")
head(mlr_tasks$get("iris")$data())

# Add a new task, based on a subset of iris:
data = iris
data$Species = factor(ifelse(data$Species == "setosa", "1", "0"))
task = TaskClassif$new("iris.binary", data, target = "Species", positive = "1")

# add to dictionary
mlr_tasks$add("iris.binary", task)

# list available tasks
mlr_tasks$keys()

# retrieve from dictionary
mlr_tasks$get("iris.binary")

# remove task again
mlr_tasks$remove("iris.binary")
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_generators}}, \code{\link{mlr_learners}},
  \code{\link{mlr_measures}}, \code{\link{mlr_resamplings}}

Other Task: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{TaskSupervised}},
  \code{\link{Task}}, \code{\link{mlr_generators}}
}
\concept{Dictionary}
\concept{Task}
