% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks.R
\name{mlr_tasks}
\alias{mlr_tasks}
\title{Dictionary of Tasks}
\description{
A simple \link{Dictionary} storing objects of class \link{Task}.
Each task has an associated help page, see \code{mlr_tasks_[id]}.
}
\section{Usage}{

See \link{Dictionary}.
}

\examples{
as.data.table(mlr_tasks)
mlr_tasks$get("iris")
head(mlr_tasks$get("iris")$data())

# Add a new task, based on a subset of iris:
data = iris
data$Species = factor(ifelse(data$Species == "setosa", "1", "0"))
b = as_data_backend(data)
task = TaskClassif$new("iris.binary", b, target = "Species", positive = "1")

# add to dictionary
mlr_tasks$add("iris.binary", task)

# list available tasks
mlr_tasks$keys()

# retrieve from dictionary
mlr_tasks$get("iris.binary")

# remove task again
mlr_tasks$remove("iris.binary")
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_learners}}, \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}},
  \code{\link{mlr_task_generators}}

Other Task: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{TaskSupervised}},
  \code{\link{Task}}, \code{\link{mlr_task_generators}}
}
\concept{Dictionary}
\concept{Task}
