% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks.R
\name{mlr_tasks}
\alias{mlr_tasks}
\title{Dictionary of example learning tasks}
\description{
Dictionary of example learning tasks
}
\section{Usage}{

See \link{Dictionary}.
}

\examples{
mlr_tasks$ids
as.data.table(mlr_tasks)
mlr_tasks$get("iris")
head(mlr_tasks$get("iris")$data())

# Add a new task, based on a subset of iris:
data = iris
data$Species = ifelse(data$Species == "setosa", "1", "0")
b = DataBackendDataTable$new(data)
task = TaskClassif$new("iris.binary", b, target = "Species")
mlr_tasks$add(task)
mlr_tasks$ids
mlr_tasks$get("iris.binary")
mlr_tasks$remove("iris.binary")
}
\seealso{
Other Dictionary: \code{\link{Dictionary}},
  \code{\link{mlr_learners}}, \code{\link{mlr_measures}},
  \code{\link{mlr_resamplings}}

Other Task: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{TaskSupervised}},
  \code{\link{Task}}
}
\concept{Dictionary}
\concept{Task}
