% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr_mtcars.R
\name{mlr_tasks_mtcars}
\alias{mlr_tasks_mtcars}
\title{Motor Trend Regression Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.
}
\description{
A regression task for the \link[datasets:mtcars]{datasets::mtcars} data set.
Target variable is \code{mpg} (Miles/(US) gallon).
Rownames are stored as variable \code{"model"} with column role \code{"name"}.
}
\section{Construction}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{mlr_tasks$get("mtcars")
tsk("mtcars")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{regr}
\item Dimensions: 32x11
\item Properties: -
\item Has Missings: \code{FALSE}
\item Target: \dQuote{mpg}
\item Features: \dQuote{am}, \dQuote{carb}, \dQuote{cyl}, \dQuote{disp}, \dQuote{drat}, \dQuote{gear}, \dQuote{hp}, \dQuote{qsec}, \dQuote{vs}, \dQuote{wt}
}
}

\seealso{
\itemize{
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}:
\url{https://mlr3book.mlr-org.com/chapters/chapter2/data_and_basic_modeling.html}
\item Package \CRANpkg{mlr3data} for more toy tasks.
\item Package \CRANpkg{mlr3oml} for downloading tasks from \url{https://www.openml.org}.
\item Package \CRANpkg{mlr3viz} for some generic visualizations.
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}
\item \code{as.data.table(mlr_tasks)} for a table of available \link[=Task]{Tasks} in the running session (depending on the loaded packages).
\item \CRANpkg{mlr3fselect} and \CRANpkg{mlr3filters} for feature selection and feature filtering.
\item Extension packages for additional task types:
\itemize{
\item Unsupervised clustering: \CRANpkg{mlr3cluster}
\item Probabilistic supervised regression and survival analysis: \url{https://mlr3proba.mlr-org.com/}.
}
}

Other Task:
\code{\link{Task}},
\code{\link{TaskClassif}},
\code{\link{TaskRegr}},
\code{\link{TaskSupervised}},
\code{\link{TaskUnsupervised}},
\code{\link{california_housing}},
\code{\link{mlr_tasks}},
\code{\link{mlr_tasks_breast_cancer}},
\code{\link{mlr_tasks_diabetes}},
\code{\link{mlr_tasks_german_credit}},
\code{\link{mlr_tasks_iris}},
\code{\link{mlr_tasks_penguins}},
\code{\link{mlr_tasks_sonar}},
\code{\link{mlr_tasks_spam}},
\code{\link{mlr_tasks_wine}},
\code{\link{mlr_tasks_zoo}}
}
\concept{Task}
