% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr_mtcars.R
\name{mlr_tasks_mtcars}
\alias{mlr_tasks_mtcars}
\title{"Motor Trend" Car Road Tests Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.}
\description{
A regression task for the \link[datasets:mtcars]{datasets::mtcars} data set.
Target variable is \code{mpg} (Miles/(US) gallon).
}
\section{Construction}{
\preformatted{mlr_tasks$get("mtcars")
tsk("mtcars")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.
}
