% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr_mtcars.R
\name{mlr_tasks_mtcars}
\alias{mlr_tasks_mtcars}
\title{Motor Trend Regression Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskRegr}.
}
\description{
A regression task for the \link[datasets:mtcars]{datasets::mtcars} data set.
Target variable is \code{mpg} (Miles/(US) gallon).
Rownames are stored as variable \verb{"..rownames} with column role \code{"model"}.
}
\section{Construction}{
\preformatted{mlr_tasks$get("mtcars")
tsk("mtcars")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.

Other Task: 
\code{\link{TaskClassif}},
\code{\link{TaskRegr}},
\code{\link{TaskSupervised}},
\code{\link{TaskUnsupervised}},
\code{\link{Task}},
\code{\link{mlr_tasks_boston_housing}},
\code{\link{mlr_tasks_german_credit}},
\code{\link{mlr_tasks_iris}},
\code{\link{mlr_tasks_pima}},
\code{\link{mlr_tasks_sonar}},
\code{\link{mlr_tasks_spam}},
\code{\link{mlr_tasks_wine}},
\code{\link{mlr_tasks_zoo}},
\code{\link{mlr_tasks}}
}
\concept{Task}
