% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks.R
\name{mlr_tasks_mtcars}
\alias{mlr_tasks_mtcars}
\title{"Motor Trend" Car Road Tests Task}
\description{
A regression task for the \link[datasets:mtcars]{datasets::mtcars} data set.
Target variable is \code{mpg} (Miles/(US) gallon).
}
