% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_pima.R
\name{mlr_tasks_pima}
\alias{mlr_tasks_pima}
\title{Pima Indian Diabetes Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.}
\description{
A classification task for the \link[mlbench:PimaIndiansDiabetes2]{mlbench::PimaIndiansDiabetes2} data set.
Positive class is set to \code{"pos"}.
}
\section{Usage}{
\preformatted{mlr_tasks$get("pima")
tsk("pima")
}
}

