% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_sonar.R
\name{mlr_tasks_sonar}
\alias{mlr_tasks_sonar}
\title{Sonar Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.}
\description{
A classification task for the \link[mlbench:Sonar]{mlbench::Sonar} data set.
Positive class is set to "M" (Mine).
}
\section{Construction}{
\preformatted{mlr_tasks$get("sonar")
tsk("sonar")
}
}

\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}
}
