% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks_spam.R
\name{mlr_tasks_spam}
\alias{mlr_tasks_spam}
\title{Spam Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.}
\description{
A classification task for the \link[kernlab:spam]{kernlab::spam} data set.
Positive class is set to "spam".
}
\section{Usage}{
\preformatted{mlr_tasks$get("spam")
}
}

