% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_spam.R
\name{mlr_tasks_spam}
\alias{mlr_tasks_spam}
\title{Spam Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.}
\source{
Creators:
Mark Hopkins, Erik Reeber, George Forman, Jaap Suermondt.
Hewlett-Packard Labs, 1501 Page Mill Rd., Palo Alto, CA 94304

Donor:
George Forman (gforman at nospam hpl.hp.com) 650-857-7835

Preprocessing:
Columns have been renamed. Preprocessed data taken from the \CRANpkg{kernlab} package.
}
\description{
Spam data set from the UCI machine learning repository (\url{http://archive.ics.uci.edu/ml/datasets/spambase}).
Data set collected at Hewlett-Packard Labs to classify emails as spam or non-spam.
57 variables indicate the frequency of certain words and characters in the e-mail.
The positive class is set to "spam".
}
\section{Construction}{
\preformatted{mlr_tasks$get("spam")
tsk("spam")
}
}

\references{
\cite{mlr3}{dua_2017}
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.

Other Task: 
\code{\link{TaskClassif}},
\code{\link{TaskRegr}},
\code{\link{TaskSupervised}},
\code{\link{Task}},
\code{\link{mlr_tasks_boston_housing}},
\code{\link{mlr_tasks_german_credit}},
\code{\link{mlr_tasks_iris}},
\code{\link{mlr_tasks_mtcars}},
\code{\link{mlr_tasks_pima}},
\code{\link{mlr_tasks_sonar}},
\code{\link{mlr_tasks_wine}},
\code{\link{mlr_tasks_zoo}},
\code{\link{mlr_tasks}}
}
\concept{Task}
