% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr_tasks_wine.R
\name{mlr_tasks_wine}
\alias{mlr_tasks_wine}
\title{Spam Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.}
\description{
A classification task for the \link[rattle.data:wine]{rattle.data::wine} data set.
}
\section{Usage}{
\preformatted{mlr_tasks$get("wine")
}
}

