% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassif_zoo.R
\name{mlr_tasks_zoo}
\alias{mlr_tasks_zoo}
\title{Zoo Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.}
\description{
A classification task for the \link[mlbench:Zoo]{mlbench::Zoo} data set.
}
\section{Construction}{
\preformatted{mlr_tasks$get("zoo")
tsk("zoo")
}
}

