% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\docType{data}
\name{Prediction}
\alias{Prediction}
\title{Abstract Prediction Object}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
This is the abstract base class for task objects like \link{PredictionClassif} and \link{PredictionRegr}.
}
\section{Construction}{
\preformatted{Prediction$new()
}
}

\section{Fields}{

\itemize{
\item \code{row_ids} :: (\code{integer()} | \code{character()})\cr
Vector of row ids for which predictions are stored.
\item \code{truth} :: \code{any}\cr
True (observed) outcome.
\item \code{response} :: \code{any}\cr
Predicted outcome.
\item \code{task_type} :: \code{character(1)}\cr
Stores the type of the \link{Task}.
\item \code{predict_types} :: \code{character()}\cr
Vector of predict types this object stores.
}
}

\section{S3 Methods}{

\itemize{
\item \code{as.data.table(rr)}\cr
\link{BenchmarkResult} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Converts the data to a \code{data.table()}.
}
}

\seealso{
Other Prediction: \code{\link{PredictionClassif}},
  \code{\link{PredictionRegr}}
}
\concept{Prediction}
\keyword{datasets}
