% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{Prediction}
\alias{Prediction}
\title{Abstract Prediction Object}
\format{\link{R6Class} object.}
\description{
This is the abstract base class for task objects like \link{PredictionClassif} and \link{PredictionRegr}.
}
\seealso{
Other Prediction: \code{\link{PredictionClassif}},
  \code{\link{PredictionRegr}}
}
\concept{Prediction}
