% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionClassif.R
\name{PredictionClassif}
\alias{PredictionClassif}
\title{Prediction Object for Classification}
\description{
This object stores the predictions returned by a learner of class \link{LearnerClassif}.
}
\section{Usage}{
\preformatted{p = PredictionClassif$new(task, response, prob = NULL)

p$response
p$truth
p$prob

as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\link{TaskRegr}): Used to extract essential information to assert the correctness of \code{response}.
\item \code{response} (\code{numeric}): Predicted class labels as \code{factor} of the same length as number of observations in the test set.
\item \code{prob} (\code{matrix}): Numeric matrix of class probabilities with \code{task$class_n} columns and rows equal to the number of
observations in the test set. Columns must be named with class levels.
}
}

\section{Details}{

\code{$new()} initializes a new object of class \link{Prediction}.

\code{$response} stores the predicted class labels.

\code{$truth} stores the true class labels.

\code{$prob} stores the label probabilities (if available), or is \code{NULL}.

Object can be transformed to a simple \code{\link[data.table:data.table]{data.table::data.table()}} with \code{data.table::as.data.table()}.
}

\seealso{
Other Prediction: \code{\link{PredictionRegr}},
  \code{\link{Prediction}}
}
\concept{Prediction}
\keyword{internal}
