% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionClassif.R
\name{PredictionClassif}
\alias{PredictionClassif}
\title{Prediction Object for Classification}
\description{
This object stores the predictions returned by a learner of class \link{LearnerClassif}.
}
\section{Usage}{
\preformatted{p = PredictionClassif$new(truth, response, prob = NULL)

p$truth
p$response
p$prob

as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{truth} (\link{factor}): Factor of true class labels (as returned by \code{task$truth()}). Note that empty levels may not be dropped.
\item \code{response} (\link{factor} | \link{character}): Vector of predicted class labels. Must have length \code{length(truth)}.
\item \code{prob} (\link{matrix}): Numeric matrix of class probabilities with \code{levels(truth)} columns and \code{length(truth)} rows.
Columns must be named with class levels.
}
}

\section{Details}{

\code{$new()} initializes a new object of class \link{Prediction}.

\code{$truth} stores the predicted class labels as factor.

\code{$response} stores the predicted class labels as factor.

\code{$prob} stores the label probabilities (if available) as matrix, or is \code{NULL}.

Object can be transformed to a simple \code{\link[data.table:data.table]{data.table::data.table()}} with \code{data.table::as.data.table()}.
}

\seealso{
Other Prediction: \code{\link{PredictionRegr}},
  \code{\link{Prediction}}
}
\concept{Prediction}
\keyword{internal}
