% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionClassif.R
\name{PredictionClassif}
\alias{PredictionClassif}
\title{Prediction Object for Classification}
\description{
This object stores the predictions returned by a learner of class \link{LearnerClassif}.
}
\section{Usage}{
\preformatted{# Construction
p = PredictionClassif$new(task, response, prob = NULL)
#
p$truth
p$response
p$prob
#
as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{task} [\link{Task}]:\cr
Task used for prediction. Used to extract \code{row_ids} and \code{truth}.
Set to \code{NULL} to skip all argument checks during initialization.
Slots \code{p$row_ids} and \code{p$truth} need to be set manually in this case
\item \code{response} [\link{factor} | \link{character}]:\cr
Vector of predicted class labels.
\item \code{prob} [\link{matrix}]:\cr
Numeric matrix of class probabilities with one column for each class in \code{task$all_classes}
and one row for each observation in the test set.
}
}

\section{Details}{

\itemize{
\item \code{$new()} initializes a new object of class \link{Prediction}.
\item \code{$truth} stores the predicted class labels as factor.
\item \code{$response} stores the predicted class labels as factor.
\item \code{$prob} stores the label probabilities (if available) as matrix, or is \code{NULL}.
\item The prediction object can be transformed to a simple \code{\link[data.table:data.table]{data.table()}}
with \code{\link[data.table:as.data.table]{as.data.table()}}.
}
}

\seealso{
Other Prediction: \code{\link{PredictionRegr}},
  \code{\link{Prediction}}
}
\concept{Prediction}
