% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionRegr.R
\name{PredictionRegr}
\alias{PredictionRegr}
\title{Prediction Object for Regression}
\description{
This object stores the predictions returned by a learner of class \link{LearnerRegr}.
}
\section{Usage}{
\preformatted{# Construction
p = PredictionRegr$new(truth, response, se = NULL)
#
p$truth
p$response
p$se
#
as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{truth} [\link{numeric}]:\cr
Numeric vector of true response.
\item \code{response} [\link{numeric}]:\cr
Numeric vector of predictions. Must have length \code{length(truth)}.
\item \code{se} [\link{numeric}]:\cr
Numeric vector of predicted standard error. Must have length \code{length(truth)}.
}
}

\section{Details}{

\itemize{
\item \code{$new()} initializes a new object of class \link{Prediction}.
\item \code{$truth} stores the true values.
\item \code{$response} stores the predicted values.
\item \code{$se} stores the predicted standard errors (if available), or is \code{NULL}.
\item The prediction object can be transformed to a simple \code{\link[data.table:data.table]{data.table()}}
with \code{\link[data.table:as.data.table]{as.data.table()}}.
}
}

\seealso{
Other Prediction: \code{\link{PredictionClassif}},
  \code{\link{Prediction}}
}
\concept{Prediction}
\keyword{internal}
