% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionRegr.R
\name{PredictionRegr}
\alias{PredictionRegr}
\title{Prediction Object for Regression}
\description{
This object stores the predictions returned by a learner of class \link{LearnerRegr}.
}
\section{Usage}{
\preformatted{# Construction
p = PredictionRegr$new(task = NULL, response = NULL, se = NULL)
#
p$truth
p$response
p$se
#
as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\link{Task}):\cr
Task used for prediction. Used to extract \code{row_ids} and \code{truth}.
Set to \code{NULL} to skip all argument checks during initialization.
Slots \code{p$row_ids} and \code{p$truth} need to be set manually in this case
\item \code{truth} (\code{numeric()}):\cr
Numeric vector of true response.
\item \code{response} (\code{numeric()}):\cr
Numeric vector of predictions. One element for each observation in the test set.
\item \code{se} (\code{numeric()}):\cr
Numeric vector of predicted standard error. One element for each observation in the test set.
}
}

\section{Details}{

\itemize{
\item \code{$new()} initializes a new object of class \link{Prediction}.
\item \code{$truth} stores the true values.
\item \code{$response} stores the predicted values.
\item \code{$se} stores the predicted standard errors (if available), or is \code{NULL}.
\item The prediction object can be transformed to a simple \code{\link[data.table:data.table]{data.table::data.table()}}
with \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.
}
}

\seealso{
Other Prediction: \code{\link{PredictionClassif}},
  \code{\link{Prediction}}
}
\concept{Prediction}
