% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionRegr.R
\name{PredictionRegr}
\alias{PredictionRegr}
\title{Prediction Object for Regression}
\description{
This object stores the predictions returned by a learner of class \link{LearnerRegr}.
}
\section{Usage}{
\preformatted{p = PredictionRegr$new(task, response, se = NULL)

p$response
p$se

as.data.table(p)
}
}

\section{Arguments}{

\itemize{
\item \code{task} (\link{TaskRegr}): Used to extract essential information to assert the correctness of \code{response}.
\item \code{response} (\code{numeric}): Predicted response of the same length as number of observations in the test set.
\item \code{se} (\code{numeric}): Predicted standard error of the same length as number of observations in the test set.
}
}

\section{Details}{

\code{$new()} initializes a new object of class \link{Prediction}.

\code{$response} stores the predicted values.

\code{$se} stores the predicted standard errors (if available), or is \code{NULL}.

Object can be transformed to a simple \code{\link[data.table:data.table]{data.table::data.table()}} with \code{data.table::as.data.table()}.
}

\seealso{
Other Prediction: \code{\link{PredictionClassif}},
  \code{\link{Prediction}}
}
\concept{Prediction}
\keyword{internal}
