% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionRegr.R
\docType{data}
\name{PredictionRegr}
\alias{PredictionRegr}
\title{Prediction Object for Regression}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Prediction}.}
\description{
This object stores the predictions returned by a learner of class \link{LearnerRegr}.
The field \code{task_type} is set to \code{"classif"}.
}
\note{
It is possible to initialize this object without any arguments.
This allows to manually construct \link{Prediction} objects in a piecemeal fashion.
Required are "row_ids", "truth", and "predict_type".
Depending on the value of "predict_types", "response" and "se" must also be set.
}
\section{Construction}{
\preformatted{p = PredictionRegr$new(task = NULL, response = NULL, se = NULL)
}
\itemize{
\item \code{task} :: \link{TaskRegr}\cr
Task for which the predictions are made. Used to extract the row ids and the true
response. Must be subsetted to test set.
\item \code{response} :: \code{numeric()}\cr
Vector of numeric response values.
One element for each observation in the test set.
\item \code{se} :: \code{numeric()}\cr
Numeric vector of predicted standard error.
One element for each observation in the test set.
}
}

\section{Fields}{

See \link{Prediction}.
}

\examples{
task = mlr_tasks$get("bh")
learner = mlr_learners$get("regr.featureless")
learner$predict_type = "se"
e = Experiment$new(task, learner)$train()$predict()
p = e$prediction
p$predict_types
head(as.data.table(p))
}
\seealso{
Other Prediction: \code{\link{PredictionClassif}},
  \code{\link{Prediction}}
}
\concept{Prediction}
\keyword{datasets}
