% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample a Learner on a Task}
\usage{
resample(task, learner, resampling)
}
\arguments{
\item{task}{(\link{Task})\cr
Object of type \link{Task}.}

\item{learner}{(\link{Learner})\cr
Object of type \link{Learner}.}

\item{resampling}{(\link{Resampling})\cr
Object of type \link{Resampling}.}
}
\value{
\link{ResampleResult}.
}
\description{
Runs a resampling (possibly in parallel).
}
\examples{
task = mlr_tasks$get("iris")
learner = mlr_learners$get("classif.rpart")
resampling = mlr_resamplings$get("cv")
rr = resample(task, learner, resampling)
print(rr)
rr$aggregated
rr$performance

# Repeat resampling with dummy learner and combine
# the ResampleResults into a BenchmarkResult
learner = mlr_learners$get("classif.dummy")
rr.dummy = resample(task, learner, resampling)

bmr = rr$combine(rr.dummy)
bmr$resamplings
}
