% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample a Learner on a Task}
\usage{
resample(task, learner, resampling, ctrl = list())
}
\arguments{
\item{task}{(\link{Task}):
Object of type \link{Task}.
Instead if a \link{Task} object, it is also possible to provide a key to retrieve a task from the \link{mlr_tasks} dictionary.}

\item{learner}{(\link{Learner}):
Object of type \link{Learner}.
Instead if a \link{Learner} object, it is also possible to provide a key to retrieve a task from the \link{mlr_learners} dictionary.}

\item{resampling}{(\link{Resampling}):
Object of type \link{Resampling}.
Instead if a \link{Resampling} object, it is also possible to provide a key to retrieve a task from the \link{mlr_resamplings} dictionary.}

\item{ctrl}{(named \code{list()}, e.g. as returned by \code{\link[=mlr_control]{mlr_control()}}):
Object to control various parts of the execution. See \code{\link[=mlr_control]{mlr_control()}}.}
}
\value{
\link{ResampleResult}.
}
\description{
Runs a resampling (possibly in parallel).
}
\examples{
\dontshow{
   .threshold = logger::log_threshold(namespace = "mlr3")
   logger::log_threshold(logger::WARN, namespace = "mlr3")
}
task = mlr_tasks$get("iris")
learner = mlr_learners$get("classif.rpart")
resampling = mlr_resamplings$get("cv")

# explicitly instantiate the resampling for this task for reproduciblity
set.seed(123)
resampling$instantiate(task)

rr = resample(task, learner, resampling)
print(rr, digits = 2)
rr$aggregated
rr$performance("classif.ce")

# Repeat resampling with featureless learner and combine
# the ResampleResults into a BenchmarkResult
learner = mlr_learners$get("classif.featureless")
rr.featureless = resample(task, learner, resampling)

bmr = rr$combine(rr.featureless)
bmr$aggregated(objects = FALSE)


\dontshow{
   logger::log_threshold(.threshold, namespace = "mlr3")
}
}
