% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\name{ResampleResult}
\alias{ResampleResult}
\title{Container for Results of \code{resample()}.}
\format{\link{R6Class} object.}
\description{
This is the object returned by \code{\link[=resample]{resample()}}.
}
\section{Usage}{
\preformatted{# Construction
# Users of mlr3 generally don't need to construct this class themselves, but rather use objects returned by [resample()].
rr = ResampleResult$new(data, hash = NULL)

# Members
rr$aggregated
rr$data
rr$errors
rr$hash
rr$learner
rr$measures
rr$resampling
rr$task

# Methods
rr$combine(rr)
rr$experiment(iter)
rr$experiments(iters)
rr$performance(id)
rr$print()
}
}

\section{Arguments}{

\itemize{
\item \code{data} (\code{\link[=data.table]{data.table()}}):
\code{\link[=data.table]{data.table()}} with columns matching the data of an \link{Experiment}.
Each row corresponds to a single experiment.
\item \code{hash} (\code{NULL} | \code{character(1)}): Pre-calculated hash for the combination of \code{task}, \code{learner} and \code{resampling}.
If \code{NULL}, the checksum will be calculated on-demand.
\item \code{rr} (\code{ResampleResult}): Second \link{ResampleResult}.
\item \code{iter} (\code{integer(1)}): Iteration of the experiment to retrieve.
\item \code{iters} (\code{integer}): Iterations of experiments to retrieve as \code{list()}.
\item \code{id} (\code{character(1)}): Identifier of a performance measure.
}
}

\section{Details}{

\itemize{
\item \code{$aggregated} (named \code{numeric()}) returns the aggregated performance measures. The aggregation method is part of the \link{Measure}.
\item \code{$combine()} takes a second \link{ResampleResult} and combines both \link{ResampleResult}s to a \link{BenchmarkResult}.
\item \code{$errors} (\code{logical()}) returns a vector with the i-th element \code{TRUE} if the i-th experiment had at least on error during train or predict.
\item \code{$experiment()} returns an \link{Experiment} for the \code{iter}-th resampling iteration.
\item \code{$experiments()} returns a \code{list} with the slice of \link{Experiment}s for the provided \code{iters}.
\item \code{$hash} (\code{character(1))} stores a hash for the combination of task, learner and resampling.
\item \code{$performance(id)} retrieves the performance values for the measure with id \code{id} as numeric vector.
\item \code{$task}, \code{$learner}, \code{$resampling} and \code{$measure} allow access to the \link{Task}, \link{Learner}, \link{Resampling} and \link{Measure} used in the resampling.
\item \code{as.data.table()} converts the \link{BenchmarkResult} to a \code{\link[=data.table]{data.table()}}.
}
}

