% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\name{ResampleResult}
\alias{ResampleResult}
\title{Container for Results of resample}
\description{
This is the object returned by \link{resample}.
}
\section{Usage}{
\preformatted{rr$task
rr$learner
rr$resampling
rr$measures
rr$experiment(iter)
rr$experiments(iters)
rr$performance
rr$aggregated
}
}

\section{Arguments}{

\itemize{
\item \code{i} (\code{integer}):
Iteration(s) of the experiment(s) to retrieve.
}
}

\section{Details}{

\code{$task}, \code{learner}, \code{resampling} and \code{measure} allow access to the \link{Task}, \link{Learner}, \link{Resampling} and \link{Measure} used in
the resampling.

\code{$experiment()} returns an \link{Experiment} for the \code{iter}-th resampling iteration.

\code{$experiments()} returns a list with the slice of \link{Experiment}s for the provided \code{iters}.

\code{$performance} provides a \link[data.table:data.table]{data.table::data.table} with column \code{iteration} (integer) and a numeric column for each
performance measure (columns named using the measure ids).

\code{$aggregated} returns the aggregated performance measures. The aggregation method is part of the \link{Measure}.
}

