% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\docType{data}
\name{ResampleResult}
\alias{ResampleResult}
\title{Container for Results of \code{resample()}}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
This is the result container object returned by \code{\link[=resample]{resample()}}.
}
\section{Construction}{
\preformatted{rr = ResampleResult$new(data)
}
\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table with the data of one \link{Experiment} per row.
See description of field \code{data} of \link{Experiment} for the exact structure.
}
}

\section{Fields}{

\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Experiment data with one \link{Experiment} per line.
\item \code{task} :: \link{Task}\cr
The task \code{\link[=resample]{resample()}} operated on.
\item \code{learner} :: \link{Learner}\cr
The learner \code{\link[=resample]{resample()}} operated on.
\item \code{resampling} :: \link{Resampling}\cr
The resampling splits \code{\link[=resample]{resample()}} operated on.
\item \code{measures} :: \code{list()} of \link{Measure}\cr
The performance measures \code{\link[=resample]{resample()}} operated on.
\item \code{errors} :: \code{logical()}\cr
Logical vector where the i-th element is \code{TRUE} if an error for the i-th resampling iteration has been captured.
\item \code{hash} :: \code{character(1)}\cr
Hash (unique identifier) for this object.
\item \code{aggregated} :: named \code{numeric()}\cr
Returns a single score for each measure, named with measure ids.
\item \code{prediction} :: \link{Prediction}\cr
Combined \link{Prediction} of all individual experiments.
Note that the performance is not calculated on this object,
but instead on each experiment separately and then combined with an aggregate function.
}
}

\section{Methods}{

\itemize{
\item \code{combine(rr)}\cr
\link{ResampleResult} -> \link{BenchmarkResult}\cr
Takes a second \link{ResampleResult} and combines both \link{ResampleResult}s to a \link{BenchmarkResult}.
\item \code{experiment(iter)}\cr
\code{integer(1)} -> \link{Experiment}\cr
Returns the \code{iter}-th \link{Experiment}.
\item \code{experiments(iters)}\cr
\code{integer()} -> \code{list()} of \link{Experiment}.
Returns a slice of \link{Experiment}s with provided resampling iterations \code{iters}.
Defaults to all experiments.
\item \code{performance(id)}\cr
\code{character(1)} -> \code{numeric(1)}\cr
Retrieves the performance values for the measure with id \code{id} as numeric vector.
}
}

\section{S3 Methods}{

\itemize{
\item \code{as.data.table(rr)}\cr
\link{ResampleResult} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Converts the data to a \code{data.table()}.
}
}

\keyword{datasets}
