% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleResult.R
\docType{data}
\name{ResampleResult}
\alias{ResampleResult}
\title{Container for Results of \code{resample()}}
\format{\link[R6:R6Class]{R6::R6Class} object.}
\description{
This is the result container object returned by \code{\link[=resample]{resample()}}.

Note that all stored objects are accessed by reference.
Do not modify any object without cloning it first.
}
\section{Construction}{
\preformatted{rr = ResampleResult$new(data)
}
\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Table with data for one resampling iteration per row:
\link{Task}, \link{Learner}, \link{Resampling}, iteration (\code{integer(1)}), and \link{Prediction}.
}
}

\section{Fields}{

\itemize{
\item \code{data} :: \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Internal data storage.
We discourage users to directly work with this field.
\item \code{task} :: \link{Task}\cr
The task \code{\link[=resample]{resample()}} operated on.
\item \code{resampling} :: \link{Resampling}\cr
The resampling splits \code{\link[=resample]{resample()}} operated on.
\item \code{errors} :: \code{logical()}\cr
Logical vector where the i-th element is \code{TRUE} if an error for the i-th resampling iteration has been captured.
\item \code{hash} :: \code{character(1)}\cr
Hash (unique identifier) for this object.
\item \code{prediction} :: \link{Prediction}\cr
Combined \link{Prediction} of all individual experiments.
Note that the performance is not calculated on this object,
but instead on each experiment separately and then combined with an aggregate function.
}
}

\section{Methods}{

\itemize{
\item \code{combine(rr)}\cr
\link{ResampleResult} -> \link{BenchmarkResult}\cr
Takes a second \link{ResampleResult} and combines both \link{ResampleResult}s to a \link{BenchmarkResult}.
\item \code{performance(measures = NULL, ids = TRUE)}\cr
(\code{list()} of \link{Measure}, \code{logical(1)}) -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a table with one row for each resampling iteration, including all involved objects.
Additionally calculates the provided performance measures and binds the performance as extra column.
If no measure is provided, defaults to the measure defined in \link[=mlr_reflections]{mlr_reflections$default_measures}
(\link{mlr_measures_classif.ce} for classification and \link{mlr_measures_regr.mse} for regression).
If \code{ids} is \code{TRUE}, character column of id names are added to the table for convenient filtering.
\item \code{aggregate(measures = NULL)}\cr
\code{list()} of \link{Measure} -> named \code{numeric()}\cr
Calculates and aggregates performance values for all provided measures.
See \link{Measure} for the aggregation function.
}
}

\section{S3 Methods}{

\itemize{
\item \code{as.data.table(rr)}\cr
\link{ResampleResult} -> \code{\link[data.table:data.table]{data.table::data.table()}}\cr
Returns a copy of the internal data.
}
}

\keyword{datasets}
