% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{Resampling}
\alias{Resampling}
\title{Abstract resampling class}
\description{
Abstraction for resampling strategies.

Predefined resamplings are stored in \link{mlr_resamplings}.
}
\section{Usage}{
\preformatted{r = Resampling$new(id)
r$id
r$is_instantiated
r$checksum
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\link{Task}):
Task to train/predict on.
\item \code{model} (any):
Fitted model as returned by \code{train}.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{Learner}.

\code{$id} (\code{character(1)}) stores the identifier of the object.

FIXME: Not finished
}

\concept{Resampling}
\keyword{internal}
