% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\docType{data}
\name{Resampling}
\alias{Resampling}
\title{Base Class for Resampling Objects}
\format{\code{\link{R6Class}} object}
\usage{
Resampling
}
\value{
\link{Resampling}.
}
\description{
A \code{\link[R6]{R6Class}} to construct resampling measures.
This is the abstract base class, do not use directly!

Predefined resampling measures are stored in \code{\link{mlr_resamplings}}.
}
\keyword{datasets}
