% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\name{Resampling}
\alias{Resampling}
\title{Abstract resampling class}
\description{
Abstraction for resampling strategies.

Predefined resamplings are stored in \link{mlr_resamplings}.
}
\section{Usage}{
\preformatted{r = Resampling$new(id)
r$id
r$iters
r$par_set
r$par_vals
r$instantiate(task)
r$is_instantiated
r$train_set(i)
r$test_set(i)
r$hash
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}):
identifier for this object.
\item \code{i} (\code{integer(1)}):
Get the \code{i}-th training/test set.
}
}

\section{Details}{

\code{$new()} creates a new object of class \link{Resampling}.

\code{$id} (\code{character(1)}) stores the identifier of the object.

\code{$iters} (\code{integer(1)}) calculates the resulting number of iterations, given the current \code{par_vals}.

\code{$par_set} (\code{[paradox::ParamSet()]}) describes available parameters.

\code{$par_vals} (\code{named list}) stores the currently set parameter values.
You can set parameters by assigning a named list of new parameters to this slot.

\code{$instantiate} materializes fixed training and test splits for a given task.

\code{$is_instantiated} returns \code{TRUE} if the resampling has been instantiated, and \code{FALSE} otherwise.

\code{$train_set()} returns the training set for the \code{i}-th iteration.

\code{$test_set()} returns the test set for the \code{i}-th iteration.

\code{$hash} stores a checksum (\code{character(1)}) calculated on the \code{id}, \code{par_vals} and the instantiation.
If the object is not instantiated yet, \code{NA} is returned.
}

\seealso{
Other Resampling: \code{\link{mlr_resamplings}}
}
\concept{Resampling}
\keyword{internal}
