% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resampling.R
\docType{data}
\name{Resampling}
\alias{Resampling}
\title{Base Class for Resampling Measures}
\format{\code{\link{R6Class}} object}
\usage{
Resampling
}
\value{
[\code{Measure}].
}
\description{
A \code{\link[R6]{R6Class}} to construct resampling measures.
This is the abstract base class, do not use directly!

Predefined resampling measures are stored in \code{\link{mlr.resamplings}}.
}
\section{Fields}{

\describe{
\item{\code{id}}{[\code{character(1)}]: Identifier of the measure.}

\item{\code{task.types}}{[\code{character}]: Set of compatible task types.}

\item{\code{fun}}{[\code{function(truth, predicted)}]: function to compute the measure.}
}}

\keyword{datasets}
