% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingBootstrap.R
\docType{data}
\name{ResamplingBootstrap}
\alias{ResamplingBootstrap}
\alias{mlr_resamplings_bootstrap}
\title{Bootstrap Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\usage{
ResamplingBootstrap
}
\description{
Simple Bootstrap sampling.
Hyperparameters are the number of bootstrap iterations (\code{repeats}, default: 30)
and the ratio of observations to draw per iteration (\code{ratio}, default: 1).
}
\section{Fields}{

See \link{Resampling}.
}

\section{Methods}{

See \link{Resampling}.
}

\examples{
# Create a task with 10 observations
task = mlr_tasks$get("iris")
task$filter(1:10)

# Instantiate Resampling
rb = mlr_resamplings$get("bootstrap")
rb$param_set$values = list(repeats = 2, ratio = 1)
rb$instantiate(task)

# Individual sets:
rb$train_set(1)
rb$test_set(1)
intersect(rb$train_set(1), rb$test_set(1))

# Internal storage:
rb$instance$M # Matrix of counts
}
\keyword{datasets}
