% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingCV.R
\docType{data}
\name{ResamplingCV}
\alias{ResamplingCV}
\alias{mlr_resamplings_cv}
\title{Cross Validation Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\description{
Cross validation with \code{folds} folds (default: 10 folds).
}
\section{Construction}{
\preformatted{ResamplingCV$new()
mlr_resamplings$get("cv")
rsmp("cv")
}
}

\section{Fields}{

See \link{Resampling}.
}

\section{Methods}{

See \link{Resampling}.
}

\examples{
# Create a task with 10 observations
task = tsk("iris")
task$filter(1:10)

# Instantiate Resampling
rcv = rsmp("cv", folds = 3)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)
rcv$test_set(1)
intersect(rcv$train_set(1), rcv$test_set(1))

# Internal storage:
rcv$instance # table
}
\keyword{datasets}
