% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_validate.R
\name{set_validate}
\alias{set_validate}
\title{Enable Validation for Learner}
\usage{
set_validate(learner, validate, ...)
}
\arguments{
\item{learner}{(any)\cr
The learner.}

\item{validate}{(\code{numeric(1)}, \code{"inner_valid"} or \code{NULL})\cr
Which validation set to use.}

\item{...}{(any)\cr
Additional arguments passed to the methods.}
}
\value{
The modified learner.
}
\description{
Helper function to configure validation for a learner.
}
\details{
This is especially useful for learners such as \code{\link[mlr3tuning:AutoTuner]{mlr3tuning::AutoTuner}} or \code{\link[mlr3pipelines:GraphLearner]{mlr3pipelines::GraphLearner}},
where configuring, where the validation set needs to be configured on different levels.
}
\examples{
learner = lrn("classif.debug")
set_validate(learner, 0.2)
learner$validate
}
