% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\docType{data}
\name{Task}
\alias{Task}
\title{Basic Tasks}
\format{\code{\link{R6Class}} object}
\usage{
Task
}
\value{
[\code{\link{Task}}].
}
\description{
This is the abstract base class for task objects.
Use \code{\link{TaskClassif}} or \code{\link{TaskRegr}} to construct tasks instead of this class.
}
\examples{
task = Task$new("iris", data = iris)
task$formula
}
\seealso{
Other Tasks: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}
}
\concept{Tasks}
\keyword{datasets}
