% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskGenerator.R
\name{TaskGenerator}
\alias{TaskGenerator}
\title{Task Generators}
\format{\link{R6Class} object.}
\description{
Creates a \link{Task} of arbitrary size.
Predefined task generators are stored in \link{mlr_task_generators}.
}
\section{Usage}{
\preformatted{# Construction
g = TaskGenerator$new(id, task_type, packages = character(0L), param_set = ParamSet$new(), param_vals = list())

g$id
g$task_type
g$param_set
g$param_vals
g$generate(n)
}
}

\section{Arguments}{

\itemize{
\item \code{id} (\code{character(1)}): Identifier for this object.
\item \code{task_type} (\code{character(1)}): Type of the task generated. E.g., \code{"classif"} or \code{"regr"}.
\item \code{packages} (\code{character()}]: Set of required packages.
\item \code{param_set} (\link[paradox:ParamSet]{paradox::ParamSet}): Set of hyperparameters.
\item \code{param_vals} (named \code{list()}): List of hyperparameter settings.
\item \code{n} (\code{integer(1)}): Number of observations to generate.
}
}

\section{Details}{

\itemize{
\item \code{$id} (\code{character(1)}) stores the identifier of the object.
\item \code{$task_type} (\code{character()}) stores the type of generated task.
\item \code{$packages} (\code{character()}) stores the names of required packages.
\item \code{$param_set} (\link[paradox:ParamSet]{paradox::ParamSet}) describes the available hyperparameter and possible settings.
\item \code{$param_vals} (named \code{list()}) stores the list set hyperparameter values.
\item \code{$generate(n)} creates a task of type \code{task_type} with \code{n} observations.
}
}

\concept{TaskGenerators}
