% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\name{TaskRegr}
\alias{TaskRegr}
\title{Regression Task}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for regression problems.
The target column is assumed to be numeric.
}
\section{Usage}{

In addition to the interface of \link{Task}/\link{TaskSupervised}, this class implements:\preformatted{t = TaskRegr$new(id, backend, target)
}
}

\section{Details}{

\itemize{
\item \code{$task_type} is \code{"regr"}.
}
}

\examples{
b = as_data_backend(iris)
task = TaskRegr$new("iris", backend = b, target = "Sepal.Length")
task$task_type
task$formula
task$truth()
}
\seealso{
Other Task: \code{\link{TaskClassif}},
  \code{\link{TaskSupervised}}, \code{\link{Task}},
  \code{\link{mlr_tasks}}
}
\concept{Task}
