% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\name{TaskRegr}
\alias{TaskRegr}
\title{Regression task}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for regression problems.
The target column is assumed to be numeric.
}
\section{Usage}{
\preformatted{t = TaskRegr$new(id, backend, target)
}
}

\examples{
b = BackendDataTable$new(iris)
task = TaskClassif$new("iris", backend = b, target = "Species")
task$formula
task$class_names
}
\seealso{
Other Tasks: \code{\link{TaskClassif}},
  \code{\link{TaskSupervised}}, \code{\link{Task}}
}
\concept{Tasks}
