% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\name{TaskRegr}
\alias{TaskRegr}
\title{Regression task}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for regression problems.
The target column is assumed to be numeric.
}
\section{Usage}{
\preformatted{t = TaskRegr$new(id, backend, target)

t$type
}
}

\section{Details}{

\code{type} is \code{"classif"}
}

\examples{
b = BackendDataTable$new(iris)
task = TaskRegr$new("iris", backend = b, target = "Sepal.Length")
task$type
task$formula
}
\seealso{
Other Tasks: \code{\link{TaskClassif}},
  \code{\link{TaskSupervised}}, \code{\link{Task}}
}
\concept{Tasks}
