% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegr.R
\name{TaskRegr}
\alias{TaskRegr}
\title{Regression Task}
\format{\link{R6Class} object inheriting from \link{Task}/\link{TaskSupervised}.}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for regression problems.
The target column is assumed to be numeric.
}
\section{Usage}{

Inherits from \link{Task}/\link{TaskSupervised}.\preformatted{t = TaskRegr$new(id, backend, target)
}
}

\section{Details}{

\code{$task_type} is \code{"regr"}.
}

\examples{
b = as_data_backend(iris)
task = TaskRegr$new("iris", backend = b, target = "Sepal.Length")
task$task_type
task$formula
task$truth()
}
\seealso{
Other Task: \code{\link{TaskClassif}},
  \code{\link{TaskSupervised}}, \code{\link{Task}},
  \code{\link{mlr_tasks}}
}
\concept{Task}
