% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSupervised.R
\docType{data}
\name{TaskSupervised}
\alias{TaskSupervised}
\title{Abstract Supervised Tasks}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaskSupervised
}
\description{
Abstract Supervised Tasks
}
\examples{
b = BackendDataTable$new(iris)
task = TaskSupervised$new("iris", backend = b, target = "Species")
task$target_names
task$formula
}
\keyword{datasets}
