% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSupervised.R
\name{TaskSupervised}
\alias{TaskSupervised}
\title{Abstract learning task}
\description{
This is the abstract base class for task objects like \link{TaskClassif} and \link{TaskRegr}.
It extends \link{Task} with some handling of target columns.
}
\section{Usage}{
\preformatted{t = TaskSupervised$new(id, backend, targets)

t$truth(subset = NULL)
}
}

\section{Arguments}{

\itemize{
\item \code{subset} (\code{vector}):
Subset of row ids to subset rows from the \link{Backend} using its primary key.
}
}

\section{Details}{

\code{$truth()} returns (a subset of) the columns which are labeled as target as \code{data.table}.
}

\seealso{
Other Tasks: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{Task}}
}
\concept{Tasks}
\keyword{internal}
