% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSupervised.R
\name{TaskSupervised}
\alias{TaskSupervised}
\title{Abstract supervised task}
\description{
This is the abstract base class for task objects like \link{TaskClassif} and \link{TaskRegr}.
It extends \link{Task} with some handling of target columns.
}
\section{Usage}{
\preformatted{# Construction
t = TaskSupervised$new(id, backend, targets)
#
t$truth(subset = NULL)
}
}

\section{Arguments}{

\itemize{
\item \code{subset} (\code{vector}):\cr
Subset of row ids to subset rows from the \link{DataBackend} using its primary key.
}
}

\section{Details}{

\itemize{
\item \code{$truth()} returns the true labels. The type depends on the type of the task.
}
}

\seealso{
Other Task: \code{\link{TaskClassif}},
  \code{\link{TaskRegr}}, \code{\link{Task}},
  \code{\link{mlr_tasks}}
}
\concept{Task}
\keyword{internal}
