% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_learner.R
\name{test_learner}
\alias{test_learner}
\title{Runs Various Unit Tests on a Learner}
\usage{
test_learner(learner)
}
\arguments{
\item{learner}{(\link{Learner}).}
}
\value{
None.
}
\description{
A learner is tested on many different tasks which cover the learner's specified feature types and properties.
}
\examples{
\dontshow{
   set.seed(123)
   .threshold = logger::log_threshold(namespace = "mlr3")
   logger::log_threshold(logger::WARN, namespace = "mlr3")
}
lrn = mlr_learners$get("classif.featureless")
test_learner(lrn)
\dontshow{
   logger::log_threshold(.threshold, namespace = "mlr3")
}
}
