% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Train a Learner on a Task}
\usage{
train(task, learner, subset = NULL)
}
\arguments{
\item{task}{(\code{\link[=Task]{Task()}})\cr
Object of type \code{\link[=Task]{Task()}}.}

\item{learner}{(\code{\link[=Learner]{Learner()}})\cr
Object of type \code{\link[=Learner]{Learner()}}.}

\item{subset}{(\code{integer} | \code{logical})\cr
Subset of \code{task} to train the data on.}
}
\value{
\link{Experiment}.
}
\description{
Fits a model on a subset of the task data.
}
