% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkResult.R
\name{uhashes}
\alias{uhashes}
\alias{uhash}
\title{Obtain specific uhashes from a \link{BenchmarkResult}}
\usage{
uhashes(bmr, learner_ids = NULL, task_ids = NULL, resampling_ids = NULL)

uhash(bmr, learner_id = NULL, task_id = NULL, resampling_id = NULL)
}
\arguments{
\item{learner_ids}{(\code{character()})\cr
Learner ids.}

\item{task_ids}{(\code{character()})\cr
Task ids.}

\item{resampling_ids}{(\code{character()})\cr
Resampling ids.}

\item{learner_id}{(\code{character(1)})\cr
Learner id.}

\item{task_id}{(\code{character(1)})\cr
Task id.}

\item{resampling_id}{(\code{character(1)})\cr
Resampling id.}
}
\description{
Calculates one or more uhashes (unique identifiers) from the IDs of a learner, task, and resampling.
}
\examples{
design = benchmark_grid(
  tsks(c("sonar", "iris")),
  lrns(c("classif.debug", "classif.featureless", "classif.rpart")),
  rsmp("holdout")
)
bmr = benchmark(design)
bmr
bmr$uhashes
uhash(bmr, learner_id = "classif.debug", task_id = "sonar", resampling_id = "holdout")
uhashes(bmr, learner_ids = c("classif.debug", "classif.featureless"))
}
